% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_byte.R
\name{add.extra_byte}
\alias{add.extra_byte}
\title{Add extra byte field}
\usage{
add.extra_byte(.las, data, datanames = colnames(data), ...)
}
\arguments{
\item{.las}{an object of class LAS}

\item{data}{data.table. The Extra Byte data to be added.}

\item{datanames}{character. A character string with field name to update.}

\item{...}{see details.}
}
\value{
an object of class LAS.
}
\description{
Add extra byte field and specify description in \code{header@VLR}
Avalaible parameters for \code{...} are \code{data_type}, \code{min}, \code{max}, \code{scale},
\code{offset}. These should be vectors of length the number of columns of \code{data} argument.
\code{data_type}: integer describing following code of LAS v1.4. Default is 10, i.e. \code{double}.
Codes are the following 1=U8, 2=I8, 3=U16, 4=I16, 5=U32, 6=I32, 7=U64, 8=I64, 9=F32, 10=F34.
}
\details{
\code{min, max}: minimum and maximum value (in data_type format).

\code{scale, offset}: factor with which data will be scaled before writing to file.
}
