% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DTM.r
\name{as.DTM}
\alias{as.DTM}
\title{Transform a raster into a DTM}
\usage{
as.DTM(r)
}
\arguments{
\item{r}{a RasterLayer from package raster}
}
\description{
Transform a raster (formal class RasterLayer from package raster) into a
DTM format used by the package lidR.
}
\seealso{
\link[raster:raster]{raster}
}

