% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gridmetrics.r
\name{as.grid_metrics}
\alias{as.grid_metrics}
\title{Set the class grid_metrics to a data.frame or a data.table}
\usage{
as.grid_metrics(x)
}
\arguments{
\item{x}{A data.frame}
}
\description{
Set the class grid_metrics to a data.frame. Useful when reading data from a file.
In this case the data.frame does not have the class grid_metrics and cannot easly be
plotted or transformed into a matrix.
}

