% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r
\name{as.gridMetrics}
\alias{as.gridMetrics}
\title{Set the class gridMetrics to a data.frame}
\usage{
as.gridMetrics(x)
}
\arguments{
\item{x}{A data.frame}
}
\description{
Set the class gridMetrics to a data.frame. Usefull when reading data from a file.
In this case the data.frame does not have the class gridMetrics and cannot easly be
plotted or transformed into a matrix
}

