% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_typecast.r
\name{as.lasmetrics}
\alias{as.lasmetrics}
\title{Set the class 'lasmetrics' to a data.frame or a data.table}
\usage{
as.lasmetrics(x, res)
}
\arguments{
\item{x}{A data.frame or a data.table}

\item{res}{numeric the original resolution}
}
\value{
Nothing. Object is updated in place by reference.
}
\description{
Set the class lasmetrics to a data.frame. Useful when reading data from a file.
In this case the data.frame does not have the class lasmetrics and cannot easily be
plotted or transformed into a raster
}
\seealso{
Other cast: \code{\link{as.SpatialPixelsDataFrame}},
  \code{\link{as.SpatialPointsDataFrame}},
  \code{\link{as.raster.lasmetrics}}
}
