% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.gridmetrics.r
\name{as.matrix.grid_metrics}
\alias{as.matrix.grid_metrics}
\title{Tranform a grid_metrics object into a spatial matrix}
\usage{
\method{as.matrix}{grid_metrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{a grid_metrics object}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{\dots}{other parameters for \link[data.table:dcast]{dcast}}
}
\description{
Tranform a grid_metrics object into a spatial matrix
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

meanHeight = grid_metrics(lidar, 20, mean(Z))
mtx = as.matrix(meanHeight)
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}
\link[lidR:canopyModel]{canopyModel}
\link[data.table:dcast]{dcast}
}

