% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r
\name{as.matrix.gridMetrics}
\alias{as.matrix.gridMetrics}
\title{Tranform a gridMetrics object into a spatial matrix}
\usage{
\method{as.matrix}{gridMetrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{a gridMetrics object}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{\dots}{Other parameters for \code{acast}}
}
\description{
Tranform a gridMetrics object into a spatial matrix
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = LoadLidar(LASfile)

meanHeight = gridMetrics(lidar, 20, mean(Z))
mtx = as.matrix(meanHeight)
}
\seealso{
\link[lidR:gridMetrics]{gridMetrics}
\link[lidR:canopyModel]{canopyModel}
\link[reshape2:acast]{acast}
}

