% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.raster.gridmetrics}
\alias{as.raster.gridmetrics}
\title{Transform a grid_metrics object into a spatial raster object}
\usage{
\method{as.raster}{gridmetrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{a grid_metrics object}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{\dots}{other parameters for \link[data.table:dcast]{dcast}}
}
\value{
A RasterLayer object from package  \link[raster:raster]{raster}
}
\description{
Transform a grid_metrics object into a spatial raster object
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

meanHeight = grid_metrics(lidar, mean(Z))
rmeanHeight = as.raster(meanHeight)
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}
\link[lidR:grid_canopy]{grid_canopy}
\link[data.table:dcast]{dcast}
\link[raster:raster]{raster}
}

