% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_typecast.R
\name{as.spatial}
\alias{as.spatial}
\title{Transform a LAS* object into an sp object}
\usage{
as.spatial(x)
}
\arguments{
\item{x}{an object from the lidR package}
}
\value{
An object from sp
}
\description{
LAS and LAScatalog objects are transformed into SpatialPointsDataFrame and SpatialPolygonsDataFrame objects, respectively.
}
\concept{cast}
