% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.SpatialPixelsDataFrame}
\alias{as.SpatialPixelsDataFrame}
\title{Tranform a gridmetric object into a SpatialPixelsDataFrame object}
\usage{
as.SpatialPixelsDataFrame(.data)
}
\arguments{
\item{.data}{an object of class gridmetric}
}
\value{
An object of class SpatialPixelDataFrame
}
\description{
Tranform a gridmetric object into a SpatialPixelsDataFrame object
}
\seealso{
\link[sp:SpatialPixelsDataFrame-class]{SpatialPixelsDataFrame}
}

