% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.SpatialPointsDataFrame}
\alias{as.SpatialPointsDataFrame}
\title{Tranform a LAS object into a SpatialPointsDataFrame object}
\usage{
as.SpatialPointsDataFrame(.las)
}
\arguments{
\item{.las}{an object of class LAS}
}
\value{
An object of class SpatialPointsDataFrame
}
\description{
Tranform a LAS object into a SpatialPointsDataFrame object
}
\seealso{
\link[sp:SpatialPointsDataFrame-class]{SpatialPointsDataFrame}
}

