% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_closure.r
\name{canopy_closure}
\alias{canopy_closure}
\title{Canopy Closure}
\usage{
canopy_closure(x, threshold, na.rm = TRUE)
}
\arguments{
\item{x}{numeric matrix. A canopy surface model}

\item{threshold}{numeric. Threshold value}

\item{na.rm}{logical. If TRUE NA values in the number of cells are not counted (useful for circular plots, for example)}
}
\value{
A numeric percentage value between 0 and 100
}
\description{
canopy_closure computes estimates of canopy closure using a canopy surface model.
Output values for cover estimates range from 0.0 to 100.0 percent. Canopy closure
is defined as the number of cells over a specified height threshold divided by
the total cells of the canopy.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

myClosureMetrics = function(x, y, z)
{
   canopy = local_maximum(x,y,z, 2)

   CC2  = canopy_closure(canopy, 2)
   CC10 = canopy_closure(canopy, 10)
   CC20 = canopy_closure(canopy, 20)

   return(list(CC2 = CC2, CC10 = CC10, CC20 = CC20))
}

closures = grid_metrics(lidar, 20, myClosureMetrics(X,Y,Z))
plot(closures, "CC2")
plot(closures, "CC10")
plot(closures, "CC20")
}
\seealso{
\link[lidR:local_maximum]{local_maximum}
\link[lidR:grid_metrics]{grid_metrics}
\link[lidR:cloud_metrics]{cloud_metrics}
}

