% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las-class.r
\docType{methods}
\name{$,LAS-method}
\alias{$,LAS-method}
\title{Extract parts of a LAS object}
\usage{
\S4method{$}{LAS}(x, name)
}
\arguments{
\item{x}{object from which to extract element(s).}

\item{name}{A literal character string or a name (possibly backtick quoted).}
}
\description{
Extract parts of a LAS object
}

