% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\docType{class}
\name{Catalog-class}
\alias{Catalog-class}
\title{An S4 class to represent a catalog of las tiles.}
\description{
An S4 class to represent a set of las tiles, to plot them and to process them in with
multicore.
}
\details{
A \code{Lidar} object contains a \code{data.frame} in the slot \code{@headers} with the data
read from the headers of all your \code{.las} files.
}
\section{Slots}{

\describe{
\item{\code{headers}}{data.frame. A table representing the las header data}
}}
\seealso{
\link[lidR:processParallel]{processParallel}
\link[lidR:extractGroundInventory]{extractGroundInventory}
}

