% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_boundaries.R
\name{catalog_boundaries}
\alias{catalog_boundaries}
\title{Computes the polygon that encloses the points}
\usage{
catalog_boundaries(ctg, concavity = 5, length_threshold = 5)
}
\arguments{
\item{ctg}{A LAScatalog}

\item{concavity, length_threshold}{see \link{concaveman}}
}
\value{
A LAScatalog with true boundaries
}
\description{
Computes the polygon that encloses the points. It reads all the files one by one and computes a
concave hull using the \link{concaveman} function. When all the hulls are computed it updates the
LAScatalog to set the true polygons instead of the bounding boxes.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
ctg <- readLAScatalog(LASfile, filter = "-drop_z_below 0.5")
ctg2 <- catalog_boundaries(ctg, 1, 15)
plot(ctg)
plot(ctg2, add = TRUE)
}
