% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_index.r
\name{catalog_index}
\alias{catalog_index}
\title{It only works for tiles that are arranged in gridlines. This function is used by \link[lidR:catalog_queries]{catalog_queries}.
Users do not really need it.}
\usage{
catalog_index(catalog, x, y, r, r2, buffer, roinames)
}
\arguments{
\item{catalog}{A LAScatalog object}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{r}{numeric or vector. A radius or a set of radii of the ROI. If only
r is provided (r2 = NULL) it will extract data falling onto a disc.}

\item{r2}{numeric or vector. A radius or a set of radii of plots. If r2
is provided, the selection turns into a rectangular ROI. If r= r2 it is a square.}

\item{roinames}{vector. A set of ROI names}
}
\description{
It only works for tiles that are arranged in gridlines. This function is used by \link[lidR:catalog_queries]{catalog_queries}.
Users do not really need it.
}
\seealso{
\link[lidR:catalog]{LAScatalog}
\link[lidR:catalog_queries]{catalog_queries}
}
\keyword{internal}
