% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_intersect.R
\name{catalog_intersect}
\alias{catalog_intersect}
\title{Subset a LAScatalog with a spatial object}
\usage{
catalog_intersect(
  ctg,
  y,
  ...,
  subset = c("subset", "flag_unprocessed", "flag_processed")
)
}
\arguments{
\item{ctg}{A \link[=LAScatalog-class]{LAScatalog} object}

\item{y}{`bbox`, `sf`, `sfc`, `Extent`, `Raster*`, `Spatial*` objects}

\item{...}{ignored}

\item{subset}{character. By default subset of the catalog. It is also possible to flag
the files to maintain the catalog as a whole but process only a subset of its content.
\code{flag_unprocessed} flags the files that will not be processed.
\code{flag_processed} flags the files that will be processed.}
}
\value{
A LAScatalog
}
\description{
Subset a LAScatalog with a spatial object to keep only the tiles of interest. It can be
used to select tiles of interest that encompass spatial objects.
}
