% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_makecluster.R
\name{catalog_makechunks}
\alias{catalog_makechunks}
\title{Subdivide a LAScatalog into chunks}
\usage{
catalog_makechunks(ctg)
}
\arguments{
\item{ctg}{a object of class \code{LAScatalog}}
}
\value{
A list containing objects of class \code{LAScluster}.
}
\description{
Virtually subdivide a LAScatalog into chunks. This function is an internal function exported to
users in version 3.0.0 because it might be useful for some debugging. I might also be useful for
some advanced developers. Regular users are not expected to use this function. The chunks are made
according to the \link[lidR:catalog_options_tools]{catalog processing options}.
}
