% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_merge.R
\name{catalog_merge_results}
\alias{catalog_merge_results}
\title{Merge a list of outputs returned by cluster_apply into a single object. This function is called
at the end of catalog_apply and only there. It merges all the independent outputs into a single
one. It cannot fail, at worst it returns the list.}
\usage{
catalog_merge_results(ctg, any_list, ...)
}
\arguments{
\item{ctg}{LAScatalog for processing options}

\item{any_list}{a list of same type objects that can be Raster*, sf, sfc, Spatial*, LAS, stars, data.frame}

\item{...}{unused}
}
\description{
Merge a list of outputs returned by cluster_apply into a single object. This function is called
at the end of catalog_apply and only there. It merges all the independent outputs into a single
one. It cannot fail, at worst it returns the list.
}
