% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_catalog_options.r
\name{catalog_options_tools}
\alias{catalog_options_tools}
\alias{get_buffer}
\alias{set_buffer<-}
\alias{get_tiling_size}
\alias{set_tiling_size<-}
\alias{get_alignment}
\alias{set_alignment<-}
\alias{get_cores}
\alias{set_cores<-}
\alias{get_progress}
\alias{set_progress<-}
\alias{get_stop_early}
\alias{set_stop_early<-}
\alias{get_output_files}
\alias{set_output_files<-}
\alias{get_laz_compression}
\alias{set_laz_compression<-}
\alias{get_select}
\alias{set_select<-}
\alias{get_filter}
\alias{set_filter<-}
\title{Get or set LAScatalog processing engine options}
\usage{
get_buffer(ctg)

set_buffer(ctg) <- value

get_tiling_size(ctg)

set_tiling_size(ctg) <- value

get_alignment(ctg)

set_alignment(ctg) <- value

get_cores(ctg)

set_cores(ctg) <- value

get_progress(ctg)

set_progress(ctg) <- value

get_stop_early(ctg)

set_stop_early(ctg) <- value

get_output_files(ctg)

set_output_files(ctg) <- value

get_laz_compression(ctg)

set_laz_compression(ctg) <- value

get_select(ctg)

set_select(ctg) <- value

get_filter(ctg)

set_filter(ctg) <- value
}
\arguments{
\item{ctg}{An object of class \link[lidR:LAScatalog-class]{LAScatalog}}

\item{value}{An appropriated value depending on the expected input.}
}
\value{
Functions \code{get_*} return the value of the associated options
}
\description{
The names of the options and their roles are documented in \link[lidR:LAScatalog-class]{LAScatalog}.
The options are used by all the functions that take a \code{LAScatalog} as input.
}
