% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_boundaries.R
\name{catalog_split_clusters}
\alias{catalog_split_clusters}
\title{Split a catalog into spatially clustered subsets}
\usage{
catalog_split_clusters(ctg, buffer = 1)
}
\arguments{
\item{ctg}{A LAScatalog}

\item{buffer}{Numeric. Buffer distance applied to tile geometry before clustering. Defaults to 1.}
}
\value{
A list of LAScatalog
}
\description{
This function groups elements of a catalog into spatial clusters based on
contiguity of the tiles. The original catalog is then split into `n` catalogs
}
