% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\docType{methods}
\name{clipRectangle}
\alias{clipCircle}
\alias{clipCircle,LAS-method}
\alias{clipPolygon}
\alias{clipPolygon,LAS-method}
\alias{clipRectangle}
\alias{clipRectangle,LAS-method}
\title{Clip LiDAR points}
\usage{
clipRectangle(obj, xleft, ybottom, xright, ytop, inside = TRUE)

\S4method{clipRectangle}{LAS}(obj, xleft, ybottom, xright, ytop,
  inside = TRUE)

clipPolygon(obj, x, y, inside = TRUE)

\S4method{clipPolygon}{LAS}(obj, x, y, inside = TRUE)

clipCircle(obj, xcenter, ycenter, radius, inside = TRUE)

\S4method{clipCircle}{LAS}(obj, xcenter, ycenter, radius, inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{xleft}{scalar. of left x position.}

\item{ybottom}{scalar. of bottom y position.}

\item{xright}{scalar. of right x position.}

\item{ytop}{scalar. of top y position.}

\item{inside}{logical. Keep data inside or outside the shape}

\item{x}{numerical array. x-coordinates of polygon}

\item{y}{numerical array. y-coordinates of polygon}

\item{xcenter}{scalar. x disc center}

\item{ycenter}{calar. y disc center}

\item{radius}{a scalar. Disc radius}
}
\value{
An object of class \code{LAS}
}
\description{
Clip LiDAR points within a given geometry
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

subset = lidar \%>\% clipRectangle(xleft=685000, ybottom=5018000,
                                 xright=685100, ytop =5018100)

plot(subset)
}

