% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\docType{methods}
\name{clipCircle}
\alias{clipCircle}
\alias{clipCircle,LAS-method}
\title{Clip LiDAR points within a disc}
\usage{
clipCircle(obj, xcenter, ycenter, radius, inside = TRUE)

\S4method{clipCircle}{LAS}(obj, xcenter, ycenter, radius, inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{xcenter}{a scalar. x disc center}

\item{ycenter}{a scalar. y disc center}

\item{radius}{a scalar. Disc radius}

\item{inside}{logical. Keep data inside or outside the shape}
}
\value{
An object of class \code{LAS}
}
\description{
Clip LiDAR points within a disc
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

subset = lidar \%>\% clipCircle(685000, 5018000, 25)

plot(subset)
}
\seealso{
\link[lidR:clipRectangle]{clipRectangle}
\link[lidR:clipCircle]{clipCircle}
\link[lidR:clipCircle]{clipPolygon}
}

