% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\docType{methods}
\name{clipRectangle}
\alias{clipRectangle}
\alias{clipRectangle,LAS-method}
\title{Clip LiDAR points within a rectangle}
\usage{
clipRectangle(obj, xleft, ybottom, xright, ytop, inside = TRUE)

\S4method{clipRectangle}{LAS}(obj, xleft, ybottom, xright, ytop,
  inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{xleft}{a scalar of left x position.}

\item{ybottom}{a scalar of bottom y position.}

\item{xright}{a scalar of right x position.}

\item{ytop}{a scalar of top y position.}

\item{inside}{logical. Keep data inside or outside the shape}
}
\value{
An object of class \code{LAS}
}
\description{
Clip LiDAR points within a rectangle
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

subset = lidar \%>\% clipRectangle(xleft=685000, ybottom=5018000,
                                 xright=685100, ytop =5018100)

plot(subset)
}
\seealso{
\link[lidR:clipRectangle]{clipRectangle}
\link[lidR:clipCircle]{clipCircle}
\link[lidR:clipCircle]{clipPolygon}
}

