% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{cloudMetrics}
\alias{cloudMetrics}
\alias{cloudMetrics,Lidar-method}
\title{Compute metrics for a cloud of points}
\usage{
cloudMetrics(obj, func)

\S4method{cloudMetrics}{Lidar}(obj, func)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{func}{The function to be apply to cloud of points}
}
\value{
It returns a \code{data.table} with containing the metrics
}
\description{
Computes a series of descriptive statistics for a LIDAR dataset
}
\details{
Computes a series of descriptive statistics for a LIDAR data set. Cloudmetrics
computes a single set of metrics for the entire data set.#'
}

