% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexHull.r
\name{convexHull}
\alias{convexHull}
\title{Compute the convex hull of set of points}
\usage{
convexHull(x, y)
}
\arguments{
\item{x}{coordinates vectors of points}

\item{y}{coordinates vectors of points}
}
\value{
A data frame with the coordinates of the unique points lying on the convex hull, in clockwise order. The first point is repeated to close the hull.
}
\description{
Computes the convex hull of the set of points specified.
}
\examples{
x = runif(20)
y = runif(20)
hull = convexHull(x,y)

plot(x,y)
lines(hull)
}
\seealso{
\link[grDevices:chull]{chull}
}

