% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_ground_segmentation.r
\name{csf}
\alias{csf}
\title{Cloth Simulation Filter}
\usage{
csf(sloop_smooth = FALSE, class_threshold = 0.5, cloth_resolution = 0.5,
  rigidness = 1L, iterations = 500L, time_step = 0.65)
}
\arguments{
\item{sloop_smooth}{logical. When sharp slopes exist, set this parameter to TRUE to perform a
post-processing which will reduced errors.}

\item{class_threshold}{scalar. The distance to the simulated cloth to classify point cloud into ground
and non-ground. The default is 0.5.}

\item{cloth_resolution}{scalar. The distance between paticles in cloth. This is usually set to the
average distance of the points in the point cloud. The default value is 0.5.}

\item{rigidness}{integer. The rididness of the cloth. 1 stands for very soft cloth (to fit rugged
terrain), 2 stands for medium cloth and 3 stands for hard cloth (for flat terrain). The default is 1.}

\item{iterations}{integer. Maximum iteration for simulating cloth. The default value is 500. Usually,
users do not need to change this.}

\item{time_step}{scalar. Time step when simulating the cloth under the gravity. The default value
is 0.65. Usually, do not change this value. It is suitable for most cases.}
}
\description{
This function is made to be used in \link{lasground}. It implements an algorithms for segmentation
of ground points. This method is the a strict implementation of the CSF algorithm made by Zhang
et al. (2016) (see references) that relies on the orginal source code written by the original
author and exposed to R via the the \code{RCSF} package.
}
\references{
W. Zhang, J. Qi*, P. Wan, H. Wang, D. Xie, X. Wang, and G. Yan, “An Easy-to-Use Airborne LiDAR Data
Filtering Method Based on Cloth Simulation,” Remote Sens., vol. 8, no. 6, p. 501, 2016.
(http://www.mdpi.com/2072-4292/8/6/501/htm)
}
\seealso{
Other Algorithm: \code{\link{dalponte2016}},
  \code{\link{hamraz2016}}, \code{\link{highest}},
  \code{\link{homogenize}}, \code{\link{li2012}},
  \code{\link{lmf}}, \code{\link{manual}},
  \code{\link{multichm}}, \code{\link{pmf}},
  \code{\link{ptrees}}, \code{\link{random}},
  \code{\link{silva2016}}, \code{\link{tin}},
  \code{\link{watershed}}, \code{\link{wing2015}}

Other Ground Segmentation: \code{\link{lasground}},
  \code{\link{pmf}}
}
