% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentify.r
\name{detect_pulse}
\alias{detect_pulse}
\title{Retrieve individual pulse}
\usage{
detect_pulse(obj)
}
\arguments{
\item{obj}{A LAS object}
}
\value{
Return nothing. The original object is modified in place by reference.
}
\description{
Retrieve each individual pulse by attributing to each point a number. The
function rely on the GPS time and the return number to retrieve each indivudual
beam. Once points are ordered by GPS time the pattern of return number enable
to attribute a new number to the points every first return.
}

