% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_digital_surface_model.r
\name{p2r}
\alias{p2r}
\alias{dsmtin}
\alias{pitfree}
\title{Canopy Surface Model algorithms}
\usage{
p2r(subcircle = 0)

dsmtin(max_edge = 0)

pitfree(thresholds = c(0, 2, 5, 10, 15), max_edge = c(0, 1),
  subcircle = 0)
}
\arguments{
\item{subcircle}{numeric. radius of the circles. To obtain fewer empty pixels the algorithm
can replace each return with a circle composed of 8 points (see details).}

\item{max_edge}{numeric. Maximum edge-length of a triangle in the Delaunay triangulation.
If a triangle has an edge greater than this value it will be removed. It is used to drive
the pit-free algorithm and to trim dummy interpolation on non-convex areas.
The first number is the value for the classical triangulation (threshold = 0), the second number
is the value for the pit-free algorithm (for thresholds > 0). If \code{max_edge = 0} no trimming
is done (see examples).}

\item{thresholds}{numeric. Set of height thresholds accoring to Khosravipour et al. (2014) algorithm
description (see references)}
}
\description{
\describe{
\item{p2r}{points-to-raster based method: for each pixel of the ouput raster the function attribute
the height of the highest point found.}
\item{dsmtin}{triangulation based method: Delaunay triangulation of first returns with a linear
interpolation within each triangle.}
\item{pitfree}{the pit-free algorithm developed by Khosravipour et al. (2014), which is based on
the computation of a set of classical triangulations at different heights (see references).}
}
\cr The \code{subcircle} tweak replaces each point with 8 points around the original one. This allows
for virtual 'emulation' of the fact that a lidar point is not a point as such, but more
realistically a disc. This tweak densifies the point cloud and the resulting canopy model is
smoother and contains fewer 'pits' and empty pixels.
}
\references{
Khosravipour, A., Skidmore, A. K., Isenburg, M., Wang, T., & Hussin, Y. A. (2014).
Generating pit-free canopy height models from airborne lidar. Photogrammetric Engineering &
Remote Sensing, 80(9), 863-872.
}
