% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{dim,Lidar-method}
\alias{dim}
\alias{dim,Lidar-method}
\title{Dimensions of a Lidar object}
\usage{
\S4method{dim}{Lidar}(x)
}
\arguments{
\item{x}{An object of the class \code{Lidar}}
}
\description{
Retrieve the dimension of a Lidar object.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)

dim(lidar)
}

