% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efi_metrics.R
\name{efi_metrics}
\alias{efi_metrics}
\title{Computes metrics for each plot of a ground inventory}
\usage{
efi_metrics(las, func, geometry, ...)
}
\arguments{
\item{las}{An object of class \link[lidR:LAS-class]{LAS} or \link[lidR:LAScatalog-class]{LAScatalog}.}

\item{func}{formula. An expression to be applied to each cell (see also \link{grid_metrics}).}

\item{geometry}{a spatial object. Many types are supported (see also \link{clip_roi})}

\item{...}{optional supplementary options (see also \link{clip_roi})}
}
\value{
A `data.frame` with all the metrics for each plot as well as all the attributes of the spatial
object used to query the inventory such as plot ID, basal area, above ground biomass
}
\description{
Computes metrics for each plot of a ground inventory by 1. clipping the plots inventories 2. computing
user's metrics to each plot 3. combining spatial data and metrics into one data.frame ready for
statistical modelling. `efi_metrics` is basically a seamless wrapper around \link{clip_roi},
\link{cloud_metrics}, `cbind` and adequate processing settings.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
SHPfile <- system.file("extdata", "efi_plot.shp", package="lidR")
las <- readLAS(LASfile)
inventory <- sf::st_read(SHPfile, quiet = TRUE)
inventory # contains an ID and a Value Of Interest (VOI) per plot

M <- efi_metrics(las, ~list(q85 = quantile(Z, probs = 0.85)), inventory, radius = 11.28)
model <- lm(VOI ~ q85, M)

M <- efi_metrics(las, .stdmetrics_z, inventory, radius = 11.28)

\dontrun{
# Works with polygons as well
inventory <- sf::st_buffer(inventory, 11.28)
plot(las@header)
plot(sf::st_geometry(inventory), add = TRUE)
M <- efi_metrics(las, .stdmetrics_z, inventory)
}
}
\seealso{
Other metrics: 
\code{\link{cloud_metrics}()},
\code{\link{grid_metrics}()},
\code{\link{point_metrics}()},
\code{\link{tree_metrics}()},
\code{\link{voxel_metrics}()}
}
\concept{metrics}
