% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasextent.r
\docType{methods}
\name{extent,LAS-method}
\alias{extent}
\alias{extent,LAS-method}
\alias{lasextent}
\title{Extent}
\usage{
\S4method{extent}{LAS}(x)

extent(x, ...)
}
\arguments{
\item{x}{An object of the class \code{LAS}}

\item{\dots}{Unused}
}
\value{
Extent object
}
\description{
Returns an Extent object of a \code{LAS} object.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

extent(lidar)
}
\seealso{
\code{\link[raster:extent]{raster::extent} }
}

