% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.r
\docType{methods}
\name{extent}
\alias{extent}
\alias{extent,Lidar-method}
\title{Extent}
\usage{
extent(x)

\S4method{extent}{Lidar}(x)
}
\arguments{
\item{x}{An object of the class \code{Lidar}}

\item{\dots}{Unused}
}
\value{
Extent object
}
\description{
Returns an Extent object of a \code{Lidar} object.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

extent(lidar)
}
\seealso{
\code{\link[raster:extent]{raster::extent} }
}

