% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_points.R
\name{filter_points}
\alias{filter_points}
\title{Return points with matching conditions}
\usage{
filter_points(las, ...)
}
\arguments{
\item{las}{An object of class \code{\link[lidR:LAS-class]{LAS}}}

\item{\dots}{Logical predicates. Multiple conditions are combined with '&' or ','}
}
\value{
An object of class \code{\link[lidR:LAS-class]{LAS}}
}
\description{
Return points with matching conditions
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Select the first returns classified as ground
firstground = filter_points(lidar, Classification == 2L & ReturnNumber == 1L)

# Multiple arguments are equivalent to &
firstground = filter_points(lidar, Classification == 2L, ReturnNumber == 1L)

# Multiple criteria
first_or_ground = filter_points(lidar, Classification == 2L | ReturnNumber == 1L)
}
\seealso{
Other filters: 
\code{\link{filter_duplicates}()},
\code{\link{filter_surfacepoints}()},
\code{\link{filters}}
}
\concept{filters}
