% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{forest.colors}
\alias{forest.colors}
\title{Color palette of green}
\usage{
forest.colors(n)
}
\arguments{
\item{n}{The number of colors (> 1) to be in the palette}
}
\description{
Create a vector of n contiguous colors of green from darkgreen to lightgreen.
}
\seealso{
\link[grDevices:colorRamp]{colorRampPalette}
}

