% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.dimension.r
\name{fractal.dimension}
\alias{fractal.dimension}
\title{fractal.dimension}
\usage{
fractal.dimension(mtx)
}
\arguments{
\item{mtx}{a matrix which is the representation of a surface model}
}
\value{
A number between 0 and 3. 3 Being the dimension of a volume
}
\description{
Computes the fractal dimention of a surface. The fractal dimension is a measure
of roughness.
}
\details{
fractal dimension compute the roughness based on box counting method  (see Taud and Parrot)
}
\references{
Taud, H., & Parrot, J.-F. (2005). Mesure de la rugosite des MNT a l'aide de la dimension fractale. Geomorphologie : Relief, Processus, Environnement, 4, 327-338. http://doi.org/10.4000/geomorphologie.622
}

