% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.r
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,LAS-method}
\title{Get LiDAR data}
\usage{
getData(obj)

\S4method{getData}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
It returns a \code{data.table} containing the LiDAR data
}
\description{
Return the slot @data from a \code{LAS} object
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = readLAS(LASfile)

getData(lidar)
}

