% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstOfMany.r
\docType{methods}
\name{getFirstOfMany}
\alias{getFirstOfMany}
\alias{getFirstOfMany,LAS-method}
\title{Filter first returns from pulses which returned multiple points}
\usage{
getFirstOfMany(obj)

\S4method{getFirstOfMany}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the first returns from pulses which returned multiple points
}
\note{
\code{getFirstOfMany(obj)} is an alias for \code{lasfilter(obj, NumberOfReturns > 1, ReturnNumber == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

firstOfManyReturns = lidar \%>\% getFirstOfMany
}
\seealso{
\link[lidR:lasfilter]{lasfilter}

Other getters: \code{\link{getFirstLast}},
  \code{\link{getFirst}}, \code{\link{getGround}},
  \code{\link{getLast}}, \code{\link{getNth}},
  \code{\link{getSingle}}
}

