% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{getLast,Lidar-method}
\alias{getLast}
\alias{getLast,Lidar-method}
\title{Filter last returns}
\usage{
\S4method{getLast}{Lidar}(obj)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}
}
\value{
An object of class \code{Lidar}
}
\description{
Select only the last returns form a \code{Lidar} object i.e. the last returns and the single returns
}
\note{
\code{getLast(obj)} is an alias for \code{leach(obj, ReturnNumber == NumberOfReturns))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)

lastReturns = lidar \%>\% getLast
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:leach]{leach} }
}

