% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLast.r
\docType{methods}
\name{getLast}
\alias{getLast}
\alias{getLast,LAS-method}
\title{Filter last returns}
\usage{
getLast(obj)

\S4method{getLast}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the last returns i.e. the last returns and the single returns
}
\note{
\code{getLast(obj)} is an alias for \code{extract(obj, ReturnNumber == NumberOfReturns))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

lastReturns = lidar \%>\% getLast
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:extract]{extract} }
}

