% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{getNth,Lidar-method}
\alias{getNth}
\alias{getNth,Lidar-method}
\title{Filter returns by their position in the return sequence}
\usage{
\S4method{getNth}{Lidar}(obj, n)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{n}{numeric. The position in the return sequence}
}
\value{
An object of class \code{Lidar}
}
\description{
Select the returns from their position in the return sequence form a \code{Lidar} object
}
\details{
function for \code{Lidar} objects which replace the function
\code{\link[dplyr:filter]{filter}} from \code{\link[dplyr:dplyr]{dplyr}}
package.
}
\note{
\code{getNth(obj, n)} is an alias for \code{leach(obj, ReturnNumber == n)}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)

secondReturns = lidar \%>\% getNth(2)
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:leach]{leach} }
}

