% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{getNth}
\alias{getNth}
\alias{getNth,Lidar-method}
\title{Filter returns by their position in the return sequence.}
\usage{
getNth(obj, n)

\S4method{getNth}{Lidar}(obj, n)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{n}{numeric. The position in the return sequence}
}
\value{
An object of class \code{Lidar}
}
\description{
Select the returns from their position in the return sequence. Point density
pulse density and area are recomputed on the fly
}
\note{
\code{getNth(obj, n)} is an alias for \code{extract(obj, ReturnNumber == n)}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

secondReturns = lidar \%>\% getNth(2)
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:extract]{extract} }
}

