% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingle.r
\docType{methods}
\name{getSingle}
\alias{getSingle}
\alias{getSingle,LAS-method}
\title{Filter single returns}
\usage{
getSingle(obj)

\S4method{getSingle}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the returns which return only one point.
}
\note{
\code{getSingle(obj)} is an alias for \code{lasfilter(obj, NumberOfReturns == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

singleReturns = lidar \%>\% getSingle
}
\seealso{
\link[lidR:lasfilter]{lasfilter}

Other getters: \code{\link{getFirstLast}},
  \code{\link{getFirstOfMany}}, \code{\link{getFirst}},
  \code{\link{getGround}}, \code{\link{getLast}},
  \code{\link{getNth}}
}

