% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasgetters.r
\docType{methods}
\name{getters}
\alias{getFirst}
\alias{getFirst,LAS-method}
\alias{getFirstLast}
\alias{getFirstLast,LAS-method}
\alias{getFirstOfMany}
\alias{getFirstOfMany,LAS-method}
\alias{getLast}
\alias{getLast,LAS-method}
\alias{getNth}
\alias{getNth,LAS-method}
\alias{getSingle}
\alias{getSingle,LAS-method}
\alias{getters}
\title{Predefined filters}
\usage{
getFirst(.las)

\S4method{getFirst}{LAS}(.las)

getFirstLast(.las)

\S4method{getFirstLast}{LAS}(.las)

getFirstOfMany(.las)

\S4method{getFirstOfMany}{LAS}(.las)

getLast(.las)

\S4method{getLast}{LAS}(.las)

getNth(.las, n)

\S4method{getNth}{LAS}(.las, n)

getSingle(.las)

\S4method{getSingle}{LAS}(.las)

getFirstOfMany(.las)

\S4method{getFirstOfMany}{LAS}(.las)
}
\arguments{
\item{.las}{An object of class \code{\link[lidR:LAS-class]{LAS}}}
}
\value{
An object of class \code{\link[lidR:LAS-class]{LAS}}
}
\description{
Select only the some returns
}
\details{
\itemize{
\item{\code{getFirst} Select only the first returns.}
\item{\code{getFistLast} Select only the first and last returns.}
\item{\code{getGround} Select only the returns classified as ground according to LAS specification v1.3.}
\item{\code{getLast} Select only the last returns i.e. the last returns and the single returns.}
\item{\code{getNth} Select the returns from their position in the return sequence.}
\item{\code{gerFirstOfMany} Select only the first returns from pulses which returned multiple points.}
\item{\code{getSingle} Select only the returns which return only one point.}
}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

firstReturns  = lidar \%>\% getFirst
groundReturns = lidar \%>\% getGround
}
\seealso{
Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}

Other lasfilters: \code{\link{lasfilter}}
}

