% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_density.r
\name{grid_density}
\alias{grid_density}
\title{Map the pulse or point density}
\usage{
grid_density(las, res = 4, ...)
}
\arguments{
\item{las}{An object of class \code{LAS} or \code{LAScatalog}.}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 4 = 16
square meters.}

\item{...}{Additional argument for \link{readLAS} to reduce the amount of data loaded (only with a
\code{LAScatalog} object)}
}
\value{
A \code{RasterLayer} or a \code{RasterBrick} containing the metrics for each cell.
}
\description{
Creates a map of the point density. If a "pulseID" field is found, return also a map of the pulse
density.
}
\section{Working with a \code{LAScatalog}}{

In \code{lidR} when the input is a \link[lidR:LAScatalog-class]{LAScatalog} the functions run following
the catalog processing options. The user can modify the processing options using the
\link[lidR:catalog]{available options}.\cr\cr
\code{lidR} supports .lax files that \strong{significantly} improve the computation speed of
spatial queries using a spatial index. Users should really take advange a .lax files but this is
not mandatory.
}

\section{Supported processing options for a LAScatalog}{

\itemize{
\item \strong{cores}: See \link{LAScatalog-class}.
\item \strong{progress}: See \link{LAScatalog-class}.
\item \strong{stop_early}: See \link{LAScatalog-class}.
\item \strong{tiling_size}: See \link{LAScatalog-class}.
}
}

\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

d = grid_density(lidar, 5)
plot(d)
d = grid_density(lidar, 10)
plot(d)
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}
}
