% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{height.colors}
\alias{height.colors}
\title{height.colors}
\usage{
height.colors(n)
}
\arguments{
\item{n}{The number of colors (> 1) to be in the palette}
}
\description{
Create a vector of n contiguous colors of elevations.
}
\seealso{
\link[lidR:height.colors]{height.colors}
\link[lidR:forest.colors]{forest.colors}
\link[grDevices:heat.colors]{heat.colors}
\link[grDevices:colorRamp]{colorRampPalette}
}

