% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-generic.r, R/LASmethods-LAS.r
\docType{methods}
\name{is.empty}
\alias{is.empty}
\alias{is.empty,LAS-method}
\title{Test if a \code{LAS} object is empty}
\usage{
is.empty(object, ...)

\S4method{is.empty}{LAS}(object, ...)
}
\arguments{
\item{object}{A \code{LAS} object}

\item{...}{Unused}
}
\value{
TRUE or FALSE
}
\description{
An empty \code{LAS} object is a point cloud with 0 points
}
\examples{
LASfile <- system.file("extdata", "example.laz", package="rlas")
las = readLAS(LASfile)
is.empty(las)

las = new("LAS")
is.empty(las)
}
