% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\alias{knnx}
\title{Search Nearest Neighbors}
\usage{
knn(data, k = 10)

knnx(data, query, k = 10)
}
\arguments{
\item{data}{LAS object for input point cloud}

\item{k}{number of nearest neighbors to search.}

\item{query}{LAS object for query locations}
}
\value{
a list contains:
\enumerate{
\item **nn.index** an n x k matrix for the nearest neighbor indice.
\item **nn.dist** an n x k matrix for the nearest neighbor Euclidean distances.
}
}
\description{
Fast parallelized k-neareast neighbor searching algorithms for point cloud in LAS format
}
