% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAD.r
\name{LAD}
\alias{LAD}
\title{Leaf area density}
\usage{
LAD(z, dz = 1, k = 0.5)
}
\arguments{
\item{z}{vector of positive z coordinates}

\item{dz}{numeric. The thickeness of the layers used (height bin)}

\item{k}{numeric. is the extinction coefficient}
}
\value{
A data.frame containing the bins elevations (z) and leaf area density for each bin (lad)
}
\description{
Computes a leaf area density profile based on Bouvier et al. method (see reference)
}
\details{
The function assessing the number of laser points that actually reached the layer
z+dz and those that when through by the layer [z, z+dz]. Then it compute the log
of this quantity and divided it by the extinction coefficient k as described in Bouvier
et al. By definition the layer 0 will always return 0 because no return went through
the ground. So the layer 0 is removed from the returned results.
}
\examples{
z = dbeta(seq(0, 0.8, length = 1000), 6, 6)*10

lad = LAD(z)

plot(lad, type="l", xlab="Elevation", ylab="Leaf area density")
}
\references{
Bouvier, M., Durrieu, S., Fournier, R. a, & Renaud, J. (2015).  Generalizing predictive models of forest inventory attributes using an area-based approach with airborne LiDAR data. Remote Sensing of Environment, 156, 322-334. http://doi.org/10.1016/j.rse.2014.10.004
}
\seealso{
\link[lidR:gapFractionProfile]{gapFractionProfile}
}

