% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\docType{methods}
\name{-,LAS,ANY-method}
\alias{-,LAS,ANY-method}
\title{Conveniant operator to normalize}
\usage{
\S4method{-}{LAS,ANY}(e1, e2)
}
\arguments{
\item{e1}{a LAS object}

\item{e2}{a digital terrain model}
}
\description{
Conveniant operator to normalize
}
\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

plot(lidar)

# --- First possibility: read the DTM from a file -----
DTMfile <- system.file("extdata", "Topography.tif", package="lidR")
dtm = raster::raster(DTMfile)

lidar_norm = lidar - dtm # is synonyme with normalize(lidar, dtm)

plot(lidar_norm)

# --- Second possibility: compute the DTM with grid_terrain -----

# Linear interpolation is fast, linear = FALSE for spline interpolation
dtm = grid_terrain(lidar, linear = TRUE)

lidar_norm = lidar - dtm

plot(lidar_norm)
}

