% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\docType{methods}
\name{-,LAS,RasterLayer-method}
\alias{-,LAS,RasterLayer-method}
\title{Conveniant operator to normalize}
\usage{
\S4method{-}{LAS,RasterLayer}(e1, e2)
}
\arguments{
\item{e1}{a LAS object}

\item{e2}{a RasterLayer}
}
\description{
Conveniant operator to normalize
}

