% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasarea.r
\name{lasarea}
\alias{lasarea}
\title{Compute the area covered by of a set a points.}
\usage{
lasarea(.las)
}
\arguments{
\item{.las}{An object of the class \code{LAS} or a numeric array of x coordinates}
}
\value{
numeric. The area of the object computed with a convex hull in coordinates units
}
\description{
The area is computed with a convex hull. It is only an approximation if the
shape of the data is not convex.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

lasarea(lidar)
}

