% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.r
\name{lascatalog}
\alias{lascatalog}
\title{Build a catalog of las tiles/files}
\usage{
lascatalog(folder, ...)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of .las files}

\item{\dots}{Extra parameters to \link[base:list.files]{list.files}. Typically `recursive = TRUE`.}
}
\value{
A data.frame with the class Catalog
}
\description{
Methods to creates a \code{Catalog} object from a folder name
}
\details{
A catalog is the representation of a set of las files. A computer cannot load all
the data at the same time. A catalog is a simple way to manage all the file sequentially
reading only the headers.
}
\seealso{
\link[lidR:plot.Catalog]{plot}
\link[lidR:process_parallel]{process_parallel}
\link[lidR:roi_query]{roi_query}
}

