% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lascheck.r
\name{lascheck}
\alias{lascheck}
\title{Inspect a LAS object}
\usage{
lascheck(las)
}
\arguments{
\item{las}{a LAS object}
}
\description{
Performs a deep inspection of a LAS object and print a report. It checks:
\itemize{
\item if the point cloud is valid according to las specification
\item if the header is valid according to las specification
\item if the point cloud is in accordance with the header
\item if the point cloud have duplicated point and degenerated ground points
\item if some pre-processing such as normalization or ground filtering are already done.
}
For th last one the check may produce a wrong estimation in rare case and should only be considered
as an indication. Also this check does not guarantee the full validity of the object. Some tests
may be missing.
}
