% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lascluster.r
\docType{class}
\name{LAScluster-class}
\alias{LAScluster-class}
\title{An S4 class to represent an arbitrary region in the catalog.}
\description{
An S4 class to represent an arbitrary region in the catalog.
}
\keyword{internal}
