% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasfilter.r
\docType{methods}
\name{lasfilter}
\alias{lasfilter}
\alias{lasfilter,LAS-method}
\title{Return points with matching conditions}
\usage{
lasfilter(.data, ...)

\S4method{lasfilter}{LAS}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{LAS}}

\item{\dots}{Logical predicates. Multiple conditions are combined with &.}
}
\value{
An object of class \code{LAS}
}
\description{
Return points with matching conditions. \code{lasfilter} is an overloading
function for \code{LAS} objects which replaces the function
\code{\link[dplyr:filter]{filter}} from \code{\link[dplyr:dplyr]{dplyr}} package.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

# Select the first returns classified as ground
firstground = lidar \%>\% lasfilter(Classification == 1, ReturnNumber == 1)
}
\seealso{
\link[dplyr:filter]{filter}
\link[lidR:LAS]{Class LAS}
\link[lidR:getFirst]{getFirst}
\link[lidR:getFirstLast]{getFirstLast}
\link[lidR:getFirstOfMany]{getFirstOfMany}
\link[lidR:getSingle]{getSingle}
\link[lidR:getLast]{getLast}
\link[lidR:getGround]{getGround}
\link[lidR:getNth]{getNth}
}

