% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasindentify.r
\name{lasflightline}
\alias{lasflightline}
\title{Retrieve individual flightlines}
\usage{
lasflightline(.las, dt = 30)
}
\arguments{
\item{.las}{A LAS object}

\item{dt}{numeric. The threshold time lag used to retrieve flightlines}
}
\value{
Return nothing. The original object is modified in place by reference.
}
\description{
Retrieve each individual flightline by attributing to each point a number. The
function rely on the GPS time to retrieve each indivudual flighlines. In a
continuous dataset, once points are ordered by GPS time, the time between two
consecutive points cannot be important. If the time between two consecutive
points is too long it means that is comes from another flightine. The defaut
thresohold is 30 seconds
}

