% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasmergespatial.r
\name{lasmergespatial}
\alias{lasmergespatial}
\title{Merge a point cloud and a source of spatial data}
\usage{
lasmergespatial(las, source, attribute = NULL)
}
\arguments{
\item{las}{An object of the class \code{LAS}}

\item{source}{An object of class \code{SpatialPolygonsDataFrame} or \code{RasterLayer} or a
\code{RasterStack} or \code{RasterBrick} with RGB colors.}

\item{attribute}{characters. The name of a attribute in the table of attributes of the shapefile or
the name of a new column in the LAS object. Not relevant for RGB colorization.}
}
\value{
An object of the class \code{LAS}.
}
\description{
Merge a point cloud and a source of spatial data. It adds an attribute along each point based on
a value found in the spatial data. Sources of spatial data can be a \code{SpatialPolygonsDataFrame})
or a \code{RasterLayer}.\cr
\itemize{
\item{\code{SpatialPolygonsDataFrame}: it checks if the points belong within each polygons. If
the parameter \code{attribute} is the name of a an attribute in the table of attributes of the shapefile
it assigns to the points the values of that attribute. Otherwise it classifies the points as boolean.
TRUE if the points are in a polygon, FALSE otherwise.}
\item{\code{RasterLayer}: it attributes to each point the value found in each pixel of the \code{RasterLayer}}.
\item{\code{RasterStack} or \code{RasterBrick} must have 3 channels for RGB colors. It colorize the
point cloud with RGB values.}
}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
shapefile_dir <- system.file("extdata", package = "lidR")

las   <- readLAS(LASfile)
lakes <- rgdal::readOGR(shapefile_dir, "lake_polygons_UTM17")

# The attribute "inlake" does not exist in the shapefile.
# Points are classified as TRUE if in a polygon
las    <- lasmergespatial(las, lakes, "inlakes")     # New attribut 'inlakes' is added.
forest <- lasfilter(las, inlakes == FALSE)
plot(las)
plot(forest)

# The attribute "LAKENAME_1" exists in the shapefile.
# Points are classified with the values of the polygons
las <- lasmergespatial(las, lakes, "LAKENAME_1")     # New column 'LAKENAME_1' is added.
}
