% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasnormalize.r
\name{lasnormalize}
\alias{lasnormalize}
\title{Substract digital terrain model}
\usage{
lasnormalize(.las, dtm = NULL, ...)
}
\arguments{
\item{.las}{a LAS objet}

\item{dtm}{a RasterLayer from package \link[raster:raster]{raster}. If NULL the function will
automatically compute it on the fly using the function \link[lidR:grid_terrain]{grid_terrain}.}

\item{...}{optionnal parameters for \link[lidR:grid_terrain]{grid_terrain} if
\code{dtm} parameter is NULL.}
}
\value{
A LAS object.
}
\description{
Substract digital terrain model (DTM) to the LiDAR data to create a dataset
normalized with the ground at 0. The digital terrain model can comes from
several sources such as external file or own computaion. It can also be computed on the
fly.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")
lidar = readLAS(LASfile)

plot(lidar)

# --- First possibility: compute the DTM on the fly -----

lidar_norm = lasnormalize(lidar)

plot(lidar_norm)

# --- Second possibility: read the DTM from a file -----

dtm = raster::raster(terrain.tiff)

lidar_norm = lidar - dtm # is synonyme with lasnormalize(lidar, dtm)

plot(lidar_norm)
}
}
\seealso{
\link[raster:raster]{raster}
\link[lidR:grid_terrain]{grid_terrain}
}

