% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasindentify.r
\name{laspulse}
\alias{laspulse}
\title{Retrieve individual pulses}
\usage{
laspulse(.las)
}
\arguments{
\item{.las}{A LAS object}
}
\value{
Return nothing. The original object is modified in place by reference.
}
\description{
Retrieve each individual pulse by attributing a number to each point. The
function depends on the GPS time to retrieve each individual
beam. A column '\code{pulseID}' is added in the slot \code{@data}
}

