% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasrangecorrection.R
\name{lasrangecorrection}
\alias{lasrangecorrection}
\title{Normalize intensity with a range correction}
\usage{
lasrangecorrection(las, flightlines, Rs = 1000, f = 2.3)
}
\arguments{
\item{las}{An object of class \link[lidR:LAS-class]{LAS} or \link[lidR:LAScatalog-class]{LAScatalog}.}

\item{flightlines}{SpatialPointsDataDrame containing the coordinates of the sensor at different
instant t. The time is stored in the 'gpstime' attribute and the elevation is stored into 'Z'
attribute. Can be computed with \link{sensor_tracking}.}

\item{Rs}{numeric. Range of reference.}

\item{f}{numeric. Exponent. Usually between 2 and 3 in vegetation context.}
}
\value{
An object of class LAS. The attribute 'Intensity' records the normalised intensity. A extra
attribute 'RawIntensity' records the original intensities.
}
\description{
Normalize intensity with a range correction according to the formula (see references):
\deqn{I_{norm} = I_{obs} (\frac{R}{Rs})^f)}
To achieve the range correction the position of the sensor must be known at different descrete times.
Using the 'gpstime' of each point, the position of the sensor is interpolated from the reference
and a range correction is applied.
}
\references{
Gatziolis, D. (2013). Dynamic Range-based Intensity Normalization for Airborne, Discrete Return
Lidar Data of Forest Canopies. Photogrammetric Engineering & Remote Sensing, 77(3), 251–259.
https://doi.org/10.14358/pers.77.3.251
}
