% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasroi.r
\name{lasroi}
\alias{lasroi}
\title{Select an region of interest interactively}
\usage{
lasroi(.las, ...)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{\dots}{Optionnal parameters for the plot function}
}
\value{
An object of class \code{LAS}
}
\description{
Select an region of interest contained into a rectangle using interactively the mouse
}
\details{
\code{lasroi} enable the user to select a region of interest (ROI) drawing
a rectangle with the mouse
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

subset = lasroi(lidar)
}
}

