% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastrees_li2.r
\name{lastrees_li2}
\alias{lastrees_li2}
\title{Individual tree segmentation}
\usage{
lastrees_li2(las, dt1 = 1.5, dt2 = 2, R = 2, Zu = 15, hmin = 2,
  speed_up = 10, ...)
}
\arguments{
\item{las}{An object of the class \code{LAS}.}

\item{dt1}{numeric. Threshold number 1. See reference page 79 in Li et al. (2012). Default 1.5.}

\item{dt2}{numeric. Threshold number 2. See reference page 79 in Li et al. (2012). Default 2.}

\item{R}{numeric. Search radius. See reference page 79 in Li et al. (2012). Default 2.}

\item{Zu}{numeric. If point elevation is greater than Zu, \code{dt2} is used, otherwise \code{dt1} is
used. See reference page 79 in Li et al. (2012). Default 15.}

\item{hmin}{numeric.  Minimum height of a detected tree. Default 2.}

\item{speed_up}{numeric. Maximum radius of a crown. Any value greater than a crown is
good because this parameter does not affect the result. However, it greatly affects the
computation speed. The lower the value, the faster the method. Default is 10.}

\item{...}{Supplementary options. Currently \code{field} is supported to change the default name of
the new column.}
}
\value{
Nothing (NULL), the point cloud is updated by reference. The original point cloud
has a new column named \code{treeID} containing an ID for each point that refers to a segmented tree.
}
\description{
Individual tree segmentation using the Li et al. (2012) algorithm (see reference). This method replaces
the \link[lidR:lastrees_li]{former algorithm} being a slightly closer implementation of the original
paper. This method is a growing region method working at the point cloud level. It is an implementation
(as strict as possible) of the Li et al. 2012 (see references) algorithm made by the \code{lidR}
author but with the addition of a parameter \code{hmin} to stop over-segmentation for objects
that are too low. The classification is done at the point cloud level and the function always
returns nothing (NULL). The original point cloud is updated in place with an ID for each point in
a new column \code{treeID}. Users are free to post-process this output the way they want.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
col = pastel.colors(200)

# Li 2012
lastrees_li2(las)
plot(las, color = "treeID", colorPalette = col)

}
\references{
Li, W., Guo, Q., Jakubowski, M. K., & Kelly, M. (2012). A new method for segmenting individual
trees from the lidar point cloud. Photogrammetric Engineering & Remote Sensing, 78(1), 75-84.
}
\seealso{
Other tree_segmentation: \code{\link{lastrees_dalponte}},
  \code{\link{lastrees_silva}},
  \code{\link{lastrees_watershed}}, \code{\link{lastrees}}
}
