% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastrees_ptrees.r
\name{lastrees_ptrees}
\alias{lastrees_ptrees}
\title{Individual tree segmentation}
\usage{
lastrees_ptrees(las, k)
}
\arguments{
\item{las}{An object of the class \code{LAS}.}

\item{k}{integer vector. A serie of k-nearest neighbors to use.}
}
\value{
X
}
\description{
Individual tree segmentation
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 5")

k_values = c(5,6,7,8,10,12,15,20,25,30,40,60,80,100)
lastrees_ptrees(las, 10)
}
