% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{leach}
\alias{leach}
\alias{leach,Lidar-method}
\title{Return points with matching conditions}
\usage{
leach(.data, ...)

\S4method{leach}{Lidar}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{Lidar}}

\item{\dots}{Logical predicates. Multiple conditions are combined with &.}
}
\value{
An object of class \code{Lidar}
}
\description{
Return points with matching conditions. \code{leach} is an overloading
}
\details{
function for \code{Lidar} objects which replace the function
\code{\link[dplyr:filter]{filter}} from \code{\link[dplyr:dplyr]{dplyr}}
package.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

# Select the first returns classified as ground
firstground = lidar \%>\% leach(Classification == 1, ReturnNumber == 1)
}
\seealso{
\code{\link[dplyr:filter]{filter}}
}

