% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{liblasWriteLAS}
\alias{liblasWriteLAS}
\title{Write a .las file}
\usage{
liblasWriteLAS(file, LASheader, X, Y, Z, I = integer(0), RN = integer(0),
  NoR = integer(0), SDF = integer(0), EoF = integer(0), C = integer(0),
  SA = integer(0), UD = integer(0), PSI = integer(0), T = numeric(0),
  R = integer(0), G = integer(0), B = integer(0))
}
\arguments{
\item{file}{character. filename of .las file}
}
\value{
void
}
\description{
Methods to write .las files
}
\details{
This function is an interface between R and the library liblas.
Users must avoid to use this function standalone. Use w
}

