% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\name{Lidar}
\alias{Lidar}
\title{Create a \code{Lidar} object}
\usage{
Lidar(input, fields = "standard", ...)
}
\arguments{
\item{input}{character. filename(s) of .las file(s)}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}}

\item{\dots}{Unused}
}
\value{
An object of the class \code{Lidar}
}
\description{
Methods to read and creates a \code{Lidar} object from a .las file or from a data.table
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)
}
\seealso{
\link[lidR:Lidar]{Class Lidar}
}

