% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLiDARProject.r
\name{LiDARProject}
\alias{LiDARProject}
\title{Build a LiDARProject object}
\usage{
LiDARProject(folder, ...)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of .las files}

\item{\dots}{Unused}
}
\value{
A LiDARProject object
}
\description{
Build a LiDARProject object
}

