% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\docType{package}
\name{lidR-deprecated}
\alias{lidR-deprecated}
\alias{cloud_metrics}
\alias{lidR-deprecated-package}
\alias{lidR-deprecated}
\alias{lasarea}
\title{Deprecated function(s) in the lidR package}
\usage{
cloud_metrics(...)

lasarea(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the lidR package.  They may eventually be completely
removed.
}
\section{Details}{

\itemize{
\item{\code{\link{lasmetrics}} replace old \code{cloud_metrics}}
\item{\code{\link{area}} replace old \code{lasarea}}
}
}

