% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_index.R
\docType{data}
\name{lidR-spatial-index}
\alias{lidR-spatial-index}
\alias{LIDRAUTOINDEX}
\alias{LIDRGRIDPARTITION}
\alias{LIDRVOXELPARTITION}
\alias{LIDRQUADTREE}
\alias{LIDROCTREE}
\alias{UKNLAS}
\alias{ALSLAS}
\alias{TLSLAS}
\alias{UAVLAS}
\alias{DAPLAS}
\alias{MLSLAS}
\alias{sensor}
\alias{sensor<-}
\alias{index}
\alias{index<-}
\title{Spatial index}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
LIDRAUTOINDEX

LIDRGRIDPARTITION

LIDRVOXELPARTITION

LIDRQUADTREE

LIDROCTREE

UKNLAS

ALSLAS

TLSLAS

UAVLAS

DAPLAS

MLSLAS

sensor(las, h = FALSE)

sensor(las) <- value

index(las, h = FALSE)

index(las) <- value
}
\arguments{
\item{las}{An object of class \link[lidR:LAS-class]{LAS} or \link[lidR:LAScatalog-class]{LAScatalog}.}

\item{h}{boolean. Human readable}

\item{value}{integer. A code for refereeing to a sensor type or a spatial index type. Use
one of \code{UKNLAS}, \code{ALSLAS}, \code{TLSLAS}, \code{UAVLAS}, \code{DAPLAS}, \code{MLSLAS} for sensor and one of
\code{LIDRAUTOINDEX}, \code{LIDRGRIDPARTITION}, \code{LIDRVOXELPARTITION}, \code{LIDRQUADTREE} \code{LIDROCTREE}
for spatial index.}
}
\description{
This document explains how to process point-clouds taking advantage of different spatial
indexes available in the lidR package. lidR can use several types of spatial indexes to
apply algorithms that need a spatial indexing as fast as possible. The choice of the spatial
index depends on the type of point-cloud that is processed and the algorithm that is performed.
lidR can use a grid partition, a voxel partition, a quadtree or an octree. See details.
}
\details{
In lidR (>= 3.1.0), a \link[=LAS-class]{LAS} object records the sensor used to sample
the point-cloud (ALS, TLS, UAV, DAP) as well as the spatial index that must be used
for processing the point cloud. This can be set manually by the user but the simplest is
to use one of the \link[=readLAS]{read*LAS()} function. By default a point-cloud is associated
to a sensor and the best spatial index is chosen on-the-fly depending on the algorithm
applied. It is possible to force the use of a specific spatial index.

Information relative to the spatial indexing are stored in slot \verb{@index} that contains
a \code{list} with two elements:
\itemize{
\item \code{sensor}: an integer that records the sensor type
\item \code{index}: an integer that records the spatial index to used
}

Some global variable defined in the package can be used to attribute manually the adequate
codes if needed. By default the spatial index code is 0 meaning that each function is free
to choose a spatial index depending on the recorded sensor. If the code is not 0 then each
function will be forced to used the spatial index that is imposed.

\link[=LAScatalog-class]{LAScatalog} objects also record such information that is automatically
propagated to the LAS objects when processing.

Notice that sensor code 0 (not registered) is equivalent to ALS for legacy and backward
compatibility reasons. Consequently \code{readLAS()} and \code{readALSLAS()} are equivalent.
}
\examples{
LASfile <- system.file("extdata", "example.laz", package="rlas")
las <- readLAS(LASfile)

# By default the sensor and spatial index codes are 0
sensor(las)
index(las)

# Codes are used internally and not intended to be known by users
# Use h option for human readable output
sensor(las, h = TRUE)
index(las, h = TRUE)

# Modification of the sensor enables to select a better spatial index
# when processing the point-cloud.
sensor(las) <- TLSLAS
sensor(las, h = TRUE)
index(las, h = TRUE)

# Modification of the spatial index forces to use one of the avaialble
# spatial index.
index(las) <- LIDRQUATREE
sensor(las, h = TRUE)
index(las, h = TRUE)

# The simplest way to take advantage of appropriated spatial indexing is
# to use one of read*LAS().
las <- readTLSLAS(LASfile)
sensor(las, h = TRUE)
index(las, h = TRUE)

# But for some specific point-cloud / algorithm it might be advised to force
# the use of a specific spatial index to perform the computation faster
index(las) <- LIDRVOXELPARTITION
index(las, h = TRUE)

# With a LAScatalog, spatial indexing informations are propagated to the
# different chunks
ctg = readTLSLAScatalog(LASfile)
index(ctg) <- LIDRVOXELPARTITION
sensor(ctg, h = TRUE)
index(ctg, h = TRUE)

}
\keyword{datasets}
