% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.r
\name{lidr_options}
\alias{lidr_options}
\alias{lidr_reset}
\title{Set or get global options for lidR package}
\usage{
lidr_options(...)

lidr_reset()
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get global options for lidR package
}
\section{Supported options}{

The following options are supported:
\itemize{
 \item{\code{verbose} (\code{logical}) Make the package talkative. }
 \item{\code{progress} (\code{logical}) Display progress bars when avaible. }
 \item{\code{multicore} (\code{numeric}) For parallel processe process fix the number of core to use.}
}
}

\examples{
lidr_options(verbose = TRUE)
lidr_options(progress = TRUE)

# Reset default options
lidr_reset()
}
