% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.r
\name{lidrpalettes}
\alias{forest.colors}
\alias{height.colors}
\alias{lidrpalettes}
\alias{pastel.colors}
\alias{random.colors}
\title{Palettes}
\usage{
height.colors(n)

forest.colors(n)

random.colors(n)

pastel.colors(n)
}
\arguments{
\item{n}{The number of colors (> 1) to be in the palette}
}
\description{
Create a vector of n contiguous colors
}
\seealso{
\link[grDevices:heat.colors]{grDevices::Palettes}
}

