% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_tree_detection.R
\name{lmfauto}
\alias{lmfauto}
\title{Individual Tree Detection Algorithm}
\usage{
lmfauto(plot = FALSE, hmin = 2)
}
\arguments{
\item{plot}{logical set it to \code{TRUE} is processing a plot instead of a large area. What change
is the estimation of the local number of trees. It should be based the local neighborhood for general
case but this does not make sense for a plot.}

\item{hmin}{numeric. Minimum height of a tree. Threshold below which a point cannot be a local
maxima. Default is 2.}
}
\description{
This function is made to be used in \link{tree_detection}. It implements a fast and parameter-free
algorithm for individual tree detection on wide coverage. It is based on two local maximum filters
(LMF). The first pass performs a very raw estimation of the number of trees with a fixed windows
size. Based on this raw estimation it automatically compute a variable windows size LMF with workable
parameters. This way this algorithm is parameter-free and properly parametrized for many contexts.
This algorithm is made to process wide areas not small plots. See references for more details.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las <- readLAS(LASfile)
ttops <- tree_detection(las, lmfauto())
}
\references{
Roussel Jean-Romain, Development of a parameter-free algorithm for automatic tree
detection on wide territories (in prep.)
}
\seealso{
Other individual tree detection algorithms: \code{\link{lmf}},
  \code{\link{manual}}
}
\concept{individual tree detection algorithms}
