% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\name{LoadLidar}
\alias{LoadLidar}
\title{Load a las file and create a \code{Lidar} object}
\usage{
LoadLidar(input, fields = "standard", ...)
}
\arguments{
\item{input}{character. filename(s) of .las file(s)}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}}

\item{\dots}{Unused}
}
\value{
An object of the class \code{Lidar}
}
\description{
Methods to read and creates a \code{Lidar} object from a vector of .las filename(s)
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)
}
\seealso{
\link[lidR:Lidar]{Class Lidar}
}

