% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maximum.R
\name{local_maximum}
\alias{local_maximum}
\title{Local Maximum Filter}
\usage{
local_maximum(las, w, filter = NULL)
}
\arguments{
\item{las}{An object of class LAS}

\item{w}{numeric. Window shape. 1 number for the diameter of a disc, 2 numbers for a rectangle
(width, height), 3 numbers for an oriented rectangle (width, height, angle). The angle must be
in radian.}
}
\value{
\code{SpatialPointsDataFrame} with attribute from the corresponding point in the LAS object,
}
\description{
Generic local maximum filter. For individual tree detection use \link{tree_detection} with the
\link{lmf} algorithm that is more adequate for ITD. This function is a more generic method for
multiple purpose other than tree segmentation. This function is natively parallelized with OpenMP
}
