% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_intensity.R
\name{normalize_intensity}
\alias{normalize_intensity}
\title{Normalize intensity}
\usage{
normalize_intensity(las, algorithm)
}
\arguments{
\item{las}{An object of class \link[lidR:LAS-class]{LAS} or \link[lidR:LAScatalog-class]{LAScatalog}.}

\item{algorithm}{an intensity normalizaton algorithm. \code{lidR} currently have \link{range_correction}.}
}
\value{
Returns an object of class LAS. The attribute 'Intensity'
records the normalised intensity. An extra attribute named 'RawIntensity' records the original
intensities.
}
\description{
Normalize intensity value using multiple methods.
}
\examples{
# A valid file properly populated
LASfile <- system.file("extdata", "Topography.laz", package="lidR")
las <- readLAS(LASfile)

# pmin = 15 because it is an extremely tiny file
# strongly decimated to reduce its size. There are
# actually few multiple returns
sensor <- track_sensor(las, pmin = 15)

# Here the effect is virtually null because the size of
# the sample is too tiny to notice any effect of range
las <- normalize_intensity(las, range_correction(sensor, Rs = 2000))
}
\seealso{
Other normalize: 
\code{\link{normalize_elevation}()}
}
\concept{normalize}
