% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{plot,Lidar,ANY-method}
\alias{plot}
\alias{plot,Lidar,ANY-method}
\alias{plot-LiDAR}
\title{Plot LiDAR data}
\usage{
\S4method{plot}{Lidar,ANY}(x, y, color = "Z",
  colorPalette = "height.colors", ...)
}
\arguments{
\item{x}{An object of the class \code{Lidar}}

\item{y}{Unused (inherited from base plot)}

\item{color}{characters. The field used to colorize the points. Default is Z coordinates}

\item{colorPalette}{characters. A color palette name. Default is \code{height.colors} provided by the package lidR}

\item{\dots}{supplementary parameters for \link[rgl:points3d]{points3d}}
}
\description{
Plot in 3D a LiDAR cloud of point
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)

plot(lidar)
plot(lidar, color = "Intensity", colorPalette = "heat.colors")
}
\seealso{
\code{\link[rgl:points3d]{points3d} }
}

