% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLiDARProject.r
\docType{methods}
\name{plot,LiDARProject,ANY-method}
\alias{plot,LiDARProject,ANY-method}
\title{Plot a LiDARProject object}
\usage{
\S4method{plot}{LiDARProject,ANY}(x, y, ...)
}
\arguments{
\item{x}{A LiDARProject object}

\item{y}{Unused (inherited from base plot)}

\item{\dots}{Unused (inherited from base plot)}
}
\description{
Plot a LiDARProject object
}

