% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.catalog.r
\name{plot.Catalog}
\alias{plot.Catalog}
\title{Plot a Catalog object}
\usage{
\method{plot}{Catalog}(x, y, type = "terrain", ...)
}
\arguments{
\item{x}{A Catalog object}

\item{y}{A proj4 string for the location of the catalog. This enable to load a map from
google map and use it as background.}

\item{type}{character. Choose from 'roadmap', 'satellite', 'hybrid', 'terrain'}

\item{\dots}{inherited from base plot}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for Catalog objects
}
\examples{
\dontrun{
ctg = catalog("<Path to a folder containing a set of .las files>")
plot(catalog)

# Exemple with a single file
proj4   <- "+proj=utm +zone=17 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

ctg = catalog(LASfile)
plot(ctg, proj4)
}
}
