% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DTM.r
\name{plot.DTM}
\alias{plot.DTM}
\title{Plot a wireframe of a DTM object}
\usage{
\method{plot}{DTM}(x, y, add = FALSE, bg = "black", ...)
}
\arguments{
\item{x}{An object of the class \code{DTM}}

\item{y}{Unused (inherited from R base)}

\item{add}{logical. if TRUE, add to current 3D plot.}

\item{bg}{The color for the background. Default is black.}

\item{\dots}{Supplementary parameters for \link[rgl:surface3d]{surface3d}}
}
\description{
Plot a wireframe of a DTM object
}

