% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gridmetrics.r
\name{plot.gridmetrics}
\alias{plot.gridmetrics}
\title{Plot an object of class gridmetrics in 2D}
\usage{
\method{plot}{gridmetrics}(x, z = NULL, colorPalette = height.colors, ...)
}
\arguments{
\item{x}{A data.frame or data.table of class gridmetrics.}

\item{z}{character. The field to plot. If NULL, autodetect.}

\item{colorPalette}{function. A color palette function. Default is \code{height.colors} provided by the package lidR}

\item{\dots}{Supplementary parameters for \link[fields:image.plot]{image.plot}}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for a gridmetrics data.frame
}
\details{
The \dots param provides additional arguments to \link[fields:image.plot]{image.plot}.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Canopy surface model with 4 m^2 cells
gridmetrics(lidar, 2, max(Z)) \%>\% plot

# Mean height with 400 m^2 cells
gridmetrics(lidar, 20, mean(Z)) \%>\% plot

# Define your own metric function
myMetrics = function(z, i, angle, pulseID)
{
  ret = list(
        npulse  = length(unique(pulseID)),
        hmean   = mean(z),
        hmax    = max(z),
        imean   = mean(i),
        angle   = mean(abs(angle))
        )

   return(ret)
 }

metrics = gridmetrics(lidar, 20, myMetrics(Z, Intensity, ScanAngle, pulseID))

plot(metrics, "hmean")
plot(metrics, "hmax")
plot(metrics, "imean")
}
\seealso{
\link[lidR:gridmetrics]{gridmetrics}
\link[lidR:canopyModel]{canopyModel}
\link[fields:image.plot]{image.plot}
\link[lidR:height.colors]{height.colors}
\link[lidR:forest.colors]{forest.colors}
\link[grDevices:heat.colors]{heat.colors}
\link[grDevices:colorRamp]{colorRampPalette}
\link[lidR:plot3d]{plot3d}
}

