% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.catalog.r
\name{plot.LAScatalog}
\alias{plot.LAScatalog}
\title{Plot a LAScatalog object}
\usage{
\method{plot}{LAScatalog}(x, y = "terrain", ...)
}
\arguments{
\item{x}{A LAScatalog object}

\item{y}{character. If the catalog have a valid CRS, a google map background will be displayed.
Choose from 'roadmap', 'satellite', 'hybrid', 'terrain', 'none'.}

\item{\dots}{inherited from base plot}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for LAScatalog objects
}
\examples{
\dontrun{
ctg = catalog("<Path to a folder containing a set of .las files>")
plot(catalog)

# Exemple with a single file
proj4   <- "+proj=utm +zone=17"
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

ctg = catalog(LASfile)
plot(ctg, proj4)
}
}
