% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TIN.r
\name{plot.TIN}
\alias{plot.TIN}
\title{Plot triangular irregular network TIN}
\usage{
\method{plot}{TIN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{TIN}}

\item{...}{Unused}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for class TIN
}
\examples{
x = runif(20)
y = runif(20)
z = runif(20)
tin = TIN(x, y, z)
plot(tin)
}
\seealso{
\link[lidR:TIN]{TIN}
}

