% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Do point fall in a given polygon?}
\usage{
point_in_polygon(vertx, verty, pointx, pointy)
}
\arguments{
\item{vertx}{numerical array of x-coordinates of polygon}

\item{verty}{numerical array of y-coordinates of polygon}

\item{pointx}{numeric. x-coordinate of a point}

\item{pointy}{numeric. y-coordinate of a point}
}
\value{
Logical. FALSE, point is exterior to the polygon, TRUE, point is in interior to the polygon
}
\description{
Verifies for one point whether it fall in a given polygon
}
\references{
Adaptation of the C function written by W. Randolph Franklin
}

