% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{points_in_polygons}
\alias{points_in_polygons}
\title{Do points fall in a one of the polygons?}
\usage{
points_in_polygons(vertx, verty, pointx, pointy)
}
\arguments{
\item{vertx}{numerical list of array of x-coordinates of polygon}

\item{verty}{numerical list of array of y-coordinates of polygon}

\item{pointx}{numerical array of x-coordinates of points}

\item{pointy}{numerical array of y-coordinates of points}
}
\value{
numerical array. 0 if the points are in any polygon or the number of the polygon if points fall in a polygon
}
\description{
Verifies for a set of points whether they fall in one of the given polygon
}

