% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointsInPolygon}
\alias{pointsInPolygon}
\title{Do points fall in a given polygon?}
\usage{
pointsInPolygon(vertx, verty, pointx, pointy)
}
\arguments{
\item{vertx}{numerical array of x-coordinates of polygon}

\item{verty}{numerical array of y-coordinates of polygon}

\item{pointx}{numerical array of x-coordinates of points}

\item{pointy}{numerical array of y-coordinates of points}
}
\value{
Logical array. FALSE, point is exterior to the polygon, TRUE, point is in interior to the polygon
}
\description{
Verifies for a set of points whether they fall in a given polygon
}

