% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_area.r
\name{polygon_area}
\alias{polygon_area}
\title{Calculates the area of a polygon}
\usage{
polygon_area(x, y)
}
\arguments{
\item{x}{numeric. x coordinates of the polygon}

\item{y}{numeric. y coordinates of the polygon}
}
\description{
Calculates the area of a polygon
}

