% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonArea.r
\name{polygonArea}
\alias{polygonArea}
\title{Calculates the area of a polygon}
\usage{
polygonArea(x, y)
}
\arguments{
\item{x}{numeric. x coordinates of the polygon}

\item{y}{numeric. y coordinates of the polygon}
}
\description{
Calculates the area of a polygon
}

