% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lascoplanar.r
\name{princomp}
\alias{princomp}
\alias{lascoplanar}
\alias{lascolinear}
\title{Test the shape of the neighborhood}
\usage{
lascoplanar(las, k = 8, th1 = 25, th2 = 6, filter = NULL)

lascolinear(las, k = 8, th1 = 10, filter = NULL)
}
\arguments{
\item{las}{an object of class LAS}

\item{k}{interger. The number of k-nearest neighbors.}

\item{th1}{numeric. The threshold 1 used in labelling criteria}

\item{th2}{numeric. The threshold 2 used in labelling criteria}

\item{filter}{formula of logical predicates. Enable to run the function only on points of interest
in an optimized way. See also examples.}
}
\value{
A LAS object with a new column names \code{Coplanar} or \code{Colinear} that indicates
those points that are part of a neighborhood that is approximately coplanar/colinear (TRUE) or not
(FALSE).
}
\description{
An approximate coplanarity and colinearity tests. Coplanarity test is based on Limberger & Oliveira
(2015). For each point it looks for the k-nearest neighbors. It computes the eigenvalues of the
covariance matrix. The points that meet the following criteria are labeled as approximately coplanar:\cr
\deqn{a_2 > (th_1*a_1) && (th_2*a_2) > a_3}
\eqn{a_1, a_2, a_3} being the eigenvalues of the neighborhood (defined by k-nearest neighbors) in
ascending order.\cr\cr
Colinearity test performs the same computation but change the critera to label point as approximatly
colinear:
\deqn{th_1*a_2 < a_3 && (th_1*a_1) < a_3}
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)

las = lascoplanar(las, k = 20)
plot(las, color = "Coplanar")
}

}
\references{
Limberger, F. A., & Oliveira, M. M. (2015). Real-time detection of planar regions in unorganized
point clouds. Pattern Recognition, 48(6), 2043–2053. https://doi.org/10.1016/j.patcog.2014.12.020
}
