% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print}
\alias{print}
\alias{summary,LAS-method}
\alias{summary}
\alias{print,LAS-method}
\alias{summary,LAScatalog-method}
\title{Summary and Print for \code{LAS*} objects}
\usage{
print(x, ...)

\S4method{summary}{LAS}(object, ...)

\S4method{print}{LAS}(x)

\S4method{summary}{LAScatalog}(object, ...)
}
\arguments{
\item{...}{Unused}

\item{object, x}{A \code{LAS*} object}
}
\value{
NULL, used for its side-effect of printing information
}
\description{
Summary and Print for \code{LAS*} objects
}
