% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{pulseDensity,Lidar-method}
\alias{pulseDensity}
\alias{pulseDensity,Lidar-method}
\title{Pulse density surface model}
\usage{
\S4method{pulseDensity}{Lidar}(obj, resolution = 4)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{resolution}{numeric. The size of a grid cell in LiDAR data coordinates units}
}
\value{
It returns a \code{data.table} with the class \code{gridMetrics} which anable to plot it easily.
}
\description{
Creates a pulse density surface model using a LIDAR point cloud.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = Lidar(LASfile)

lidar \%>\% pulseDensity(5) \%>\% plot
lidar \%>\% pulseDensity(5) \%>\% plot
}
\seealso{
\code{\link[lidR:gridMetrics]{gridMetrics}}
}

