% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseDensity.r
\docType{methods}
\name{pulseDensity}
\alias{pulseDensity}
\alias{pulseDensity,LAS-method}
\title{Pulse density surface model}
\usage{
pulseDensity(obj, res = 4)

\S4method{pulseDensity}{LAS}(obj, res = 4)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 4 units i.e. 16 square units cells.}
}
\value{
It returns a \code{data.table} with the class \code{gridmetrics} which which enables easier plotting.
}
\description{
Creates a pulse density map using a LiDAR cloud of points.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

lidar \%>\% pulseDensity(5) \%>\% plot
lidar \%>\% pulseDensity(10) \%>\% plot
}
\seealso{
\code{\link[lidR:gridmetrics]{gridmetrics}}
}

