% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{random.colors}
\alias{random.colors}
\title{Random palette}
\usage{
random.colors(n)
}
\arguments{
\item{n}{The number of colors (> 1) to be in the palette}
}
\description{
Create a vector of n contiguous random colors.
}
\seealso{
\link[grDevices:colorRamp]{colorRampPalette}
}

