% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_points_decimation.r
\name{random}
\alias{random}
\title{Algorithm for point cloud decimation}
\usage{
random(density, use_pulse = FALSE)
}
\arguments{
\item{density}{numeric. The desired output density.}

\item{use_pulse}{logical. Decimate by removing random pulses instead of random points (requieres to run
\link{laspulse} first)}
}
\description{
Randomly removes points or pulses to reach the desired density over the whole area (see
\code{\link[lidR:area]{area}}).
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile, select = "xyz")

# Reach a pulse density of 1 on the overall dataset
thinned1 = lasfilterdecimate(las, random(1))
plot(grid_density(las))
plot(grid_density(thinned1))
}
\seealso{
Other Algorithm: \code{\link{csf}},
  \code{\link{dalponte2016}}, \code{\link{hamraz2016}},
  \code{\link{highest}}, \code{\link{homogenize}},
  \code{\link{li2012}}, \code{\link{lmf}},
  \code{\link{manual}}, \code{\link{multichm}},
  \code{\link{pmf}}, \code{\link{ptrees}},
  \code{\link{silva2016}}, \code{\link{tin}},
  \code{\link{watershed}}, \code{\link{wing2015}}

Other Point Cloud Decimation: \code{\link{highest}},
  \code{\link{homogenize}}, \code{\link{lasfilterdecimate}}
}
