% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_las.R
\name{rbind}
\alias{rbind}
\alias{rbind.LAS}
\alias{rbind.LAScatalog}
\title{Merge LAS* objects}
\usage{
\method{rbind}{LAS}(...)

\method{rbind}{LAScatalog}(...)
}
\arguments{
\item{...}{LAS or LAScatalog objects}
}
\description{
Merge LAS* objects
}
