% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLAS.r
\name{readLAS}
\alias{readLAS}
\title{Read las files}
\usage{
readLAS(files, Intensity = TRUE, ReturnNumber = TRUE,
  NumberOfReturns = TRUE, ScanDirectionFlag = FALSE,
  EdgeofFlightline = FALSE, Classification = TRUE, ScanAngle = TRUE,
  UserData = FALSE, PointSourceID = FALSE, RGB = TRUE)
}
\arguments{
\item{files}{character.}

\item{Intensity}{logical.}

\item{ReturnNumber}{logical.}

\item{NumberOfReturns}{logical.}

\item{ScanDirectionFlag}{logical.}

\item{EdgeofFlightline}{logical.}

\item{Classification}{logical.}

\item{ScanAngle}{logical.}

\item{UserData}{logical.}

\item{PointSourceID}{logical.}

\item{RGB}{logical.}
}
\value{
A LAS object
}
\description{
Read las files
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = readLAS(LASfile)
}

