% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readLASdata}
\alias{readLASdata}
\title{Read data in a .las file}
\usage{
readLASdata(file, Intensity = TRUE, ReturnNumber = TRUE,
  NumberOfReturns = TRUE, ScanDirectionFlag = FALSE,
  EdgeofFlightline = FALSE, Classification = TRUE, ScanAngle = TRUE,
  UserData = FALSE, PointSourceID = FALSE, RGB = TRUE)
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read .las files content
}

