% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readLASheader}
\alias{readLASheader}
\title{Read header in a .las file}
\usage{
readLASheader(file)
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read the header in a .las file
}

