% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-LAS.r
\docType{methods}
\name{$<-,LAS-method}
\alias{$<-,LAS-method}
\alias{[[<-,Spatial,ANY,missing-method}
\title{Inherited but modified methods from sp}
\usage{
\S4method{$}{LAS}(x, name) <- value

\S4method{[[}{Spatial,ANY,missing}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{LAS*} object}

\item{name}{A literal character string or a name (possibly backtick quoted).}

\item{value}{typically an array-like R object of a similar class as x.}

\item{i}{string, name of elements to extract or replace.}

\item{j}{Unused.}
}
\description{
\code{LAS*} objects being \link[sp:Spatial-class]{Spatial} object they inherit several methods
from \code{sp}. However, some have modified behaviors to forbid some unrelevant modifications. Indeed
a \code{LAS*} object cannot contains anything. The content is restricted by the LAS specifications.
If a user try to use one of these function in a bad way an informative error will be thrown.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)

las$Z = 2L
las[["Z"]] = 1:10
las$NewCol = 0
las[["NewCol"]] = 0
}
}
