% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_index.r
\docType{methods}
\name{roi_index}
\alias{roi_index}
\alias{roi_index,Catalog-method}
\title{Retrieve the tiles containing ROIs}
\usage{
roi_index(obj, x, y, radius, radius2 = NULL, roinames = NULL)

\S4method{roi_index}{Catalog}(obj, x, y, radius, radius2 = NULL,
  roinames = NULL)
}
\arguments{
\item{obj}{A Catalog object}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{radius}{numeric or vector. A radius or a set of radii of the ROI. If only
radius is provided (radius2 = NULL) it will extract data falling into a disc.}

\item{radius2}{numeric or vector. A radius or a set of radii of plots. If radius2
is provided, the selection turns into a rectangular ROI. If radius = radius2 it is a square obviouly.}

\item{roinames}{vector. A set of ROI names}
}
\description{
When the user has a set of (x, y) coordinates corresponding to a region of interest (ROI)
(a ground inventory for example), he can automatically find in which tiles the lidar data associated
to the ROIs from a Catalog are. The algorithm automatically find tiles even for ROIs falling on the edges
of one or more tiles.\cr
It works only for tiles well organized as a damier. This function is used by \link[lidR:roi_query]{roi_query}.
Users do not really need it.
}

