% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_select.r
\docType{methods}
\name{roi_select}
\alias{roi_select}
\alias{roi_select,LAS-method}
\title{Select an region of interest interactively}
\usage{
roi_select(obj, ...)

\S4method{roi_select}{LAS}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{\dots}{Optionnal parameters for the plot function}
}
\value{
An object of class \code{LAS}
}
\description{
Select an region of interest contained into a rectangle using interactively the mouse
}
\details{
\code{roi_select} enable the user to select a region of interest (ROI) drawing
a rectangle with the mouse
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

subset = roi_select(lidar)
}
}

