% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_canopy_roughtness.r
\name{rumple_index}
\alias{rumple_index}
\title{Rumple index roughness}
\usage{
rumple_index(x, res)
}
\arguments{
\item{x}{A matrix or a RasterLayer. The surface.}

\item{res}{numeric the resolution of the surface}
}
\value{
numeric.
}
\description{
Computes the roughness of a surface represented by a matrix as the ratio between visible area
over the projected area.
}
\examples{
# Perfectly flat surface, rumple_index = 1
surface = matrix(rep(2, 100), 10)
rumple_index(surface, 1)

# Rought surface, rumple_index > 1
surface = matrix(runif(100), 10)
rumple_index(surface, 1)

# This measure of roughness is scale dependent
rumple_index(surface, 1)
rumple_index(surface, 2)
}
