% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArea.r
\docType{methods}
\name{selectArea}
\alias{selectArea}
\alias{selectArea,Lidar-method}
\title{Select an region of interest interactively}
\usage{
selectArea(obj, ...)

\S4method{selectArea}{Lidar}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{\dots}{Optionnal parameters for the plot function}
}
\value{
An object of class \code{Lidar}
}
\description{
Select an region of interest contained into a rectangle using interactively the mouse
}
\details{
\code{selectArea} enable the user to select a region of interest (ROI) drawing
a rectangle with the mouse
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

subset = selectArea(lidar)
}
}

