% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{set.colors}
\alias{set.colors}
\title{Automatic colorization}
\usage{
set.colors(x, palette, ncolors = 50, trim = 1)
}
\arguments{
\item{x}{A vector}

\item{palette}{function. A color palette function. Default is \code{height.colors} provided by the package lidR}

\item{ncolors}{numeric. The number of colors in the palette.}

\item{trim}{numeric.}
}
\description{
Attribute a color to each element of a vector
}

