% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_spatial_interpolations.r
\name{tin}
\alias{tin}
\alias{SpatialInterpolationFunctions}
\alias{knnidw}
\alias{kriging}
\title{Algorithms for spatial interpolation}
\usage{
tin()

knnidw(k = 10, p = 2)

kriging(model = gstat::vgm(0.59, "Sph", 874), k = 10L)
}
\arguments{
\item{k}{numeric. Number of k-nearest neighbours. Default 10.}

\item{p}{numeric. Power for inverse distance weighting. Default 2.}

\item{model}{A variogram model computed with \link[gstat:vgm]{vgm}. If NULL it performs an ordinary
or weighted least squares prediction.}
}
\description{
These functions are made to be used in \link{lasnormalize} or \link{grid_terrain} (see details).
They implement spatial interpolation algorithms.\cr
\describe{
\item{\code{knnidw}}{Spatial interpolation is done using a k-nearest neighbour (KNN) approach with
an inverse distance weighting (IDW). This is a fast but basic method for spatial data interpolation.}
\item{\code{tin}}{Spatial interpolation is based on a Delaunay triangulation. It performs a linear
interpolation within each triangle. There are usually a few points outside the convex hull,
determined by the ground points at the very edge of the dataset, which cannot be interpolated
with a triangulation. Extrapolation is done using the nearest neighbour.}
\item{\code{kriging}}{Spatial interpolation based on universal kriging using the \link[gstat:krige]{krige}
function from gstat. This method combines the KNN approach with the kriging approach. For each
point of interest it kriges the terrain using the k-nearest neighbour ground points. This method
is more difficult to manipulate but it is also the most advanced method for interpolating spatial data. }
}
}
\seealso{
Other Algorithm: \code{\link{csf}},
  \code{\link{dalponte2016}}, \code{\link{hamraz2016}},
  \code{\link{highest}}, \code{\link{homogenize}},
  \code{\link{li2012}}, \code{\link{lmf}},
  \code{\link{manual}}, \code{\link{multichm}},
  \code{\link{pmf}}, \code{\link{ptrees}},
  \code{\link{random}}, \code{\link{silva2016}},
  \code{\link{watershed}}, \code{\link{wing2015}}
}
