% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_hull.R
\name{st_convex_hull}
\alias{st_convex_hull}
\alias{st_concave_hull}
\alias{st_convex_hull.LAS}
\alias{st_concave_hull.LAS}
\title{Concave and convex hulls for LAS objects}
\usage{
st_concave_hull(x, concavity = 2, length_threshold = 5)

\method{st_convex_hull}{LAS}(x)

\method{st_concave_hull}{LAS}(x, concavity = 2, length_threshold = 5)
}
\arguments{
\item{x}{An object of class LAS.}

\item{concavity, length_threshold}{see \link{concaveman}.}
}
\value{
A sfc_POLYGON from sf
}
\description{
Concave and convex hulls for LAS objects. \code{st_convex_hull} extends \code{sf::st_convex_hull} for LAS
objects, Both functions return a polygon geometry.
}
