% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\docType{methods}
\name{[,Catalog,ANY,ANY,ANY-method}
\alias{[,Catalog,ANY,ANY,ANY-method}
\title{Extract parts of an Catalog object}
\usage{
\S4method{[}{Catalog,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing)}
}
\description{
Extract parts of an Catalog object
}

