% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-LAS.r
\docType{methods}
\name{summary,LAS-method}
\alias{summary,LAS-method}
\alias{summary}
\alias{print,LAS-method}
\title{Summary and Print for \code{LAS*} object}
\usage{
\S4method{summary}{LAS}(object, ...)

\S4method{print}{LAS}(x)
}
\arguments{
\item{object, x}{A \code{LAS*} object}

\item{...}{Unused}
}
\value{
NULL, used for its side effect of printing informations
}
\description{
Summarize a \code{LAS*} object.
}
