% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\docType{methods}
\name{summary,Lidar-method}
\alias{summary}
\alias{summary,Lidar-method}
\title{Summary of Lidar data}
\usage{
\S4method{summary}{Lidar}(object, ...)
}
\arguments{
\item{object}{An object of the class \code{Lidar}}

\item{\dots}{Unused}
}
\description{
Summary of Lidar data
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = Lidar(LASfile)

summary(lidar)
}

