% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-las.r
\name{summary.LAS}
\alias{summary.LAS}
\title{LiDAR data summary}
\usage{
\method{summary}{LAS}(object, ...)
}
\arguments{
\item{object}{A LAS object}

\item{...}{unused}
}
\description{
Print a summary of a LAS object
}
