% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\name{summary.Lidar}
\alias{summary}
\alias{summary.Lidar}
\title{Summary of Lidar data}
\usage{
\method{summary}{Lidar}(object, ...)
}
\arguments{
\item{object}{An object of the class \code{Lidar}}

\item{\dots}{Unused}
}
\description{
This functions implements a \link[base:summary]{summary} method for raster image
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

summary(lidar)
}

