% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_delaunay.R
\name{tDelaunay}
\alias{tDelaunay}
\alias{tInterpolate}
\alias{tSearch}
\alias{tInfo}
\title{Delaunay triangulation tools}
\usage{
tDelaunay(P, trim = 0, option = "QbB")

tInterpolate(D, P, X, threads = 1L)

tSearch(D, P, X, threads = 1L)

tInfo(D, P)
}
\arguments{
\item{P}{matrix. Points triangulated or to be triangulated.}

\item{trim}{numeric. If \code{trim > 0}, removes all the triangles that have the longest edge greater than
this value. If \code{trim < 0}, the same but smaller than this value.}

\item{D}{matrix. Triangulation}

\item{X}{matrix. Points to test.}

\item{threads}{integer. Number of OpenMP threads.}
}
\description{
Internal functions: a set of tools to work with Delaunay triangulation.
\itemize{
\item \code{tDelaunay}: Delaunay triangulation (DT) with triangle size filter
\item \code{tInterpolate}: linear interpolation within DT at coordinates X.
\item \code{tSerach}: finds the index of the triangule in DT containing the points X.
\item \code{tInfo}: compute informations for each triangle including normal vector, area, longest
edge.
}
}
\examples{
P <- matrix(runif(30, 0, 10), ncol = 3)
X <- matrix(runif(20, 0, 10), ncol = 2)
D <- tDelaunay(P, trim = -4)
geometry::trimesh(D, P)
points(X)
tInfo(D,P)
tInterpolate(D, P, X)
}
\keyword{internal}
