% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTIN.r
\name{TIN}
\alias{TIN}
\title{Triangular irregular network}
\usage{
TIN(x, y, z)
}
\arguments{
\item{x}{numeric. n points x coordinates}

\item{y}{numeric. n points y coordinates}

\item{z}{numeric. n points z coordinates}
}
\value{
A data.frame containig the list of triangles coordinates in each row
}
\description{
Triangular irregular network computed from Delaunay triangulation
}
\examples{
x = runif(20)
y = runif(20)
z = runif(20)
tin = TIN(x, y, z)
plot(tin)
}
\seealso{
\link[geometry:delaunayn]{delaunayn}
}

