% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/utils_base.R
\name{print.LAS}
\alias{print.LAS}
\alias{print.LAScatalog}
\alias{print.lidRAlgorithm}
\alias{print.raster_template}
\alias{summary.LAS}
\alias{summary.LAScatalog}
\alias{tools}
\alias{dim.LAS}
\alias{dim.LAScatalog}
\alias{ncol.LAS}
\alias{nrow.LAScatalog}
\alias{names.LAS}
\alias{names.LASheader}
\alias{rbind.LAS}
\alias{npoints}
\alias{density}
\alias{density,LAS-method}
\alias{density,LASheader-method}
\alias{density,LAScatalog-method}
\title{Tools inherited from base R for LAS* objects}
\usage{
\method{print}{LAS}(x, ...)

\method{print}{LAScatalog}(x, ...)

\method{print}{lidRAlgorithm}(x, ...)

\method{print}{raster_template}(x, ...)

\method{summary}{LAS}(object, ...)

\method{summary}{LAScatalog}(object, ...)

\method{dim}{LAS}(x)

\method{dim}{LAScatalog}(x)

ncol.LAS(x)

nrow.LAScatalog(x)

\method{names}{LAS}(x)

\method{names}{LASheader}(x)

\method{rbind}{LAS}(...)

npoints(x, ...)

density(x, ...)

\S4method{density}{LAS}(x, ...)

\S4method{density}{LASheader}(x, ...)

\S4method{density}{LAScatalog}(x, ...)
}
\arguments{
\item{x}{a LAS* object}

\item{...}{LAS* objects if it is the sole argurment (e.g. in rbind())}

\item{object}{A \code{LAS*} object or other lidR related objects.}
}
\value{
Use for side effect of printing
}
\description{
Tools inherited from base R for LAS* objects
}
