% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_detection.r
\name{tree_detection_manual}
\alias{tree_detection_manual}
\title{Tree top detection based on manual selection}
\usage{
tree_detection_manual(las, detected = NULL, ...)
}
\arguments{
\item{las}{An object of the class LAS}

\item{detected}{\code{SpatialPointsDataFrame} or \code{data.table} or \code{data.frame} or \code{matrix}
containing X,Y,Z coordinates of already found tree tops that need manual corrections.}

\item{...}{supplementary parameters to be pass to \link{plot.LAS}.}
}
\value{
A \code{SpatialPointsDataFrame} with the 2D coordinates of the tree tops (X, Y) and the elevation of
the apex and the id of each tree in the table of attributes (attributes \code{Z} and \code{treeID} respectively)
}
\description{
Find the tree top positions manually and interactively using the mouse. This is only suitable for
small to medium size plots. First the point cloud is displayed, then the user is invited to select
a rectangular region of interest in the scene using the right button of the mouse. Within the selected
points the highest will be flaged as 'tree top' in the scene. Once all the tree are labelled the user can
exit the tools by selecting an empty region. Points can also be unflagged.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile)

# Full manual tree finding
ttops = tree_detection_manual(las)

# Automatic finding with manual correction
ttops = tree_detection_lmf(las, 5)
ttops = tree_detection_manual(las, ttops)
}
}
