% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.r
\name{voxelize}
\alias{voxelize}
\title{convert lidar point cloud into voxels}
\usage{
voxelize(pointcloud, voxel_size = 1)
}
\arguments{
\item{pointcloud}{lidar pointcloud, object of class "Lidar".}

\item{voxel_size}{size of the voxels}
}
\value{
an object of class "data.table" with coordinates of the voxels.
}
\description{
convert lidar point cloud into voxels
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar <-  LoadLidar(LASfile)
lidar_vox <- voxelize(lidar)
}
\author{
Piotr Tompalski
}

