% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-LAS.r, R/LASmethods-LAScatalog.r
\docType{methods}
\name{area,LAS-method}
\alias{area,LAS-method}
\alias{area,LAScatalog-method}
\title{Surface covered by a \code{LAS*} object.}
\usage{
\S4method{area}{LAS}(x, ...)

\S4method{area}{LAScatalog}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{LAS*}}

\item{...}{unused}
}
\value{
numeric. The area of the object computed in the same units as the coordinate reference system
}
\description{
Surface covered by a \code{LAS*} object.
}
