% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{area}
\alias{area}
\alias{area,Lidar-method}
\title{Area of a Lidar object}
\usage{
area(obj)

\S4method{area}{Lidar}(obj)
}
\arguments{
\item{obj}{An object of the class \code{Lidar}}
}
\value{
numeric. The area of the object computed with a convex hull in Lidar coordinates units
}
\description{
Retrieve the area of a Lidar object.
}
\details{
area is computed with a convex hull. It is only an approximation if the shape of
the data is not convex.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

area(lidar)
}
\seealso{
\code{\link[lidR:convexHull]{convexHull} }
\code{\link[lidR:polygonArea]{polygonArea} }
}

