% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.r
\docType{methods}
\name{area}
\alias{area}
\alias{area,LAS-method}
\title{Area of a LAS object}
\usage{
area(obj)

\S4method{area}{LAS}(obj)
}
\arguments{
\item{obj}{An object of the class \code{LAS}}
}
\value{
numeric. The area of the object computed with a convex hull in LAS coordinates units
}
\description{
Retrieve the area of a LAS object.
}
\details{
area is computed with a convex hull. It is only an approximation if the shape of
the data is not convex.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

area(lidar)
}
\seealso{
\code{\link[lidR:convexHull]{convexHull} }
\code{\link[lidR:polygonArea]{polygonArea} }
}

