% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-generic.r, R/LASmethods-LAS.r,
%   R/LASmethods-LAScatalog.r
\docType{methods}
\name{area}
\alias{area}
\alias{area,LAS-method}
\alias{area,LAScatalog-method}
\title{Surface covered by a LAS* object.}
\usage{
area(x, ...)

\S4method{area}{LAS}(x, ...)

\S4method{area}{LAScatalog}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{LAS*}}

\item{...}{unused}
}
\value{
numeric. The area of the object computed in the same units as the coordinate reference system
}
\description{
Surface covered by a \code{LAS*} object. For \code{LAS} point clouds it is computed based on the
convex hull of the points. For a \code{LAScatalog} it is computed as the sum of the bounding boxes
of the files. For overlapping tiles the value may be larger than the total covered area because
some regions are sampled twice.
}
