% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.gridmetrics}
\alias{as.gridmetrics}
\title{Set the class gridmetrics to a data.frame or a data.table}
\usage{
as.gridmetrics(x, res)
}
\arguments{
\item{x}{A data.frame or a data.table}

\item{res}{numeric the original resolution}
}
\description{
Set the class gridmetrics to a data.frame. Useful when reading data from a file.
In this case the data.frame does not have the class gridmetrics and cannot easly be
plotted or transformed into a raster
}

