% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-LASheader.R
\name{as.list.LASheader}
\alias{as.list.LASheader}
\title{Transform to a list}
\usage{
\method{as.list}{LASheader}(x, ...)
}
\arguments{
\item{x}{A LASheader object}

\item{...}{unused}
}
\description{
Functions to construct, coerce and check for both kinds of R lists.
}
