% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r
\name{as.matrix.gridMetrics}
\alias{as.matrix.gridMetrics}
\title{Tranform a gridMetrics into a spatial matrix}
\usage{
\method{as.matrix}{gridMetrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{a gridMetrics object}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{\dots}{Other parameters for \code{acast}}
}
\description{
Tranform a gridMetrics into a spatial matrix
}
\seealso{
\link[reshape2:acast]{acast}
}

