% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_typecast.r
\name{as.spatial}
\alias{as.spatial}
\title{Transform a lidR object into sp object}
\usage{
as.spatial(x)
}
\arguments{
\item{x}{an object from the lidR package}
}
\value{
An object from sp
}
\description{
LAS, LAScatalog, lasmetrics are transformed respectively into SpatialPointsDataFrame,
SpatialPolygonsDataFrame, SpatialPixelsDataFrame
}
\seealso{
Other cast: \code{\link{as.lasmetrics}},
  \code{\link{as.raster.lasmetrics}}
}
