% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.SpatialPixelsDataFrame}
\alias{as.SpatialPixelsDataFrame}
\title{Transform a \code{'lasmetrics'} object into a SpatialPixelsDataFrame object}
\usage{
as.SpatialPixelsDataFrame(.data)
}
\arguments{
\item{.data}{an object of class \code{lasmetrics}}
}
\value{
An object of class \code{SpatialPixelDataFrame}
}
\description{
Transform a \code{'lasmetrics'} object into a SpatialPixelsDataFrame object
}
\seealso{
Other cast: \code{\link{as.SpatialPointsDataFrame}},
  \code{\link{as.lasmetrics}},
  \code{\link{as.raster.lasmetrics}}
}

