% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.r
\name{as.SpatialPointsDataFrame}
\alias{as.SpatialPointsDataFrame}
\title{Transform a LAS object into a SpatialPointsDataFrame object}
\usage{
as.SpatialPointsDataFrame(.las)
}
\arguments{
\item{.las}{an object of class LAS}
}
\value{
An object of class \code{SpatialPointsDataFrame}
}
\description{
Transform a LAS object into a SpatialPointsDataFrame object
}
\seealso{
Other cast: \code{\link{as.SpatialPixelsDataFrame}},
  \code{\link{as.lasmetrics}},
  \code{\link{as.raster.lasmetrics}}
}

