% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyModel.r
\docType{methods}
\name{canopyModel}
\alias{canopyModel}
\alias{canopyModel,LAS-method}
\title{Canopy surface model}
\usage{
canopyModel(obj, res = 2, start = c(0, 0))

\S4method{canopyModel}{LAS}(obj, res = 2, start = c(0, 0))
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 2 units i.e. 4 square units cells.}

\item{start}{vector of x and y coordinates for the reference raster. Default is (0,0) see \link[lidR:gridmetrics]{gridmetrics}}
}
\value{
It returns a \code{data.table} with the class \code{gridmetrics} which enables easier plotting.
}
\description{
Creates a canopy surface model using a LiDAR cloud of points.
}
\details{
The algorithm used is the local maximum algorithm. It assigns the
elevation of the highest return within each grid cell to the grid cell center.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Local maximum algorithm with a resolution of 2 meters
lidar \%>\% canopyModel(2) \%>\% plot
lidar \%>\% canopyModel(2) \%>\% plot3d
}
\seealso{
\code{\link[lidR:gridmetrics]{gridmetrics}}
\code{\link[lidR:clipCircle]{clipCircle}}
}

