% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\docType{class}
\name{Catalog-class}
\alias{Catalog-class}
\title{An S4 class to represent a catalog of las tiles.}
\description{
An S4 class to represent a set of las or laz tiles, to plot them and to process them with
multicore.
}
\details{
A \code{Catalog} object contains a \code{data.frame} in the slot \code{@headers} with the data
read from the headers of all user's .las or .laz files. A catalog is the representation
of a set of files. A computer cannot load all the data at thet same time. A catalog
is a simple way to manage all the files sequentially reading only the headers. See the
public documentation of las format specification for more informations.
}
\section{Slots}{

\describe{
\item{\code{headers}}{data.frame. A table representing the las header data}
}}
\seealso{
\link[lidR:Catalog]{Catalog}
\link[lidR:plot.Catalog]{plot}
\link[lidR:process_parallel]{process_parallel}
\link[lidR:roi_query]{roi_query}
}

