% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\name{Catalog}
\alias{Catalog}
\title{Build a Catalog object}
\usage{
Catalog(folder, ...)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of .las files}

\item{\dots}{Unused}
}
\value{
A Catalog object
}
\description{
Build a Catalog object
}

