% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\name{Catalog}
\alias{Catalog}
\title{Build a Catalog object}
\usage{
Catalog(folder, ...)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of .las files}

\item{\dots}{Unused}
}
\value{
A Catalog object
}
\description{
Methods to creates a \code{Catalog} object from a folder name
}
\details{
A catalog is the representation of a set of las files. A computer cannot load all
the data at the same time. A catalog is a simple way to manage all the file sequentially
reading only the headers.
}
\seealso{
\link[lidR:Catalog]{Catalog-class}
\link[lidR:plot.Catalog]{plot}
\link[lidR:processParallel]{processParallel}
\link[lidR:extractGroundInventory]{extractGroundInventory}
}

