% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-LAScatalog.r,
%   R/utils_catalog_options.r
\name{catalog}
\alias{catalog}
\alias{by_file}
\alias{by_file<-}
\alias{buffer.LAScatalog}
\alias{buffer<-}
\alias{tiling_size}
\alias{tiling_size<-}
\alias{alignement}
\alias{alignement<-}
\alias{cores}
\alias{cores<-}
\alias{progress}
\alias{progress<-}
\alias{stop_early}
\alias{stop_early<-}
\alias{output_files<-}
\alias{output_files}
\alias{laz_compression}
\alias{laz_compression<-}
\alias{use_select}
\alias{use_select<-}
\alias{use_filter}
\alias{use_filter<-}
\title{Create an object of class \link[lidR:LAScatalog-class]{LAScatalog}}
\usage{
catalog(folder, ...)

by_file(ctg)

by_file(ctg) <- value

buffer.LAScatalog(ctg)

buffer(ctg) <- value

tiling_size(ctg)

tiling_size(ctg) <- value

alignement(ctg)

alignement(ctg) <- value

cores(ctg)

cores(ctg) <- value

progress(ctg)

progress(ctg) <- value

stop_early(ctg)

stop_early(ctg) <- value

output_files(ctg) <- value

output_files(ctg)

laz_compression(ctg)

laz_compression(ctg) <- value

use_select(ctg)

use_select(ctg) <- value

use_filter(ctg)

use_filter(ctg) <- value
}
\arguments{
\item{folder}{string. The path of a folder containing a set of las/laz files. Can also be a vector of
file paths.}

\item{\dots}{Extra parameters to \link[base:list.files]{list.files}. Typically `recursive = TRUE`.}

\item{ctg}{A LAScatalog object.}

\item{value}{An appropriated value for catalog settings. See \link[lidR:LAScatalog-class]{LAScatalog}}
}
\value{
A \code{LAScatalog} object
}
\description{
Create an object of class \link[lidR:LAScatalog-class]{LAScatalog} from a folder or a set of filename.
A LAScatalog is a representation of a set of las/laz files. A computer cannot load all the data at
once. A \code{LAScatalog} is a simple way to manage all the files sequentially. Most function from
\code{lidR} can seamlessly be used with a LAScatalog using the internal \code{LAScatalog} processing
engine. To take advantage of the \code{LAScatalog} processing engine the user must first adjust some
processing options using the appropriated functions. The careful reading of the
\link[lidR:LAScatalog-class]{LAScatalog class documentation} is requiered to use the \code{LAScatalog}
class correclty.
}
\examples{
# A single file LAScatalog using data provided with the package
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
ctg = catalog(LASfile)
plot(ctg)

\dontrun{
ctg <- catalog("/path/to/a/folder/of/las/files")

# Internal engine will compute in parallel using two cores
cores(ctg) <- 2L

# Internal engine will process sequentially regions of interest of 500 x 500 m (clusters)
tiling_size(ctg) <- 500

# Internal engine will align the 500 x 500 m clusters on x = 250 and y = 300
alignement(ctg) <- c(250, 300)

# Internal engine will not display any progress estimation
progress(ctg) <- FALSE

# Internal engine will not return result into R. Instead it will write results in files
output_files(ctg) <- "/path/to/folder/templated_filename_{XBOTTOM}_{ID}"

# More details in the documentation
help("LAScatalog-class", "lidR")
}
}
