% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lascatalog.r
\name{catalog}
\alias{catalog}
\title{Build a catalog of las tiles/files}
\usage{
catalog(folder, ...)
}
\arguments{
\item{folder}{string. The path of a folder containing a set of .las files}

\item{\dots}{Extra parameters to \link[base:list.files]{list.files}. Typically `recursive = TRUE`.}
}
\value{
A \code{LAScatalog} object
}
\description{
Build a \link[lidR:LAScatalog-class]{LAScatalog} object from a folder name. A catalog is the representation of a set
of las files. A computer cannot load all the data at the same time. A catalog is a simple
way to manage all the file sequentially reading only the headers.
}
\seealso{
\link{LAScatalog-class}
\link[lidR:plot.LAScatalog]{plot}
\link{catalog_apply}
\link{catalog_queries}
}
