% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_index.r
\name{catalog_index}
\alias{catalog_index}
\title{Retrieve the files containing ROIs}
\usage{
catalog_index(catalog, x, y, w, h, buffer, roinames)
}
\arguments{
\item{catalog}{A LAScatalog object}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{roinames}{vector. A set of ROI names}

\item{r}{numeric or vector. A radius or a set of radii of the ROI. If only
r is provided (r2 = NULL) it will extract data falling onto a disc.}

\item{r2}{numeric or vector. A radius or a set of radii of plots. If r2
is provided, the selection turns into a rectangular ROI. If r= r2 it is a square.}
}
\value{
A list of LAScluster object.
}
\description{
\bold{This function is an internal function reserved for lidR developers.\cr\cr}
When we have a set of (x, y) coordinates corresponding to a region of interest (ROI) in a catalog
this function retrieve in which file(s) are these ROIs.
}
\keyword{internal}
