% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_intersect.R
\name{catalog_intersect}
\alias{catalog_intersect}
\title{Subset a LAScatalog with a spatial object}
\usage{
catalog_intersect(ctg, y)
}
\arguments{
\item{ctg}{A \link[=LAScatalog-class]{LAScatalog} object}

\item{y}{Extent, Raster*, Spatial*, sf, sfc, Extent or bbox objects}
}
\value{
A LAScatalog
}
\description{
Subset a LAScatalog with a spatial object to keep only the tiles of interest. It can be
used to select tiles of interest that encompass spatial objects such as Spatial* objects,
Raster* objects or sf and sfc objects
}
