% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.r
\name{catalog_options}
\alias{catalog_options}
\alias{catalog_reset}
\title{Set or get global options for \link{catalog} tools}
\usage{
catalog_options(...)

catalog_reset()
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set
options.}
}
\description{
Set or get global options for \link{catalog} tools
}
\section{Supported options}{

The following options are supported:
\itemize{
 \item{\code{return_virtual_raster} (logical) functions which return raster-alike
 data such as \link{grid_metrics}, \link{grid_terrain} and other \code{grid_*} functions
 may return huge amount of data for large catalog or hight resolution (typically
 \code{grid_terrain} with a resolution of 1 meter). Switching this options enable to
 store the data on the hard disk and return a light weight virtual raster mosaic.}
 \item{\code{buffer} (numeric) when applying a function to an entiere catalog
 processing sequentially sub-areas some algotihms (such as \link{grid_terrain}) requiere
 a buffer around the area. Default is 30 m.}
 \item{\code{multicore} (numeric) for parallel processes, fix the number of
 core to use. Default is the number of core you have.}
 \item{\code{memory_limit_warning} (numeric) when applying a function to an entiere
 catalog, an internal function try to estimate the size of the output before to run the
 algorithm in attempt to prevent memory overflow. This value (in bytes) is the threshold
 before to get a warning. Set to \code{Inf} to disable.},
 \item{\code{tiling_size} (numeric) to process an entiere catalog, the algorithm split the
 dataset in several squqre subareas to process them sequentially. This is the size of a square.
 Defaul is 1000 (1 km^2).}
 }
}

\examples{
catalog_options(multicore = 2)
catalog_options(buffer = 40)
catalog_options()

# Reset default options
catalog_reset()
}
