% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_retile.r
\name{catalog_retile}
\alias{catalog_retile}
\title{Retile a catalog}
\usage{
catalog_retile(ctg)
}
\arguments{
\item{ctg}{A \link[lidR:catalog]{LAScatalog} object}
}
\value{
A new \code{LAScatalog} object
}
\description{
Splits or merges files to reshape the original catalog files (.las or .laz) into smaller or larger
files. It also enables the addition or removal of a buffer around the tiles. The new files are written
in a dedicated folder. The function first displays the layout of the new tiling pattern and then asks
the user to validate the command.
}
\details{
Internally the function reads and writes the clusters defined by the internal processing options of a
\link[lidR:LAScatalog-class]{LAScatalog} (see also \link{catalog}). Thus the function is flexible and
enables the user to retile the dataset, retile while adding or removing a buffer (negative buffers are
allowed), or optionally to compress the data by retiling without changing the pattern but by changing
the format (las/laz).\cr\cr

Note that this function is not actually very useful since \code{lidR} manages everything
(clipping, processing, buffering, ...) internally using the proper options. Thus, retiling may be useful
for working in other software for example, but not in \code{lidR}.
}
\examples{
\dontrun{
ctg = catalog("path/to/catalog")

# Create a new set of .las files 500 by 500 wide in the folder
# path/to/new/catalog/ and iteratively named Forest_1.las, Forest_2.las
# Forest_3.las, and so on.

set_buffer(ctg) <- 0
set_tiling_size(ctg) <- 500
set_output_files(ctg) <- "path/to/new/catalog/Forest_{ID}
newctg = catalog_retile(ctg)

# Create a new set of .las files equivalent to the original one
# but extended with a 50 m buffer in the folder path/to/new/catalog/
# and iteratively named named after the original files.

set_buffer(ctg) <- 50
set_tiling_size(ctg) <- 0
set_output_files(ctg) <- "path/to/new/catalog/{ORIGINALFILENAME}_buffered
newctg = catalog_retile(ctg)

# Being flexible this function can also compress a catalog but this is
# not really useful since laszip from LAStools is a free and open source
# program.

set_buffer(ctg) <- 0
set_tiling_size(ctg) <- 0
set_laz_compression(ctg) <- TRUE
newctg = catalog_retile(ctg)
}
}
\seealso{
\link{catalog}
}
