% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_retile.r
\name{catalog_retile}
\alias{catalog_retile}
\title{Retile a catalog}
\usage{
catalog_retile(ctg, path, prefix, ext = c("las", "laz"))
}
\arguments{
\item{ctg}{A \link[lidR:catalog]{LAScatalog} object}

\item{path}{string. The folder where the new files should be saved.}

\item{prefix}{character. The initial part of the name of the written files. It can be missing and if
the catalog is processed by files the original name of the file will be retained.}

\item{ext}{character. The format of the written files. Can be ".las" or ".laz".}
}
\value{
A new \code{LAScatalog} object
}
\description{
This function splits or merges files to reshape the original catalog files (.las or .laz)
into smaller or larger files. It also enables to add or remove buffer around the tiles.
The new files are written in a dedicated folder. The function first displays the pattern of the new
tiling pattern and then asks the user to validate the command.
}
\details{
The internal processing options of a \code{LAScatalog} drive the way the catalog is retiled. See
\link{catalog}
}
\examples{
\dontrun{
ctg = catalog("path/to/catalog")

# Create a new set of .las files 500 by 500 wide in the folder
# path/to/new/catalog/ and iteratively named Forest_001.las, Forest_002.las
# Forest_003.las, and so on.

buffer(ctg) = 0
by_file(ctg) = FALSE
tiling_size(ctg) = 500
newctg = catalog_reshape(ctg, "path/to/new/catalog", "Forest_")

# Create a new set of .las files equivalent to the original one but extended with
# a 50 m buffer in the folder path/to/new/catalog/ and iteratively named Forest_001.las, Forest_002.las
# Forest_003.las, and so on.

buffer(ctg) = 50
by_file(ctg) = TRUE
newctg = catalog_reshape(ctg, "path/to/new/catalog", "Forest_")
}
}
\seealso{
\link{catalog}
}
