% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_select.r
\name{catalog_select}
\alias{catalog_select}
\title{Select LAS files interactively}
\usage{
catalog_select(x)
}
\arguments{
\item{x}{A Catalog object}
}
\value{
A Catalog object
}
\description{
Select a set of LAS tile from a Catalog using interactively the mouse
}
\details{
This function enable the user to select a set of las file from a Catalog clicking on the map of the file with the mouse.
The selected files turn red on the plot at the end of the selection.
}
\examples{
\dontrun{
catalog = lascatalog("<Path to a folder containing a set of .las files>")
selectedFiles = catalog_select(catalog)
}
}
\seealso{
\link[lidR:lascatalog]{Catalog}
}

