% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_select.r
\name{catalog_select}
\alias{catalog_select}
\title{Select LAS files interactively}
\usage{
catalog_select(x, Rbase = FALSE)
}
\arguments{
\item{x}{A LAScatalog object}

\item{Rbase}{logical. If TRUE, will use R base plot (no pan, no zoom and not convenient).}
}
\value{
A LAScatalog object
}
\description{
Select a set of LAS tiles from a LAScatalog using the mouse interactively. This function
enables the user to select a set of las files from a LAScatalog by clicking
on the map of the file using the mouse. The selected files will be highlighted in red on
the plot after selection is complete.
}
\examples{
\dontrun{
project = catalog("<Path to a folder containing a set of .las files>")
selectedFiles = catalog_select(project)
}
}
\seealso{
\link[lidR:catalog]{LAScatalog}
}
