% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_from_shapefile.r
\docType{methods}
\name{classify_from_shapefile}
\alias{classify_from_shapefile}
\alias{classify_from_shapefile,LAS,SpatialPolygonsDataFrame-method}
\title{Classify LiDAR points from the polygons in a shapefile}
\usage{
classify_from_shapefile(obj, shapefile, field = NULL)

\S4method{classify_from_shapefile}{LAS,SpatialPolygonsDataFrame}(obj, shapefile,
  field = NULL)
}
\arguments{
\item{obj}{An object of the class \code{LAS}}

\item{shapefile}{An object of class SpatialPolygonsDataFrame}

\item{field}{characters. The name of a field of the shapefile or the name of the new field in the LAS object.}
}
\value{
Nothing. The new field is added by reference in the original data.
}
\description{
Classify LAS points based on geographic data found in a shapefile. It checks
if the LiDAR points are in polygons given in the shapefile. If the parameter
\code{field} is the name of a field in the attribute data of shapefile it
classifies the points based on the attribute data. Else it classifies the points
as boolean. TRUE if the points are in a polygon, FALSE otherwise. This function
allows for filtering lakes, for example.
}
\details{
The function recognizes only SpatialPolygonsDataFrame containing polygons.
Polygons can be simple, one-part shapes, multi-part polygons, or polygons with
holes.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
shapefile_dir <- system.file("extdata", package = "lidR")

lidar = readLAS(LASfile)
lakes = rgdal::readOGR(shapefile_dir, "lake_polygons_UTM17")

# The field "inlake" does not exist in the shapefile. Points are classified as TRUE if in a polygon
classify_from_shapefile(lidar, lakes, "inlakes")
forest = lasfilter(lidar, inlakes == FALSE)
plot(lidar)
plot(forest)

# The field "LAKENAME_1" exists in the shapefile.
# Points are classified with the value of the polygon
classify_from_shapefile(lidar, lakes, "LAKENAME_1")
}
}
\seealso{
\code{\link[rgdal:readOGR]{readOGR} }
\code{\link[sp:SpatialPolygonsDataFrame-class]{SpatialPolygonsDataFrame} }
}

