% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{clipCircle}
\alias{clipCircle}
\alias{clipCircle,Lidar-method}
\title{Clip LiDAR points within a disc}
\usage{
clipCircle(obj, xcenter, ycenter, radius, inside = TRUE)

\S4method{clipCircle}{Lidar}(obj, xcenter, ycenter, radius, inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{xcenter}{a scalar of x circle center}

\item{ycenter}{a scalar of y circle center}

\item{radius}{a a scalar of circle radius}

\item{inside}{logical. Keep data inside or outside the shape}
}
\value{
An object of class \code{Lidar}
}
\description{
Clip LiDAR points within a disc
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

subset = lidar \%>\% clipCircle(685000, 5018000, 25)

plot(subset)
}
\seealso{
\link[lidR:clipRectangle]{clipRectangle}
\link[lidR:clipCircle]{clipCircle}
\link[lidR:clipCircle]{clipPolygon}
}

