% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\docType{methods}
\name{clipPolygon}
\alias{clipPolygon}
\alias{clipPolygon,LAS-method}
\title{Clip LiDAR points within a polygon}
\usage{
clipPolygon(obj, x, y, inside = TRUE)

\S4method{clipPolygon}{LAS}(obj, x, y, inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{x}{numerical array of x-coordinates of polygon}

\item{y}{numerical array of y-coordinates of polygon}

\item{inside}{logical. Keep data inside or outside the shape}
}
\value{
An object of class \code{LAS}
}
\description{
Clip LiDAR points within a polygon
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = readLAS(LASfile)

subset = lidar \%>\% clipPolygon(x=c(685000, 685200, 685050),
                               y=c(5018000, 5018100, 5018200))


plot(subset)
}
\seealso{
\link[lidR:clipRectangle]{clipRectangle}
\link[lidR:clipCircle]{clipCircle}
\link[lidR:clipCircle]{clipPolygon}
}

