% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_individual_tree_segmentation.r
\name{dalponte2016}
\alias{dalponte2016}
\title{Algorithm for individual tree segmentation}
\usage{
dalponte2016(chm, treetops, th_tree = 2, th_seed = 0.45, th_cr = 0.55,
  max_cr = 10, extra = FALSE, ID = "treeID")
}
\arguments{
\item{chm}{RasterLayer. Image of the canopy. Can be computed with \link{grid_canopy} or read from
an external file.}

\item{treetops}{\code{SpatialPointsDataFrame}. Can be computed with
\link[lidR:tree_detection]{tree_detection} or read from an external shapefile.}

\item{th_tree}{numeric. Threshold below which a pixel cannot be a tree. Default 2.}

\item{th_seed}{numeric. Growing threshold 1. See reference in Dalponte et al. 2016. A pixel
is added to a region if its height is greater than the tree height multiplied by this value.
It should be between 0 and 1. Default 0.45.}

\item{th_cr}{numeric. Growing threshold 2. See reference in Dalponte et al. 2016. A pixel
is added to a region if its height is greater than the current mean height of the region
multiplied by this value. It should be between 0 and 1. Default 0.55.}

\item{max_cr}{numeric. Maximum value of the crown diameter of a detected tree (in pixels).
Default 10.}

\item{ID}{character. If the \code{SpatialPointsDataFrame} contains an attribute with the ID for
each tree, the name of this attribute. This way, original IDs will be preserved. If there is no
such data trees will be numbered sequentially.}
}
\description{
This functions is made to be used in \link{lastress}. Individual tree segmentation using Dalponte
et al. (2016) algorithm (see reference). This is a seeds + growing region algorithm. It is
based on the constraints proposed by Dalponte and Coomes (see references). This algorithm exists
in the package \code{itcSegment}. This version is identical to the original but with superfluous
code removed and rewritten efficiently. Consequently it is hundreds to millions times faster. Note
that this algorithm strictly performs a segmentation, while the original method as implemented
in \code{itcSegment} and described in the manuscript also performs a pre- and post-process when
these tasks are expected to be done by the user in separate functions.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
col = pastel.colors(200)

chm = grid_canopy(las, 0.5, p2r(0.3))
kernel = matrix(1,3,3)
chm = raster::focal(chm, w = kernel, fun = mean, na.rm = TRUE)

ttops = tree_detection(chm, lmf(4, 2))
lastrees(las, dalponte2016(chm, ttops))
plot(las, color = "treeID", colorPalette = col)

}
\references{
Dalponte, M. and Coomes, D. A. (2016), Tree-centric mapping of forest carbon density from
airborne laser scanning and hyperspectral data. Methods Ecol Evol, 7: 1236–1245. doi:10.1111/2041-210X.12575.
}
\seealso{
Other Algorithm: \code{\link{csf}},
  \code{\link{hamraz2016}}, \code{\link{highest}},
  \code{\link{homogenize}}, \code{\link{li2012}},
  \code{\link{lmf}}, \code{\link{manual}},
  \code{\link{multichm}}, \code{\link{pmf}},
  \code{\link{ptrees}}, \code{\link{random}},
  \code{\link{silva2016}}, \code{\link{tin}},
  \code{\link{watershed}}, \code{\link{wing2015}}

Other IndividualTreeSegmentation: \code{\link{hamraz2016}},
  \code{\link{lastrees}}, \code{\link{li2012}},
  \code{\link{silva2016}}, \code{\link{watershed}}

Other RasterBasedSegmentation: \code{\link{silva2016}},
  \code{\link{watershed}}
}
