% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{lasmetrics}
\title{Deprecated functions in lidR}
\usage{
lasmetrics(...)
}
\arguments{
\item{...}{parameters}
}
\description{
These functions are provided for compatibility with older versions of lidR only, and may be
defunct as soon as the next release.
}
