% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{lasmetrics}
\alias{grid_metrics3d}
\alias{grid_hexametrics}
\title{Deprecated functions in lidR}
\usage{
lasmetrics(las, func)

grid_metrics3d(...)

grid_hexametrics(...)
}
\arguments{
\item{las, func, ...}{parameters}
}
\description{
These functions are provided for compatibility with older versions of lidR only, and may be
defunct as soon as the next release.
}
