% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{readALSLAS}
\alias{readTLSLAS}
\alias{readUAVLAS}
\alias{readDAPLAS}
\alias{readALSLAScatalog}
\alias{readTLSLAScatalog}
\alias{readUAVLAScatalog}
\alias{readDAPLAScatalog}
\alias{filter_surfacepoints}
\alias{filter_surfacepoints.LAS}
\alias{filter_surfacepoints.LAScatalog}
\title{Deprecated functions in lidR}
\usage{
readALSLAS(files, select = "*", filter = "")

readTLSLAS(files, select = "*", filter = "", sort = TRUE)

readUAVLAS(files, select = "*", filter = "", sort = TRUE)

readDAPLAS(files, select = "*", filter = "", sort = TRUE)

readALSLAScatalog(folder, ...)

readTLSLAScatalog(folder, ...)

readUAVLAScatalog(folder, ...)

readDAPLAScatalog(folder, ...)

filter_surfacepoints(las, res)

\method{filter_surfacepoints}{LAS}(las, res)

\method{filter_surfacepoints}{LAScatalog}(las, res)
}
\arguments{
\item{files, select, filter, sort}{See the new functions that replace the old ones}

\item{folder, ...}{See the new functions that replace the old ones}

\item{las, res}{See the new functions that replace the old ones}
}
\description{
These functions are provided for compatibility with older versions of lidR but are deprecated.
}
