% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-dsm.R
\name{dsm_spikefree}
\alias{dsm_spikefree}
\alias{spikefree}
\title{Digital Surface Model Algorithm}
\usage{
spikefree(freeze_distance, height_buffer = 0.5)
}
\arguments{
\item{freeze_distance}{freeze distance (see references). Recommended value: 3 times the pulse spacing or
a little higher.}

\item{height_buffer}{buffer distance (see references). Recommended value: 0.5 do not change.}
}
\description{
This function is made to be used in \link{rasterize_canopy}. It implements the spike-free algorithm
developed by Khosravipour et al. (2016), which is based on the computation of an incremental
triangulation of all returns with triangle freezing criteria.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
poi = "-drop_z_below 0 -inside 481280 3812940 481330 3812990"
las <- readLAS(LASfile, filter = poi)
col <- height.colors(50)

# Khosravipour et al. spikefree algorithm
chm <- rasterize_canopy(las, res = 0.25, spikefree(1.5))
plot(chm, col = col)
}
\references{
Khosravipour, Anahita & Skidmore, Andrew & Isenburg, Martin. (2016). Generating spike-free
digital surface models using LiDAR raw point clouds: A new approach for forestry applications.
International Journal of Applied Earth Observation and Geoinformation. 52. 104-114. 10.1016/j.jag.2016.06.005.
}
\seealso{
Other digital surface model algorithms: 
\code{\link{dsm_pitfree}},
\code{\link{dsm_point2raster}},
\code{\link{dsm_tin}}
}
\concept{digital surface model algorithms}
