% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.r
\docType{methods}
\name{extent,LAS-method}
\alias{extent,LAS-method}
\alias{extent,LAScatalog-method}
\title{Extent}
\usage{
\S4method{extent}{LAS}(x, ...)

\S4method{extent}{LAScatalog}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{LAS} or \code{LAScatalog}}

\item{\dots}{Unused}
}
\value{
Extent object from \pkg{raster}
}
\description{
Returns an Extent object of a \code{LAS} or \code{LAScatalog} object.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)
extent(las)
}
\seealso{
\code{\link[raster:extent]{raster::extent} }
}
