% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASmethods-LAS.r, R/LASmethods-LAScatalog.r
\docType{methods}
\name{extent,LAS-method}
\alias{extent,LAS-method}
\alias{extent,LAScatalog-method}
\title{Extent}
\usage{
\S4method{extent}{LAS}(x, ...)

\S4method{extent}{LAScatalog}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{LAS} or \code{LAScatalog}}

\item{\dots}{Unused}
}
\value{
Extent object from \pkg{raster}
}
\description{
Returns an Extent object of a \code{LAS*}.
}
\seealso{
\code{\link[raster:extent]{raster::extent}}
}
