% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLiDARProject.r
\name{extractGroundInventory}
\alias{extractGroundInventory}
\title{Extract inventory from a set of tiles}
\usage{
extractGroundInventory(obj, plotnames, x, y, radius, buffer = 2)
}
\arguments{
\item{obj}{A LiDARProject object}

\item{plotnames}{vector. A set of plot names}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{radius}{numeric or vector. A radius or a set of radiuses of plots}

\item{buffer}{numeric. A buffer value to expend the extent of search}
}
\value{
A list of Lidar objects
}
\description{
Extract inventory from a set of tiles
}

