% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassCatalog.r
\docType{methods}
\name{extractGroundInventory}
\alias{extractGroundInventory}
\alias{extractGroundInventory,Catalog-method}
\title{Extract inventory from a set of tiles}
\usage{
extractGroundInventory(obj, plotnames, x, y, radius, buffer = 2)

\S4method{extractGroundInventory}{Catalog}(obj, plotnames, x, y, radius,
  buffer = 2)
}
\arguments{
\item{obj}{A Catalog object}

\item{plotnames}{vector. A set of plot names}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{radius}{numeric or vector. A radius or a set of radii of plots}

\item{buffer}{numeric. A buffer value to extend the search range}
}
\value{
A list of Lidar objects
}
\description{
When the user has a set of coordinates of ground inventory, they can extract the
lidar data associated with these plot automatically.
}
\details{
When the user has a set of coordinates of ground inventory, they can extract the
lidar data associated with these plot automatically. The algorithm automatically extracts
plots falling on the edges of tiles.
}

