% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapFractionProfile.r
\name{gapFractionProfile}
\alias{gapFractionProfile}
\title{Gap fraction profile}
\usage{
gapFractionProfile(z, dz = 1)
}
\arguments{
\item{z}{vector of positive z coordinates}

\item{dz}{numeric. The thickness of the layers used (height bin)}
}
\value{
A data.frame containing the bin elevations (z) and the gap fraction for each bin (gf)
}
\description{
Computes the gap fraction profile using the method of Bouvier et al. (see reference)
}
\details{
The function assessing the number of laser points that actually reached the layer
z+dz and those that passed through the layer [z, z+dz]. By definition the layer 0
will always return 0 because no returns reached the ground. Therefore, the layer 0 is removed
from the returned results.
}
\examples{
z = dbeta(seq(0, 0.8, length = 1000), 6, 6)*10

gapFraction = gapFractionProfile(z)

plot(gapFraction, type="l", xlab="Elevation", ylab="Gap fraction")
}
\references{
Bouvier, M., Durrieu, S., Fournier, R. a, & Renaud, J. (2015).  Generalizing predictive models of forest inventory attributes using an area-based approach with airborne LiDAR data. Remote Sensing of Environment, 156, 322-334. http://doi.org/10.1016/j.rse.2014.10.004
}
\seealso{
\link[lidR:LAD]{LAD}
}

