% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ground_elevation.r
\docType{methods}
\name{get_ground_elevation}
\alias{get_ground_elevation}
\alias{get_ground_elevation,LAS-method}
\title{Get the evelation of the ground for given coordinates}
\usage{
get_ground_elevation(las, coord, k = 3L)

\S4method{get_ground_elevation}{LAS}(las, coord, k = 3L)
}
\arguments{
\item{las}{a LAS objet}

\item{coord}{data.frame containing the at least the coordinates of the points of
interest in the columns named X and Y.}

\item{k}{numeric. The number of nearest neighbours}
}
\value{
Numeric. The predicted elevations.
}
\description{
The algorithm use the k-nearest ground points of the input coordinates to
interpol the elevations at the points of interest. The interpolation is done by kriging
using a invert distance kernel.
}
\seealso{
\link[kknn:kknn]{kknn}
\link[lidR:grid_terrain]{grid_terrain}
}

