% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Lidar-method}
\title{Get LiDAR data}
\usage{
getData(obj)

\S4method{getData}{Lidar}(obj)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}
}
\value{
It returns a \code{data.table} with containing the LiDAR data
}
\description{
Return the slot @data from an \code{Lidar} object
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = LoadLidar(LASfile)

getData(lidar)
}

