% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirst.r
\docType{methods}
\name{getFirst}
\alias{getFirst}
\alias{getFirst,LAS-method}
\title{Filter first returns}
\usage{
getFirst(obj)

\S4method{getFirst}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the first returns.
}
\note{
\code{getFirst(obj)} is an alias for \code{lasfilter(obj, ReturnNumber == 1)}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

firstReturns = lidar \%>\% getFirst
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:lasfilter]{lasfilter} }
}

