% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast.r
\docType{methods}
\name{getFirstLast}
\alias{getFirstLast}
\alias{getFirstLast,LAS-method}
\title{Filter first and last returns}
\usage{
getFirstLast(obj)

\S4method{getFirstLast}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the first and last returns.
}
\note{
\code{getFirstLast(obj)} is an alias for \code{extract(obj, ReturnNumber == NumberOfReturns | ReturnNumber == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = readLAS(LASfile)

firstLastReturns = lidar \%>\% getFirstLast
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:extract]{extract} }
}

