% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast.r
\docType{methods}
\name{getFirstLast}
\alias{getFirstLast}
\alias{getFirstLast,LAS-method}
\title{Filter first and last returns}
\usage{
getFirstLast(obj)

\S4method{getFirstLast}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the first and last returns.
}
\note{
\code{getFirstLast(obj)} is an alias for \code{lasfilter(obj, ReturnNumber == NumberOfReturns | ReturnNumber == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

firstLastReturns = lidar \%>\% getFirstLast
}
\seealso{
\link[lidR:lasfilter]{lasfilter}

Other getters: \code{\link{getFirstOfMany}},
  \code{\link{getFirst}}, \code{\link{getGround}},
  \code{\link{getLast}}, \code{\link{getNth}},
  \code{\link{getSingle}}
}

