% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{getFirstOfMany}
\alias{getFirstOfMany}
\alias{getFirstOfMany,Lidar-method}
\title{Filter first returns from pulses which returned multiple points}
\usage{
getFirstOfMany(obj)

\S4method{getFirstOfMany}{Lidar}(obj)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}
}
\value{
An object of class \code{Lidar}
}
\description{
Select only the first returns from pulses which returned multiple points form a \code{Lidar} object
}
\note{
\code{getFirstOfMany(obj)} is an alias for \code{filter(obj, NumberOfReturns > 1, ReturnNumber == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

firstOfManyReturns = lidar \%>\% getFirstOfMany
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:filter]{filter} }
}

