% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGround.r
\docType{methods}
\name{getGround}
\alias{getGround}
\alias{getGround,LAS-method}
\title{Filter returns classified as ground}
\usage{
getGround(obj)

\S4method{getGround}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the returns classified as ground according to LAS specification v1.3.
}
\note{
\code{getGround(obj)} is an alias for \code{lasfilter(obj, Classification == 2)}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

ground = lidar \%>\% getGround
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:lasfilter]{lasfilter} }
}

