% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLast.r
\docType{methods}
\name{getLast}
\alias{getLast}
\alias{getLast,LAS-method}
\title{Filter last returns}
\usage{
getLast(obj)

\S4method{getLast}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the last returns i.e. the last returns and the single returns
}
\note{
\code{getLast(obj)} is an alias for \code{lasfilter(obj, ReturnNumber == NumberOfReturns))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

lastReturns = lidar \%>\% getLast
}
\seealso{
\link[lidR:lasfilter]{lasfilter}

Other getters: \code{\link{getFirstLast}},
  \code{\link{getFirstOfMany}}, \code{\link{getFirst}},
  \code{\link{getGround}}, \code{\link{getNth}},
  \code{\link{getSingle}}
}

