% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNth.r
\docType{methods}
\name{getNth}
\alias{getNth}
\alias{getNth,LAS-method}
\title{Filter returns by their position in the return sequence.}
\usage{
getNth(obj, n)

\S4method{getNth}{LAS}(obj, n)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{n}{numeric. The position in the return sequence}
}
\value{
An object of class \code{LAS}
}
\description{
Select the returns from their position in the return sequence. Point density
pulse density and area are recomputed on the fly
}
\note{
\code{getNth(obj, n)} is an alias for \code{lasfilter(obj, ReturnNumber == n)}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

secondReturns = lidar \%>\% getNth(2)
}
\seealso{
\link[lidR:lasfilter]{lasfilter}

Other getters: \code{\link{getFirstLast}},
  \code{\link{getFirstOfMany}}, \code{\link{getFirst}},
  \code{\link{getGround}}, \code{\link{getLast}},
  \code{\link{getSingle}}
}

