% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingle.r
\docType{methods}
\name{getSingle}
\alias{getSingle}
\alias{getSingle,LAS-method}
\title{Filter single returns}
\usage{
getSingle(obj)

\S4method{getSingle}{LAS}(obj)
}
\arguments{
\item{obj}{An object of class \code{LAS}}
}
\value{
An object of class \code{LAS}
}
\description{
Select only the returns which return only one point.
}
\note{
\code{getSingle(obj)} is an alias for \code{extract(obj, NumberOfReturns == 1))}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = readLAS(LASfile)

singleReturns = lidar \%>\% getSingle
}
\seealso{
\code{\link[lidR:getFirst]{getFirst} }
\code{\link[lidR:getFirstLast]{getFirstLast} }
\code{\link[lidR:getFirstOfMany]{getFirstOfMany} }
\code{\link[lidR:getSingle]{getSingle} }
\code{\link[lidR:getLast]{getLast} }
\code{\link[lidR:getGround]{getGround} }
\code{\link[lidR:getNth]{getNth} }
\code{\link[lidR:extract]{extract} }
}

