% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_canopy.r
\name{grid_canopy}
\alias{grid_canopy}
\title{Canopy surface model}
\usage{
grid_canopy(.las, res = 2, subcircle = 0)
}
\arguments{
\item{.las}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is
2 meters i.e. 4 square meters.}

\item{subcircle}{numeric radius of the circles. To fill empty pixels the algorithm can
replaces each return by a circle composed by 8 points before to compute the maximum elevation
in each pixel.}
}
\value{
It returns a \code{data.table} with the class \code{grid_metrics} which enables easier plotting.
}
\description{
Creates a canopy surface model using a LiDAR cloud of points. For each pixel the
function returns highest point found. This basic method could be improved by replacing
each LiDAR return with a small disk
}
\details{
The algorithm used is the local maximum algorithm. It assigns the
elevation of the highest return within each grid cell to the grid cell center.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Local maximum algorithm with a resolution of 2 meters
lidar \%>\% grid_canopy(2) \%>\% plot
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}

Other grid_alias: \code{\link{grid_density}}
}

