% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_canopy.r
\docType{methods}
\name{grid_canopy}
\alias{grid_canopy}
\alias{grid_canopy,LAS-method}
\title{Canopy surface model}
\usage{
grid_canopy(obj, res = 2, start = c(0, 0))

\S4method{grid_canopy}{LAS}(obj, res = 2, start = c(0, 0))
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 2 units i.e. 4 square units cells.}

\item{start}{vector of x and y coordinates for the reference raster. Default is (0,0) see \link[lidR:grid_metrics]{grid_metrics}}
}
\value{
It returns a \code{data.table} with the class \code{grid_metrics} which enables easier plotting.
}
\description{
Creates a canopy surface model using a LiDAR cloud of points. This function is an
alias for \code{grid_metrics(obj, res, max(Z))}.
}
\details{
The algorithm used is the local maximum algorithm. It assigns the
elevation of the highest return within each grid cell to the grid cell center.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Local maximum algorithm with a resolution of 2 meters
lidar \%>\% grid_canopy(2) \%>\% plot
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}

Other grid_alias: \code{\link{grid_density}}
}

