% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_density.r
\name{grid_density}
\alias{grid_density}
\title{Pulse density surface model}
\usage{
grid_density(obj, res = 4)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 4 units i.e. 16 square units cells.}
}
\value{
It returns a \code{data.table} with the class \code{grid_metrics} which which enables easier plotting.
}
\description{
Creates a pulse density map using a LiDAR cloud of points. This function is an alias
for \code{grid_metrics(obj, res, length(unique(pulseID))/res^2)}.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

lidar \%>\% grid_density(5) \%>\% plot
lidar \%>\% grid_density(10) \%>\% plot
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}

Other grid_alias: \code{\link{grid_canopy}}
}

