% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_density.r
\docType{methods}
\name{grid_density}
\alias{grid_density}
\alias{grid_density,LAS-method}
\title{Pulse density surface model}
\usage{
grid_density(obj, res = 4)

\S4method{grid_density}{LAS}(obj, res = 4)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 4 units i.e. 16 square units cells.}
}
\value{
It returns a \code{data.table} with the class \code{grid_metrics} which which enables easier plotting.
}
\description{
Creates a pulse density map using a LiDAR cloud of points.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

lidar \%>\% grid_density(5) \%>\% plot
lidar \%>\% grid_density(10) \%>\% plot
}
\seealso{
\code{\link[lidR:grid_metrics]{grid_metrics}}
}

