% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_terrain.r
\name{grid_terrain}
\alias{grid_terrain}
\title{Digital Terrain Model}
\usage{
grid_terrain(obj, res = 1, k = 7L, kernel = "inv", ...)
}
\arguments{
\item{obj}{LAS objet}

\item{res}{numeric resolution.}

\item{k}{numeric. The number of ground points used to interpolate (see
\link[lidR:lasterrain]{lasterrain})}

\item{kernel}{character. Kernel to use. Default is "inv". See \link[kknn:kknn]{kknn}
for possible choices.}

\item{...}{extra parameter for \link[kknn:kknn]{kknn}}
}
\value{
A RasterLayer from package raster
}
\description{
Interpol ground points using a k neareast neighbourgh approach and create a rasterized
digital terrain model. The interpolation is done by default using an invert distance weighting
(IDW). The algorithm uses the points classified as "ground" and for each pixel the
algorithm uses the k nearest ground point to compute the interpolation.
}
\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

# plot(lidar)

dtm = grid_terrain(lidar)

# plot3d(dtm)
}
\seealso{
\link[lidR:normalize]{normalize}
}

