% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_terrain.r
\docType{methods}
\name{grid_terrain}
\alias{grid_terrain}
\alias{grid_terrain,LAS-method}
\title{Digital Terrain Model}
\usage{
grid_terrain(obj, res = 1, linear = F, ...)

\S4method{grid_terrain}{LAS}(obj, res = 1, linear = F, ...)
}
\arguments{
\item{obj}{LAS objet}

\item{res}{resolution}

\item{linear}{using linear or spline interpolation. Default is spline (slower
but more realistic).}

\item{...}{optionnal parameters for \link[akima:interp]{interp}.}
}
\value{
A \code{list} containing the elevation for each cell of the
output grid. The list has the class "DTM" enabling to easily deal with it in
the \code{lidR} package.
}
\description{
Interpol ground points using linear interpolation or spline interpolation and
create a digital terrain model (DTM). This function rely on the
\link[akima:interp]{interp} function from package \code{akima}.
}
\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

plot(lidar)

# Linear interpolation is fast, linear = FALSE for spline interpolation
dtm = grid_terrain(lidar, linear = TRUE)

plot(dtm)

\dontrun{
}
}
\seealso{
\link[akima:interp]{interp}
\link[lidR:normalize]{normalize}
}

