% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_terrain.r
\docType{methods}
\name{grid_terrain}
\alias{grid_terrain}
\alias{grid_terrain,LAS-method}
\title{Digital Terrain Model}
\usage{
grid_terrain(obj, res = 1, k = 3L)

\S4method{grid_terrain}{LAS}(obj, res = 1, k = 3L)
}
\arguments{
\item{obj}{LAS objet}

\item{res}{resolution}

\item{k}{numeric. The number of ground points used to interpolate (see
\link[lidR:get_ground_elevation]{get_ground_elevation})}
}
\value{
A RasterLayer from package raster
}
\description{
Interpol ground points using a kriging approach and create a rasterized digital terrain
model. The interpolation is done by kriging using a invert distance kernel. The algorithm
uses is the points classified as ground and for each pixel the algorithm use the k nearest
ground point to compute the interpolation.
}
\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

# plot(lidar)

# Linear interpolation is fast, linear = FALSE for spline interpolation
dtm = grid_terrain(lidar, linear = TRUE)

# plot3d(dtm)

\dontrun{
}
}
\seealso{
\link[lidR:normalize]{normalize}
}

