% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{gridMetrics}
\alias{gridMetrics}
\alias{gridMetrics,Lidar-method}
\title{Rasterize the space and compute metrics for each cells}
\usage{
gridMetrics(obj, res, func, start = c(0, 0), option = NULL)

\S4method{gridMetrics}{Lidar}(obj, res, func, start = c(0, 0),
  option = NULL)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{res}{numeric. The size of the cells}

\item{func}{the function to be apply to each cells}

\item{start}{vector x and y coordinates for the reference raster. Default is (0,0).}

\item{option}{character. Could be \code{"slip_flightline"}. In this case algorithm will compute the metrics for each flightline individually. It return several times the same cells in overlaps.}
}
\value{
It returns a \code{data.table} with containing the metrics for each cells. The table have the class "gridMetrics" enabling to easily plot it.
}
\description{
Computes a series of descriptive statistics for a LIDAR dataset for each cells
of a grid.
}
\details{
Computes a series of descriptive statistics for a LIDAR data set. Output is a
data.frame represented in database form with each record corresponding to a
single grid cell. GridMetrics is similar to CloudMetrics except it computes
metrics for all returns  within each cell in the output grid. Cloudmetrics
computes a single set of metrics for the entire data set. The grid cells
coordinates are pre-determinted for a given resolution. So algorithm will always
provides the same coordinates idependently of the datset. It allows to compare
the same cells on several data set. Particularly the option \code{slit_fightline}
makes the exact same grid for each flightlines.
When start = (0,0) and res = 20 m gridMetrics will produce the following raster centers (10,10), (10,30), (30,10) etc..
When start = (-10, -10) and res = 20 m gridMetrics will produce the following raster centers (0,0), (0,20), (20,0) etc..
In Quebec (Canada) reference is (-831600,  117980) in NAD83.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = LoadLidar(LASfile)

# Canopy surface model with 4 m^2 cells
gridMetrics(lidar, 2, max(Z)) \%>\% plot

# Mean height with 400 m^2 cells
gridMetrics(lidar, 20, mean(Z)) \%>\% plot

# Define your own metric function
myMetrics = function(z, i, angle, pulseID)
{
  ret = list(
        npulse  = length(unique(pulseID)),
        hmean   = mean(z),
        hmax    = max(z),
        imean   = mean(i),
        angle   = mean(abs(angle))
        )

   return(ret)
 }

metrics = gridMetrics(lidar, 20, myMetrics(Z, Intensity, ScanAngle, pulseID))
plot(metrics, "hmean")
plot(metrics, "hmax")
plot(metrics, "imean")
#etc...)
}

