% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{height.colors}
\alias{height.colors}
\title{height.colors}
\usage{
height.colors(n)
}
\arguments{
\item{n}{The number of colors (> 1) to be in the palette}
}
\description{
Create a vector of n contiguous colors of elevations.
}
\seealso{
\link[grDevices:colorRamp]{colorRampPalette}
}

