% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAD.r
\name{LAD}
\alias{LAD}
\title{Leaf area density}
\usage{
LAD(z, by = 1, threshold = 0)
}
\arguments{
\item{z}{vector of positive z coordinates}

\item{by}{numeric. The thickeness of the layers used (height bin)}

\item{threshold}{numeric. Remove points below the threshold}
}
\value{
numeric vector containing the leaf area density for each bin
}
\description{
Computes a leaf area density profile based on Bouvier et al. method (see reference)
}
\examples{
z = dbeta(seq(0, 0.8, length = 1000), 6, 6)*10

lad = LAD(z)

plot(1:length(lad)~lad, type="l", ylab="Elevation", xlab="Leaf area density")
}
\references{
Bouvier, M., Durrieu, S., Fournier, R. a, & Renaud, J. (2015).  Generalizing predictive models of forest inventory attributes using an area-based approach with airborne LiDAR data. Remote Sensing of Environment, 156, 322-334. http://doi.org/10.1016/j.rse.2014.10.004
}

