% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LAS.r
\docType{class}
\name{LAS-class}
\alias{LAS-class}
\title{An S4 class to represent the data read in a .las or .laz files}
\description{
Class is the representation of a las/laz file.
}
\details{
A \code{LAS} object inherit a \link[sp:Spatial-class]{Spatial} object from \code{sp}. Thus it is
a \code{Spatial} object plus a \code{data.table} with the data read from  a \code{las/laz} file and
a \link[lidR:LASheader-class]{LASheader} (see the ASPRS documentation for the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}
for more information). las files being normalised the table of attributes read from the las/laz file
is also normalized. Columns are named:
\itemize{
\item{\code{X} (numeric)}
\item{\code{Y} (numeric)}
\item{\code{Z} (numeric)}
\item{\code{Intensity} (integer)}
\item{\code{ReturnNumber} (integer)}
\item{\code{NumberOfReturns} (integer)}
\item{\code{ScanDirectionFlag} (integer)}
\item{\code{EdgeOfFlightline} (integer)}
\item{\code{Classification} (integer)}
\item{\code{Synthetic_flag} (logical)}
\item{\code{Keypoint_flag} (logical)}
\item{\code{Withheld_flag} (logical)}
\item{\code{ScanAngle} (integer)}
\item{\code{UserData} (integer)}
\item{\code{PointSourceID} (integer)}
}
}
\section{Slots}{

\describe{
\item{\code{bbox}}{Object of class "matrix", with bounding box}

\item{\code{proj4string}}{Object of class \link[sp:CRS-class]{CRS}, projection string}

\item{\code{data}}{Object of class \link[data.table:data.table]{data.table}. Point cloud data accoridng to the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}}

\item{\code{header}}{Object of class \link[lidR:LASheader-class]{LASheader}. las file header accoridng to the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}}
}}

\section{Extends}{

Class  \link[sp:Spatial-class]{Spatial}, directly.
}

\seealso{
\link[lidR:LAS]{LAS}
\link[lidR:readLAS]{readLAS}
}
