% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasnormalize.r
\docType{methods}
\name{-,LAS,RasterLayer-method}
\alias{-,LAS,RasterLayer-method}
\title{Convenient operator to lasnormalize}
\usage{
\S4method{-}{LAS,RasterLayer}(e1, e2)
}
\arguments{
\item{e1}{a LAS object}

\item{e2}{a RasterLayer}
}
\description{
Convenient operator to lasnormalize
}
