% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS-methods.r
\name{LAS}
\alias{LAS}
\title{Create object of class \link[lidR:LAS-class]{LAS}}
\usage{
LAS(data, header = list(), proj4string = sp::CRS(), check = TRUE)
}
\arguments{
\item{data}{a \link[data.table:data.table]{data.table} containing the data of a las or laz file.}

\item{header}{a \code{list} or a \link[lidR:LASheader-class]{LASheader} containing the header of
a las or laz file.}

\item{proj4string}{projection string of class \link[sp:CRS-class]{CRS-class}.}

\item{check}{logical. consistency tests while building the object.}
}
\value{
An object of class \code{LAS}
}
\description{
Create object of class \link[lidR:LAS-class]{LAS}
}
\seealso{
\link[lidR:LAS]{Class LAS}
}
