% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAS.r
\name{LAS}
\alias{LAS}
\title{Create a \code{LAS} object}
\usage{
LAS(data, header = list())
}
\arguments{
\item{data}{a data.table containing the LiDAR data.}

\item{header}{a list containing the data from the header of a las file.}
}
\value{
An object of class \code{LAS}
}
\description{
When a \code{LAS} object is created sevral things are computed on the fly in addition
to the data inputed in the constructor. See \link[lidR:LAS-class]{LAS-class}.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

summary(lidar)
}
\seealso{
\link[lidR:LAS]{Class LAS}
\link[lidR:summary]{summary}
}

