% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_tools.R
\name{las_utilities}
\alias{las_utilities}
\alias{las_rescale}
\alias{las_reoffset}
\alias{las_quantize}
\alias{las_update}
\alias{quantize}
\alias{is.quantized}
\alias{count_not_quantized}
\title{LAS tools}
\usage{
las_rescale(las, xscale, yscale, zscale)

las_reoffset(las, xoffset, yoffset, zoffset)

las_quantize(las, by_reference = TRUE)

las_update(las)

quantize(x, scale, offset, by_reference = TRUE, ...)

is.quantized(x, scale, offset, ...)

count_not_quantized(x, scale, offset)
}
\arguments{
\item{las}{An object of class LAS}

\item{xscale, yscale, zscale}{scalar. Can be missing if not relevant.}

\item{xoffset, yoffset, zoffset}{scalar. Can be missing if not relevant.}

\item{by_reference}{bool. Update the data in place without allocating new memory.}

\item{x}{numeric. Coordinates vector}

\item{scale, offset}{scalar. scale and offset}

\item{...}{Unused.}
}
\description{
Tools to manipulate LAS object maintaining compliance with
\href{http://www.asprs.org/wp-content/uploads/2019/07/LAS_1_4_r15.pdf}{ASPRS specification}
}
\details{
In the specification of the LAS format the coordinates are expected to be given
with a precision e.g. 0.01 for a millimeter precision (or millifeet) meaning
that a files records 123.46 not 123.45678. Also coordinates are stored as
integer. This is made possible with a scale and offset factor. For example
123.46 with an offset of 10000 and a scale factor of 0.01 is actually stored as
123.46/0.01 - 10000 = 2346. Storing 123.45678 with a scale of 0.01 and an offset
of 1000 is invalid because it does not converts to integer: 123.45678/0.01 - 10000
= 2345.678. Having an invalid LAS object may be critical in some application.
First when writing into a LAS file, user will loose the extra precision without
warning but some algorithm in lidR use the conversion to integer to make integer-based
compilation and speed-up some algorithm. Creation of an and invalid LAS object
may cause troubles.
}
\examples{
LASfile <- system.file("extdata", "example.laz", package="rlas")
las = readLAS(LASfile)

# Manual modification of the coordinates (rotation, alignment, ...)
las$X <- las$X + 2/3
las$Y <- las$Y - 5/3

# The point cloud is no longer valid
las_check(las)

# It is important to fix that
las_quantize(las)

# Now the file is almost valid
las_check(las)

# Update the object to set up-to-date header data
las <- las_update(las)
las_check(las)

# Recompute the coordinates with new scales and offsets
las <- las_rescale(las, xscale = 0.01, yscale = 0.01)
las <- las_reoffset(las, xoffset = 300000, yoffset = 5248000)
}
\seealso{
Other las utilities: 
\code{\link{las_check}()}
}
\concept{las utilities}
