% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasarea.r
\name{lasarea}
\alias{lasarea}
\title{Compute the area covered by of a set a points.}
\usage{
lasarea(.las)
}
\arguments{
\item{.las}{An object of the class \code{LAS} or a numeric array of x coordinates}
}
\value{
numeric. The area of the object computed in the same units as the coordinate reference system
}
\description{
The area is computed by appling a convex hull. If the data are not convex, the resulting area is only an approximation.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

lasarea(lidar)
}
