% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lascatalog.r
\docType{class}
\name{LAScatalog-class}
\alias{LAScatalog-class}
\title{An S4 class to represent a set of a .las or .laz files}
\description{
A LAScatalog object is a representation of a set of las/laz files
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table. A table representing the header of each file.}

\item{\code{crs}}{A \link[sp:CRS]{CRS} object.}
}}

\seealso{
\link[lidR:LAS-class]{LAS}
\link[lidR:LAS]{LAS}
\link[lidR:readLAS]{readLAS}
}
