% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasclip.r
\name{lasclip}
\alias{lasclip}
\alias{lasclipRectangle}
\alias{lasclipPolygon}
\alias{lasclipCircle}
\title{Clip LiDAR points}
\usage{
lasclip(x, geometry, ofile = "")

lasclipRectangle(x, xleft, ybottom, xright, ytop, ofile = "")

lasclipPolygon(x, xpoly, ypoly, ofile = "")

lasclipCircle(x, xcenter, ycenter, radius, ofile = "")
}
\arguments{
\item{x}{An object of class \code{LAS} or \code{LAScatalog}}

\item{geometry}{a geomtric object. Currently \code{Polygon} from \code{sp} is supported.}

\item{ofile}{character write the result in a file or load it into R. If \code{ofile = ""}
the result is loaded in R.}

\item{xleft}{scalar of left x position of rectangle}

\item{ybottom}{scalar of bottom y position of rectangle.}

\item{xright}{scalar of right x position of rectangle}

\item{ytop}{scalar of top y position of rectangle}

\item{xpoly}{numerical array. x-coordinates of polygon}

\item{ypoly}{numerical array. y-coordinates of polygon}

\item{xcenter}{scalar x disc center}

\item{ycenter}{scalar y disc center}

\item{radius}{scalar disc radius}
}
\value{
An object of class \code{LAS} or NULL if the results is immediatly written in a file.
}
\description{
Clip LiDAR points within a given geometry and convenient wrappers most common geometries
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)
subset = lasclipRectangle(las, 684850, 5017850, 684900, 5017900)
plot(subset)
}
