% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\name{lasclip}
\alias{lasclip}
\alias{lasclipCircle}
\alias{lasclipPolygon}
\alias{lasclipRectangle}
\title{Clip LiDAR points}
\usage{
lasclipRectangle(obj, xleft, ybottom, xright, ytop, inside = TRUE)

lasclipPolygon(obj, x, y, inside = TRUE)

lasclipCircle(obj, xcenter, ycenter, radius, inside = TRUE)
}
\arguments{
\item{obj}{An object of class \code{LAS}}

\item{xleft}{scalar. of left x position.}

\item{ybottom}{scalar. of bottom y position.}

\item{xright}{scalar. of right x position.}

\item{ytop}{scalar. of top y position.}

\item{inside}{logical. Keep data inside or outside the shape}

\item{x}{numerical array. x-coordinates of polygon}

\item{y}{numerical array. y-coordinates of polygon}

\item{xcenter}{scalar. x disc center}

\item{ycenter}{calar. y disc center}

\item{radius}{a scalar. Disc radius}
}
\value{
An object of class \code{LAS}
}
\description{
Clip LiDAR points within a given geometry
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

subset = lidar \%>\% lasclipRectangle(xleft=684850, ybottom=5017850,
                                 xright=684900, ytop =5017900)

plot(subset)
}

