% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lascolor.r
\name{lascolor}
\alias{lascolor}
\title{Compute the color from RGB fields}
\usage{
lascolor(.las, nbits = 16)
}
\arguments{
\item{.las}{A LAS object}

\item{nbits}{the number of bits used to store the R G and B field. Default is 16 bits i.e.
each channel ranged between 0 and 65535 as defined in LAS specifications. But sometime this rule
is not respected. You can force the default behavior. A 8 bit color ranged between 0 and 255.}
}
\value{
Return nothing. The original object is modified in place by reference.
}
\description{
Compute the hexadecimal name of a color from RGB fields. A column 'color' is added in the
slot @data
}
