% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasfilter.r
\docType{methods}
\name{lasfilter}
\alias{lasfilter}
\alias{lasfilter,LAS-method}
\title{Return points with matching conditions}
\usage{
lasfilter(.las, ...)

\S4method{lasfilter}{LAS}(.las, ...)
}
\arguments{
\item{.las}{An object of class \code{LAS}}

\item{\dots}{Logical predicates. Multiple conditions are combined with & or ,}
}
\value{
An object of class \code{LAS}
}
\description{
Return points with matching conditions.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

# Select the first returns classified as ground
firstground = lidar \%>\% lasfilter(Classification == 1 & ReturnNumber == 1)

# Multiple arguments are equivalent to &
firstground = lidar \%>\% lasfilter(Classification == 1, ReturnNumber == 1)

# Multiple criteria
first_or_ground = lidar \%>\% lasfilter(Classification == 1 | ReturnNumber == 1)
}
\seealso{
\link[dplyr:filter]{filter}
\link[lidR:LAS]{Class LAS}
\link[lidR:getFirst]{getFirst}
\link[lidR:getFirstLast]{getFirstLast}
\link[lidR:getFirstOfMany]{getFirstOfMany}
\link[lidR:getSingle]{getSingle}
\link[lidR:getLast]{getLast}
\link[lidR:getGround]{getGround}
\link[lidR:getNth]{getNth}
}

