% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasfiltersurfacepoints.r
\name{lasfiltersurfacepoints}
\alias{lasfiltersurfacepoints}
\title{Filter the surface points}
\usage{
lasfiltersurfacepoints(las, res)
}
\arguments{
\item{las}{An object of class \link[lidR:LAS-class]{LAS} or \link[lidR:LAScatalog-class]{LAScatalog}.}

\item{res}{numeric. The resolution of the grid used to filter the point cloud}
}
\value{
If the input is a \code{LAS} object, returns a \code{LAS} object. If the input is a
\code{LAScatalog} returns a \code{LAScatalog}.
}
\description{
This function is superseded by the algorithm \link{highest} usable in \code{lasfilterdecimate}
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)
subset = lasfiltersurfacepoints(las, 2)
plot(subset)

}
\seealso{
Other lasfilters: \code{\link{lasfilterduplicates}},
  \code{\link{lasfilters}}, \code{\link{lasfilter}}
}
\concept{lasfilters}
