% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasindentify.r
\name{lasflightline}
\alias{lasflightline}
\title{Retrieve individual flightlines}
\usage{
lasflightline(.las, dt = 30)
}
\arguments{
\item{.las}{A LAS object}

\item{dt}{numeric. The threshold time lag used to retrieve flightlines}
}
\value{
Return nothing. The original object is modified in place by reference.
}
\description{
Retrieve each individual flightline by attributing a number to each point. The
function depends on the GPS time to retrieve each indivudual flightline. In a
continuous dataset, once points are ordered by GPS time, the time between two
consecutive points does not exceed a few milliseconds. If the time between two consecutive
points is too long it means that the second point is from a different flightline. The default
threshold is 30 seconds. A column  '\code{flightlineID}' is added in the slot  \code{@data}
}
