% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasground.r
\name{lasground}
\alias{lasground}
\title{Classify points as ground or not ground}
\usage{
lasground(las, algorithm, ...)
}
\arguments{
\item{las}{An object of class \code{LAS} or \code{LAScatalog}.}

\item{algorithm}{character. The name of an algorithm. Currently \code{"pmf"} and \code{csf} are
supported (see related sections).}

\item{...}{parameters for the algorithms. These depend on the algorithm used (see details
of the algorithms).}
}
\value{
Nothing, this function is used for its side effect. The original LAS object is updated by
reference. In the 'Classification' column a value of 2 denotes 'ground' according to LAS specifications.
}
\description{
Classify points as ground or not ground with several possible algorithms. The function is a wrapper
around all the existing methods. The functions update the field \code{Classification} of the LAS
input object. The points classified as 'ground' are assigned a value of 2 according to las
specifications (See the ASPRS documentation for the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}).
}
\section{Working with a \code{LAScatalog}}{

In \code{lidR} when the input is a \link[lidR:LAScatalog-class]{LAScatalog} the functions run following
the catalog processing options. The user can modify the processing options using the
\link[lidR:catalog]{available options}.\cr\cr
\code{lidR} supports .lax files that \strong{significantly} improve the computation speed of
spatial queries using a spatial index. Users should really take advange a .lax files but this is
not mandatory.
}

\section{Supported processing options for a LAScatalog}{

\itemize{
\item \strong{cores}: See \link{LAScatalog-class}.
\item \strong{progress}: See \link{LAScatalog-class}.
\item \strong{stop_early}: See \link{LAScatalog-class}.
\item \strong{by_file}: See \link{LAScatalog-class}.
\item \strong{tiling_size}: See \link{LAScatalog-class}.
\item \strong{alignment}: See \link{LAScatalog-class}.
\item \strong{output_files*}: Requiered because the output is likely to be too big to be returned
 in R and needs to be written in las/laz files.
}
}

\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")
las = readLAS(LASfile, select = "xyzrn")

ws = seq(3,12, 3)
th = seq(0.1, 1.5, length.out = length(ws))

lasground(las, "pmf", ws, th)

plot(las, color = "Classification")
}
\seealso{
Other lasground: \code{\link{lasground_csf}},
  \code{\link{lasground_pmf}}
}
