% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasground_csf.r
\name{lasground_csf}
\alias{lasground_csf}
\title{Classify points as ground based on Cloth Simulation}
\usage{
lasground_csf(las, sloop_smooth = FALSE, class_threshold = 0.5,
  cloth_resolution = 0.5, rigidness = 1L, iterations = 500L,
  time_step = 0.65, last_returns = TRUE)
}
\arguments{
\item{las}{An object of class \code{LAS} or \code{LAScatalog}.}

\item{sloop_smooth}{logical. When sharp slopes exist, set this parameter to TRUE to perform a
post-processing which will reduced errors.}

\item{class_threshold}{scalar. The distance to the simulated cloth to classify point cloud into ground
and non-ground. The default is 0.5.}

\item{cloth_resolution}{scalar. The distance between paticles in cloth. This is usually set to the
average distance of the points in the point cloud. The default value is 0.5.}

\item{rigidness}{integer. The rididness of the cloth. 1 stands for very soft cloth (to fit rugged
terrain), 2 stands for medium cloth and 3 stands for hard cloth (for flat terrain). The default is 1.}

\item{iterations}{integer. Maximum iteration for simulating cloth. The default value is 500. Usually,
users do not need to change this.}

\item{time_step}{scalar. Time step when simulating the cloth under the gravity. The default value
is 0.65. Usually, do not change this value. It is suitable for most cases.}

\item{last_returns}{logical. The algorithm will use only the last returns (including the first returns
in the cases of single return) to run the algorithm. If FALSE all the returns are used. If the fields
\code{'ReturnNumber'} or \code{'NumberOfReturns'} are not specified \code{'last_returns'} is turned
to \code{FALSE} automatically.}
}
\value{
Nothing, this function is used for its side effect. The original LAS object is updated by
reference. In the 'Classification' column a value of 2 denotes 'ground' according to LAS specifications.
}
\description{
Implements algorithms for segmentation of ground points. The function updates the field
\code{Classification} of the LAS input object. The points classified as 'ground' are
assigned a value of 2 according to las specifications (See the ASPRS documentation for the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}).
This method is the a strict implementation of the CSF algorithm made by Zhang et al. (2016) (see
references) that relies on the orginal source code written by the original author and exposed to
R via the the \code{RCSF} package.
}
\section{Working with a \code{LAScatalog}}{

In \code{lidR} when the input is a \link[lidR:LAScatalog-class]{LAScatalog} the functions run following
the catalog processing options. The user can modify the processing options using the
\link[lidR:catalog]{available options}.\cr\cr
\code{lidR} supports .lax files that \strong{significantly} improve the computation speed of
spatial queries using a spatial index. Users should really take advange a .lax files but this is
not mandatory.
}

\section{Supported processing options for a LAScatalog}{

\itemize{
\item \strong{cores}: See \link{LAScatalog-class}.
\item \strong{progress}: See \link{LAScatalog-class}.
\item \strong{stop_early}: See \link{LAScatalog-class}.
\item \strong{by_file}: See \link{LAScatalog-class}.
\item \strong{tiling_size}: See \link{LAScatalog-class}.
\item \strong{alignment}: See \link{LAScatalog-class}.
\item \strong{output_files*}: Requiered because the output is likely to be too big to be returned
 in R and needs to be written in las/laz files.
}
}

\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")
las = readLAS(LASfile, select = "xyzrn")

lasground(las, "csf")

plot(las, color = "Classification")
}
\references{
W. Zhang, J. Qi*, P. Wan, H. Wang, D. Xie, X. Wang, and G. Yan, “An Easy-to-Use Airborne LiDAR Data
Filtering Method Based on Cloth Simulation,” Remote Sens., vol. 8, no. 6, p. 501, 2016.
(http://www.mdpi.com/2072-4292/8/6/501/htm)
}
\seealso{
Other lasground: \code{\link{lasground_pmf}},
  \code{\link{lasground}}
}
