% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasheader-class.r
\docType{class}
\name{LASheader-class}
\alias{LASheader-class}
\title{An S4 class to represent the header read in a .las or .laz file}
\description{
A \code{LAS} object contains a \code{list} in the slot \code{@data} with
the data read from a \code{.las} file.
}
\section{Slots}{

\describe{
\item{\code{data}}{list. A list of information contained in the las file header.}
}}

\seealso{
\link[lidR:LAS]{LASheader}
\link[lidR:LAS]{LAS}
\link[lidR:readLAS]{readLAS}
}
