% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lasheader.r
\docType{class}
\name{LASheader-class}
\alias{LASheader-class}
\title{An S4 class to represent the header read in a .las or .laz file}
\description{
A \code{LASheader} object contains a \code{list} in the slot \code{@PHB} with
the data read from the Public Header Block and \code{list} in the slot \code{@VLR} with
the data read from the Variable Lenght Records
}
\section{Slots}{

\describe{
\item{\code{PHB}}{list. Represents the Public Header Block}

\item{\code{VLR}}{list. Represents the Variable Length Records}
}}

\seealso{
\link[lidR:LAS]{LAS}
\link[lidR:readLAS]{readLAS}
}
