% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasheader-class.r
\name{LASheader}
\alias{LASheader}
\title{Create a \code{LASheader} object}
\usage{
LASheader(data = list())
}
\arguments{
\item{data}{a list containing the data from the header of a las file.}
}
\value{
An object of class \code{LASheader}
}
\description{
Create a \code{LASheader} object
}
\seealso{
\link[lidR:LASheader]{Class LASheader}
\link[lidR:LAS]{Class LAS}
}
