% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasroi.r
\name{lasroi}
\alias{lasroi}
\title{Select a region of interest interactively}
\usage{
lasroi(.las, ...)
}
\arguments{
\item{.las}{An object of class \code{LAS}}

\item{\dots}{Optional parameters for the plot function}
}
\value{
An object of class \code{LAS}
}
\description{
Select a rectangular region of interest interactively using the mouse
}
\details{
\code{lasroi} enables the user to select a region of interest (ROI) by drawing
a rectangle with the mouse
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)

subset = lasroi(lidar)
}
}

