% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasterrain.r
\name{lasterrain}
\alias{lasterrain}
\title{Get the elevation of the ground for given coordinates}
\usage{
lasterrain(.las, coord, method = "knn_idw", k = 3L, linear = F)
}
\arguments{
\item{.las}{LAS objet}

\item{coord}{data.frame containing  the coordinates of interest in columns X and Y}

\item{method}{character can be \code{"knn_idw"} or \code{"akima"}}

\item{k}{numeric. number of k nearest neibourgh when selected method is \code{"knn_idw"}}

\item{linear}{logical indicating wether linear or spline interpolation should be used
when selected method is \code{"akima"}}
}
\value{
Numeric. The predicted elevations.
}
\description{
Interpolate ground points and return the elevation at the position of interest given by the
parameter. The interpolation can be done using two methods: \code{"knn_idw"} or
\code{"akima"} (see details). The algorithm uses the points classified as "ground" to
compute the interpolation.
}
\details{
Methods:
\itemize{
\item{\code{knn_idw}: interpolation is done using a k-nearest neighbour approach with
an inverse distance weighting (IDW).}
\item{\code{akima}: interpolation depends on the \link[akima:interp]{interp} function from
package \code{akima}. With this method no extrapolation is done outside of the convex hull
determined by the ground points.}
}
}
\seealso{
\link[kknn:kknn]{kknn}
\link[lidR:grid_terrain]{grid_terrain}
}

