% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastrees.r
\name{lastrees}
\alias{lastrees}
\title{Individual tree segmentation}
\usage{
lastrees(las, algorithm, attribute = "treeID")
}
\arguments{
\item{las}{An object of the class \link[lidR:LAS-class]{LAS}.}

\item{algorithm}{function. An algorithm of individual tree segmentation. \code{lidR} have several
possible algorithms:  \link{dalponte2016}, \link{watershed}, \link{mcwatershed}, \link{li2012},
\link{hamraz2016}, \link{silva2016} or \link{ptrees}.}

\item{attribute}{character. The original LAS object is automatically updated by the function. A new
column is added. This parameter is the name of this new column.}
}
\value{
Nothing. The point cloud is updated by reference (in place without copy). It has a new column
names after the parameter \code{attribute}.
}
\description{
Individual tree segmentation with several possible algorithms.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
col = pastel.colors(200)

# Using Li et al. (2012)
lastrees(las, li2012(R = 3, speed_up = 5))
plot(las, color = "treeID", colorPalette = col)
}
\concept{IndividualTreeSegmentation}
