% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastree_hamraz.r
\name{lastrees_hamraz}
\alias{lastrees_hamraz}
\title{Individual tree segmentation}
\usage{
lastrees_hamraz(las, nps = 0.25, th = 5, R = 15.24, SENSITIVITY = 6,
  MDCW = 1.5, epsilon = 5, CLc = 0.8, Oc = 2/3, CLs = 0.7, Os = 1/3,
  angleRefCone = 90, angleRefSphere = 32.7)
}
\arguments{
\item{las}{A LAS object}

\item{nps}{nominal point spacing (page 533 in "2.Tree segmentation approach")}

\item{th}{minimal tree height to take into account (page 534)}

\item{R}{maximum horizontal distance of vertical profiles (page 535)}

\item{SENSITIVITY}{for inter tree gap identification ( multiplied by interquartile range from the third quartile) (p535)}

\item{MDCW}{minimum detectable crown width (page 534)}

\item{epsilon}{small deviation from vertical (page 535)}

\item{CLc}{crown ratio of a narrow cone-shaped crown (page 535)}

\item{Oc}{crown radius reduction due to the overlap assuming the narrow cone-shaped tree is situated in a dense stand(page 535)}

\item{CLs}{crown ratio of a sphere-shaped crown (page 535)}

\item{Os}{crown radius reduction due to the overlap within a dense stand for the sphere- shaped tree (page 535)}

\item{angleRefCone}{(page 536 - first sentence)}

\item{angleRefSphere}{(page 536 - first sentence)}
}
\value{
A LAS object
}
\description{
Individual tree segmentation
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
col = pastel.colors(200)

tree = lastrees_hamraz(las)

plot(tree, color = "treeID", colorPalette = pastel.colors(200))
}
