% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastree_hamraz.r
\name{lastrees_hamraz}
\alias{lastrees_hamraz}
\title{Individual tree segmentation}
\usage{
lastrees_hamraz(las, nps = 0.25, th = 5, MDCW = 1.5, epsilon = 5,
  CLc = 0.8, Oc = 2/3, CLs = 0.7, Os = 1/3, gap_sensitivity = 6L,
  R = 15.24, ...)
}
\arguments{
\item{las}{A LAS object}

\item{nps}{numeric. Nominal point spacing (see reference page 533  section 2)}

\item{th}{numeric. Minimal height. Point below this threshold are not condisered for the segmentation (see reference page 534 section 2)}

\item{MDCW}{numeric. Minimum detectable crown width (page 534 section 2)}

\item{epsilon}{numeric. Small deviation from vertical (page 535 section 2.2.2)}

\item{CLc}{numeric. Crown ratio of a narrow cone-shaped crown (page 535 equation 3)}

\item{Oc}{numeric. Crown radius reduction due to the overlap assuming the narrow cone-shaped tree is situated in a dense stand (page 535 equation 3)}

\item{CLs}{numeric. Crown ratio of a sphere-shaped crown (page 535 equation 4)}

\item{Os}{numeric. Crown radius reduction due to the overlap within a dense stand for the sphere-shaped tree (page 535 equation 4)}

\item{gap_sensitivity}{integer. In the original article, page 535 section 2.2.1, gaps are detected
using six times the interquartile range of square root distance between consecutive points. This
paramter control this value. Default is 6.}

\item{R}{numeric Maximum horizontal distance of vertical profiles (page 535 sectioh 2.1). Any value greater
than a crown is good because this parameter does not affect the result. However, it greatly affects the
computation speed. The lower the value, the faster the method.}
}
\value{
Nothing (NULL). The original point cloud is updated by reference.
}
\description{
It is an strict implementation of the Hamraz et al. 2012 (see references) algorithm made by the \code{lidR}
author from the original paper. The classification is done at the point cloud level and the function
returns nothing (NULL). The original point cloud is updated in place with an ID for each point in
a new column \code{treeID}. The user is free to post-process this output the way he want.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")
col = pastel.colors(200)

lastrees_hamraz(las, 0.25)

plot(las, color = "treeID", colorPalette = pastel.colors(200))
}
\references{
Hamraz, H., Contreras, M. A., & Zhang, J. (2016). A robust approach for tree segmentation in deciduous
forests using small-footprint airborne LiDAR data. International Journal of Applied Earth Observation
and Geoinformation, 52, 532–541. https://doi.org/10.1016/j.cageo.2017.02.017
}
