% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\docType{package}
\name{lidR-deprecated}
\alias{lidR-deprecated}
\alias{lasdecimate}
\alias{lidR-deprecated-package}
\title{Deprecated function(s) in the lidR package}
\usage{
lasdecimate(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the lidR package.  They may eventually be completely
removed.
}
