% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.r
\name{lidr_options}
\alias{lidr_options}
\title{Set or get global options for lidR package}
\usage{
lidr_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get global options for lidR package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{verbose}}{(\code{logical}) Make the package talkative. }
 \item{\code{debug}}{(\code{logical}) If you encounter a bug maybe that option would enable to find the problem more easily. }
}
}

