% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\name{lidr_options}
\alias{lidr_options}
\alias{lidr_reset}
\title{Options Settings for the lidR package}
\usage{
lidr_options(...)

lidr_reset()
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Allows the user to set and examine a variety of global options that affect the way in which
lidR computes and displays its results.
}
\section{Supported options}{

The following options are supported:
\itemize{
 \item{\code{verbose} (\code{logical}) Make the package "talkative". }
 \item{\code{progress} (\code{logical}) Display progress bar when available. }
 \item{\code{debug} (\code{logical}) Switch the package to debug mode when available.}
}
}

\examples{
lidr_options(verbose = TRUE)
lidr_options(progress = TRUE)
lidr_options()

# Reset default options
lidr_reset()
}
\seealso{
\link{catalog_options}
}
