% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadLidar.r
\name{LoadLidar}
\alias{LoadLidar}
\title{Load a las file and create a 'Lidar' object}
\usage{
LoadLidar(input, fields = "standard", ...)
}
\arguments{
\item{input}{character. Filename of .las file. Use \link[base:c]{c()} to concatenate several files.}

\item{fields}{character. Can be \code{"minimal"}, \code{"standard"}, \code{"all"}. Default is standard. See details.}

\item{\dots}{Unused}
}
\value{
An object of class \code{Lidar}
}
\description{
Methods to read and create a \code{Lidar} object from a vector of .las filename(s)
}
\details{
Methods to read and create a \code{Lidar} object from a vector of .las filename(s).
The option fields allows selection of fields to be loaded. Removing redundant fields
saves memory. The option '\code{minimal}' loads only X,Y,Z and gpstime allowing
pulseID and flightlineID to be computed. The option '\code{standard}' loads all fields
apart from UserDate, EdgeofFlighline and PointSourceID. The option '\code{all}' loads everything.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

getData(lidar)
summary(lidar)
}
\seealso{
\link[lidR:Lidar]{Class Lidar}
\link[lidR:getData]{getData}
\link[lidR:summary]{summary}
}

