% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_tree_detection.R
\name{manual}
\alias{manual}
\title{Tree top detection based on manual selection}
\usage{
manual(detected = NULL, ...)
}
\arguments{
\item{detected}{\code{SpatialPointsDataFrame} or \code{data.table} or \code{data.frame} or \code{matrix}
containing X,Y,Z coordinates of already found tree tops that need manual corrections.}

\item{...}{supplementary parameters to be pass to \link{plot}.}
}
\description{
This function is made to be used in \link{tree_detection} not alone. It implements an algorithm
for tree detection. Find the tree top positions manually and interactively using the mouse. This
is only suitable for small to medium size plots. First the point cloud is displayed, then the user
is invited to select a rectangular region of interest in the scene using the right button of the mouse.
Within the selected points the highest will be flaged as 'tree top' in the scene. Once all the tree
are labelled the user can exit the tools by selecting an empty region. Points can also be unflagged.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile)

# Full manual tree finding
ttops = tree_detection(las, manual())

# Automatic finding with manual correction
ttops = tree_detection(las, lmf(5))
ttops = tree_detection(las, manual(ttops))
}
}
\seealso{
Other Algorithm: \code{\link{csf}},
  \code{\link{dalponte2016}}, \code{\link{hamraz2016}},
  \code{\link{highest}}, \code{\link{homogenize}},
  \code{\link{li2012}}, \code{\link{lmf}},
  \code{\link{multichm}}, \code{\link{pmf}},
  \code{\link{ptrees}}, \code{\link{random}},
  \code{\link{silva2016}}, \code{\link{tin}},
  \code{\link{watershed}}, \code{\link{wing2015}}

Other Individual Tree Detection: \code{\link{lmf}},
  \code{\link{multichm}}, \code{\link{ptrees}}
}
