% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_tree_detection.R
\name{manual}
\alias{manual}
\title{Individual Tree Detection Algorithm}
\usage{
manual(detected = NULL, radius = 0.5, color = "red", ...)
}
\arguments{
\item{detected}{\code{SpatialPointsDataFrame} of already found tree tops that need manual corrections.}

\item{radius}{numeric. Radius of the spheres displayed on the point cloud (aesthetic purpose only).}

\item{color}{character. Color of the spheres displayed on the point cloud (aesthetic purpose only).}

\item{...}{supplementary parameters to be passed to \link{plot}.}
}
\description{
This function is made to be used in \link{tree_detection}. It implements an algorithm for manual
tree detection. Users can pinpoint the tree top positions manually and interactively using the mouse.
This is only suitable for small-sized plots. First the point cloud is displayed, then the user is
invited to select a rectangular region of interest in the scene using the right mouse button.
Within the selected points the highest one will be flagged as 'tree top' in the scene. Once all the trees
are labeled the user can exit the tool by selecting an empty region. Points can also be unflagged.
The goal of this tool is mainly for minor correction of automatically-detected tree outputs.
}
\examples{
\dontrun{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile)

# Full manual tree finding
ttops = tree_detection(las, manual())

# Automatic finding with manual correction
ttops = tree_detection(las, lmf(5))
ttops = tree_detection(las, manual(ttops))
}
}
\seealso{
Other individual tree detection algorithms: \code{\link{lmf}}
}
\concept{individual tree detection algorithms}
