% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{Substract digital terrain model}
\usage{
normalize(las, dtm = NULL, ...)
}
\arguments{
\item{las}{a LAS objet}

\item{dtm}{a digital terrain model. It can be a RasterLayer from package
\link[raster:raster]{raster} or a DTM from computed with
\link[lidR:grid_terrain]{grid_terrain}. If NULL the function will automatocally
compute it with \link[lidR:grid_terrain]{grid_terrain}.}

\item{...}{optionnal parameters for \link[lidR:grid_terrain]{grid_terrain} if
\code{dtm} parameter is NULL.}
}
\value{
A LAS object.
}
\description{
Substract digital terrain model (DTM) to the LiDAR data to create a dataset
normalized with the ground at 0. The digital terrain model can comes from
several sources such as external file or own computaion.
}
\examples{
LASfile <- system.file("extdata", "Topography.laz", package="lidR")

lidar = readLAS(LASfile)

plot(lidar)

# --- First possibility: read the DTM from a file -----
DTMfile <- system.file("extdata", "Topography.tif", package="lidR")
dtm = raster::raster(DTMfile)

lidar_norm = lidar - dtm # is synonyme with normalize(lidar, dtm)

plot(lidar_norm)

# --- Second possibility: compute the DTM with grid_terrain -----

# Linear interpolation is fast, linear = FALSE for spline interpolation
dtm = grid_terrain(lidar, linear = TRUE)

lidar_norm = lidar - dtm

plot(lidar_norm)
}
\seealso{
\link[raster:raster]{raster}
\link[lidR:grid_terrain]{grid_terrain}
}

