% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassCatalog.r
\name{plot.Catalog}
\alias{plot.Catalog}
\title{Plot a Catalog object}
\usage{
plot.Catalog(x, y, ...)
}
\arguments{
\item{x}{A Catalog object}

\item{y}{Unused (inherited from base plot)}

\item{\dots}{Unused (inherited from base plot)}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for a Catalog objects
}

