% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.catalog.r
\name{plot.Catalog}
\alias{plot.Catalog}
\title{Plot a Catalog object}
\usage{
\method{plot}{Catalog}(x, y, ...)
}
\arguments{
\item{x}{A Catalog object}

\item{y}{Unused (inherited from base plot)}

\item{\dots}{Unused (inherited from base plot)}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for Catalog objects
}
\examples{
\dontrun{

catalog = catalog("<Path to a folder containing a set of .las files>")
plot(catalog)
}
}

