% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r
\name{plot.gridMetrics}
\alias{plot.gridMetrics}
\title{Plot an object of class gridMetrics in 2D}
\usage{
\method{plot}{gridMetrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{A data.frame or data.table of class gridMetrics.}

\item{z}{character. The field to plot. If NULL, autodetect.}

\item{\dots}{Unused}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for a gridMetrics data.frame
}
\details{
The \dots param provides additional arguments to \link[fields:image.plot]{image.plot}.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = LoadLidar(LASfile)

# Canopy surface model with 4 m^2 cells
gridMetrics(lidar, 2, max(Z)) \%>\% plot

# Mean height with 400 m^2 cells
gridMetrics(lidar, 20, mean(Z)) \%>\% plot

# Define your own metric function
myMetrics = function(z, i, angle, pulseID)
{
  ret = list(
        npulse  = length(unique(pulseID)),
        hmean   = mean(z),
        hmax    = max(z),
        imean   = mean(i),
        angle   = mean(abs(angle))
        )

   return(ret)
 }

metrics = gridMetrics(lidar, 20, myMetrics(Z, Intensity, ScanAngle, pulseID))

plot(metrics, "hmean")
plot(metrics, "hmax")
plot(metrics, "imean")
}
\seealso{
\link[lidR:gridMetrics]{gridMetrics}
\link[lidR:canopyModel]{canopyModel}
\link[fields:image.plot]{image.plot}
\link[lidR:plot3d]{plot3d}
}

