% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLidar.r
\name{plot.Lidar}
\alias{plot}
\alias{plot.Lidar}
\title{Plot LiDAR data}
\usage{
\method{plot}{Lidar}(x, y, color = "Z", colorPalette = "height.colors",
  bg = "black", ...)
}
\arguments{
\item{x}{An object of the class \code{Lidar}}

\item{y}{Unused (inherited from R base)}

\item{color}{characters. The field used to colorize the points. Default is Z coordinates}

\item{colorPalette}{characters. A color palette name. Default is \code{height.colors} provided by the package lidR}

\item{bg}{The color for the background Default is black.}

\item{\dots}{Supplementary parameters for \link[rgl:points3d]{points3d}}
}
\description{
This functions implements a 3D plot method for Lidar objects
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")

lidar = LoadLidar(LASfile)

plot(lidar)
plot(lidar, color = "Intensity", colorPalette = "heat.colors")
}
\seealso{
\link[rgl:points3d]{points3d}
\link[lidR:height.colors]{height.colors}
\link[grDevices:heat.colors]{heat.colors}
\link[lidR:Lidar]{Class Lidar}
}

