% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r, R/ClassTIN.r
\name{plot.gridMetrics}
\alias{plot.TIN}
\alias{plot.gridMetrics}
\title{Plot an object of class gridMetrics in 2D}
\usage{
\method{plot}{gridMetrics}(x, z = NULL, ...)

\method{plot}{TIN}(x, ...)
}
\arguments{
\item{x}{an object of class gridMetrics}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{...}{Unused}

\item{\dots}{Other parameters for \code{image.plot}}

\item{x}{An object of class \code{TIN}}
}
\description{
Plot an object of class gridMetrics in 2D

Plot method for object \code{TIN}.
}
\examples{
x = runif(20)
y = runif(20)
z = runif(20)
tin = TIN(x, y, z)
plot(tin)
}
\seealso{
\link[fields:image.plot]{image.plot}

\link[lidR:TIN]{TIN}
}

