% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r, R/ClassTIN.r
\name{plot.gridMetrics}
\alias{plot.TIN}
\alias{plot.gridMetrics}
\title{Plot an object of class gridMetrics in 2D}
\usage{
\method{plot}{gridMetrics}(x, z = NULL, ...)

\method{plot}{TIN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{TIN}}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{...}{Unused}
}
\description{
This functions implements a \link[graphics:plot]{plot} method for gridMetrics data.frame

This functions implements a \link[graphics:plot]{plot} method for class TIN
}
\details{
The \dots param enables to provide more arguments to \link[fields:image.plot]{image.plot} function
}
\examples{
x = runif(20)
y = runif(20)
z = runif(20)
tin = TIN(x, y, z)
plot(tin)
}
\seealso{
\link[lidR:gridMetrics]{gridMetrics}
\link[fields:image.plot]{image.plot}

\link[lidR:TIN]{TIN}
}

