% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGridMetrics.r
\name{plot3d}
\alias{plot3d}
\alias{plot3d.gridMetrics}
\title{Plot3d generic function}
\usage{
plot3d(x, z = NULL, ...)

\method{plot3d}{gridMetrics}(x, z = NULL, ...)
}
\arguments{
\item{x}{an object of class gridMetrics}

\item{z}{character. The field to plot. If NULL, autodetect}

\item{\dots}{Other parameters for \code{surface3d}}
}
\description{
Plot3d generic function

Plot an object of class gridMetrics in 3D
}
\seealso{
\link[rgl:surface3d]{surface3d}
}

