% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.r
\name{plot3d}
\alias{plot3d}
\title{Plot a wireframe of a \code{RasterLayer} or a \code{lasmetrics} object}
\usage{
plot3d(x, y, add = FALSE, bg = "black", ...)
}
\arguments{
\item{x}{An object of the class \code{RasterLayer} or \code{lasmetrics}}

\item{y}{Unused (inherited from R base)}

\item{add}{logical. if TRUE, add to current 3D plot.}

\item{bg}{The color for the background. Default is black.}

\item{\dots}{Supplementary parameters for \link[rgl:surface3d]{surface3d}}
}
\description{
Plot a wireframe of a \code{RasterLayer} or a \code{lasmetrics} object
}
