% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{summary,LAS-method}
\alias{summary}
\alias{print,LAS-method}
\alias{summary,LAScatalog-method}
\alias{print.lidRAlgorithm}
\title{Summary and Print for \code{LAS*} objects}
\usage{
print(x, ...)

\S4method{summary}{LAS}(object, ...)

\S4method{print}{LAS}(x)

\S4method{summary}{LAScatalog}(object, ...)

\method{print}{lidRAlgorithm}(x, ...)
}
\arguments{
\item{...}{Unused}

\item{object, x}{A \code{LAS*} object or other lidR related objects.}
}
\description{
Summary and Print for \code{LAS*} objects
}
