% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassLiDARProject.r
\name{processParallel}
\alias{processParallel}
\title{Apply a function to a set of tiles}
\usage{
processParallel(x, func, mc.cores = "auto", combine = "rbind", ...)
}
\arguments{
\item{x}{A LiDARProject object}

\item{func}{A function which have one parameter: the name of a .las file}

\item{mc.cores}{numeric. Number of core used. Default is "auto"}

\item{combine}{character. The function used to merge the outputs of the \code{func} function}

\item{\dots}{Other parameters for \code{mclapply}}
}
\description{
Apply a function to a set of tiles using several cores
}
\seealso{
\link[parallel:mclapply]{mclapply}
}

