% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_tree_detection.R
\name{ptrees}
\alias{ptrees}
\title{Algoritm for individual tree detection or segmentation.}
\usage{
ptrees(k, hmin = 2, nmax = 7L)
}
\arguments{
\item{k}{integer vector. A serie of k-nearest neighbors to use. In this original paper a k refers
to a 'scale' of analyse (see reference).}

\item{hmin}{scalar. This is an addition from the original paper to limit oversegmentation.
Point below this threshold cannot initiate new trees or increase a hull (see details). Set to \code{-Inf}
to strictly respect original paper.}

\item{nmax}{integer. This is an addition from the original paper to protect against uncomputable
cases (see details). Set to \code{+Inf} to strictly respect the original paper (not recommended)}
}
\description{
This function is made to be used in \link{tree_detection} or \link{lastrees} not alone. Segmentation
algorithm proposed  by Vega et al. (2014) (see references and section details). When used in the function
\link{tree_detection} it runs only the fisrt part of the method i.e. the detection of the trees
while when run in  \link{lastrees} it performs the whole segmentation including the tree segmentation
(see details).
}
\details{
This function has been written by the \code{lidR} authors from the original article. We made our
best to implement as far as possible exactly what is written in the original paper but we
cannot affirm that it is this exact original algorithm. Also, minor variations were introduced to
fix some issues that were not adressed in the original paper.
\itemize{
\item Addition of the parameter \code{hmin}: to reduce oversegmentation we introduced a minium height
threshold. Point below this thresold cannot initiate new trees during the tree detection and cannot
incrase a crown hull during the segmentation. This way the ground is not segmented but the segmentation
of each tree is prolongated to the ground anyway.
\item Addition of the parameter \code{nmax}: in the original article page 103 figures 5, the number
of possible combination is 2^n-n-1. This exponential number of combinations lead, in some cases
to an infinite computation time. During the developpement we got cases where the number of combinations
to consider was beyond a billion. If the number of tree to consider between two scales is greater
than \code{nmax} (i.e. the number of combination is greater than 2^nmax-nmax-1) then the "TreeSegment"
from the biggest scale is retained anyway, the smallest scale is considered as dummy.
}
Notice that to use the PTree strictly as originally described, the point cloud should
not be normalized (see reference). In that case the parameter '\code{hmin}' is miningless and can
be set to \code{-Inf} for example.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz")

k = c(30,15)
ttops = tree_detection(las, ptrees(k))

lastrees(las, ptrees(k))
}
\seealso{
Other Algorithm: \code{\link{csf}},
  \code{\link{dalponte2016}}, \code{\link{hamraz2016}},
  \code{\link{highest}}, \code{\link{homogenize}},
  \code{\link{li2012}}, \code{\link{lmf}},
  \code{\link{manual}}, \code{\link{multichm}},
  \code{\link{pmf}}, \code{\link{random}},
  \code{\link{silva2016}}, \code{\link{tin}},
  \code{\link{watershed}}, \code{\link{wing2015}}

Other Individual Tree Detection: \code{\link{lmf}},
  \code{\link{manual}}, \code{\link{multichm}}

Other Individual Tree Segmentation: \code{\link{dalponte2016}},
  \code{\link{hamraz2016}}, \code{\link{li2012}},
  \code{\link{silva2016}}, \code{\link{watershed}}

Other Point Cloud Based Segmentation: \code{\link{hamraz2016}},
  \code{\link{li2012}}
}
\concept{Algorithm}
\concept{Individual Tree Detection}
\concept{Individual Tree Segmentation}
\concept{Point Cloud Based Detection}
\concept{Point Cloud Based Segmentation}
