% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassLidar.r
\docType{methods}
\name{pulseDensity}
\alias{pulseDensity}
\alias{pulseDensity,Lidar-method}
\title{Pulse density surface model}
\usage{
pulseDensity(obj, res = 4)

\S4method{pulseDensity}{Lidar}(obj, res = 4)
}
\arguments{
\item{obj}{An object of class \code{Lidar}}

\item{res}{numeric. The size of a grid cell in LiDAR data coordinates units. Default is 4 units i.e. 16 square units cells.}
}
\value{
It returns a \code{data.table} with the class \code{gridMetrics} which which enables easier plotting.
}
\description{
Creates a pulse density surface model using a LiDAR cloud of points.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.las", package="lidR")
lidar = LoadLidar(LASfile)

lidar \%>\% pulseDensity(5) \%>\% plot
lidar \%>\% pulseDensity(5) \%>\% plot
}
\seealso{
\code{\link[lidR:gridMetrics]{gridMetrics}}
}

