% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLas.r
\name{readLAS}
\alias{readLAS}
\title{Read a .las file}
\usage{
readLAS(LASfile, fields = "standard")
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read .las file
}

