% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLas.r
\name{readLAS}
\alias{readLAS}
\title{Read a .las file}
\usage{
readLAS(LASfile, fields = "standard")
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read .las files
}
\note{
The function was first written by Michael Sumner and Carlos Alberto Silva.
This is a modified version of the original function that is less memory-intensive
and more flexible for the various lidR package usages. The majority of the work was done by the authors
of the rLidar package.
}
\author{
Michael Sumner and Carlos Alberto Silva.
}

