% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLas.r
\name{readLAS}
\alias{readLAS}
\title{Read a .las file}
\usage{
readLAS(LASfile, fields = "standard")
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read .las file
}
\note{
The function was firsly wrote by Michael Sumner and Carlos Alberto Silva.
This version is a function modify from the first one whichis less greedy in memory
and more flexible for the lidR package usages. But the greater work was made by the authors
of rLidar package.
}
\author{
Michael Sumner and Carlos Alberto Silva.
}

