% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLAS.r
\name{readLAS}
\alias{readLAS}
\title{Read las or laz files}
\usage{
readLAS(files, Intensity = TRUE, ReturnNumber = TRUE,
  NumberOfReturns = TRUE, ScanDirectionFlag = FALSE,
  EdgeofFlightline = FALSE, Classification = TRUE, ScanAngle = TRUE,
  UserData = FALSE, PointSourceID = FALSE, RGB = TRUE, pulseID = TRUE,
  flightlineID = TRUE, XYZonly = FALSE, all = FALSE)
}
\arguments{
\item{files}{array of character or a \link[lidR:Catalog]{Catalog} object}

\item{Intensity}{logical. do you want to load Intensity field? default: TRUE}

\item{ReturnNumber}{logical. do you want to load ReturnNumber field? default: TRUE}

\item{NumberOfReturns}{logical. do you want to load NumberOfReturns field? default: TRUE}

\item{ScanDirectionFlag}{logical. do you want to load ScanDirectionFlag field? default: FALSE}

\item{EdgeofFlightline}{logical. do you want to load EdgeofFlightline field? default: FALSE}

\item{Classification}{logical. do you want to load Classification field? default: TRUE}

\item{ScanAngle}{logical. do you want to load intensity field? default: TRUE}

\item{UserData}{logical. do you want to load UserData field? default: FALSE}

\item{PointSourceID}{logical. do you want to load PointSourceID field? default: FALSE}

\item{RGB}{logical. do you want to load intensity R,G and B? default: TRUE}

\item{pulseID}{logical. do you want to compute extra field pulseID? default: TRUE}

\item{flightlineID}{logical. do you want to compute extra field flightlineID? default: TRUE}

\item{XYZonly}{logical. Overwrite every other options. Load only X, Y, Z fields. default: FALSE}

\item{all}{logical. Overwrite every other options. Load everything. default: FALSE}
}
\value{
A LAS object
}
\description{
Read las or laz files in format 1 to 4 according to LAS specification and return an object of class LAS
}
\details{
If several files are given the returned LAS object is considered as one LAS file.
The informations keep in the header are those read only from the first file of the list.
The optional logical parameters enable to do not load some information from the files to save memory if these informations
are useless for user's purpose. Indeed, the readLAS function does no stream the data. Everything is loaded into the
computer memory (RAM) in a not completely optimized way (R does not enable to manage many data type).
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile)
}
\seealso{
\link[lidR:LAS-class]{Class LAS}
\link[lidR:Catalog]{Catalog}
}

