% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_readLAS.R
\name{readLAS}
\alias{readLAS}
\title{Read .las or .laz files}
\usage{
readLAS(files, select = "*", filter = "")
}
\arguments{
\item{files}{characters. Path(s) to one or several a file(s). Can also be a
\link[lidR:LAScatalog-class]{LAScatalog} object.}

\item{select}{character. Read only attributes of interest to save memory (see details).}

\item{filter}{character. Read only points of interest to save memory (see details).}
}
\value{
A LAS object
}
\description{
Reads .las or .laz files into an object of class \link[lidR:LAS-class]{LAS}. If several files are read at
once the returned LAS object is considered as one LAS file. The optional parameters enable the user
to save a substantial amount of memory by choosing to load only the attributes or points of interest.
The LAS formats 1.1 to 1.4 are supported. Point Data Record Format 0,1,2,3,5,6,7,8 are supported.
}
\details{
\strong{Select:} the 'select' argument specifies the data that will actually be loaded. For example,
'xyzia' means that the x, y, and z coordinates, the intensity and the scan angle will be loaded.
The supported entries are t - gpstime, a - scan angle, i - intensity, n - number of returns,
r - return number, c - classification, s - synthetic flag, k - keypoint flag, w - withheld flag,
o - overlap flag (format 6+), u - user data, p - point source ID, e - edge of flight line flag,
d - direction of scan flag, R - red channel of RGB color, G - green channel of RGB color,
B - blue channel of RGB color, N - near-infrared channel. C - scanner channel (format 6+).
Also numbers from 1 to 9 for the extra bytes data numbers 1 to 9. 0 enables all extra bytes to be
loaded and '*' is the wildcard that enables everything to be loaded from the LAS file. \cr
Note that x, y, z are implicit and always loaded. 'xyzia' is equivalent to 'ia'.\cr\cr
\strong{Filter:} the 'filter' argument allows filtering of the point cloud while reading files.
This is much more efficient than \link{filter_poi} in many ways. If the desired filters are known
before reading the file, the internal filters should always be preferred. The available filters are
those from \code{LASlib} and can be found by running the following command: \code{readLAS(filter = "-help")}.
(see also \link[rlas:read.las]{rlas::read.las}). From \code{rlas} v1.4.0 the transformation commands
can also be passed via the argument filter.
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)
las = readLAS(LASfile, select = "xyz")
las = readLAS(LASfile, select = "xyzi", filter = "-keep_first")
las = readLAS(LASfile, select = "xyziar", filter = "-keep_first -drop_z_below 0")

# Negation of attributes is also possible (all except intensity and angle)
las = readLAS(LASfile, select = "* -i -a")
}
