% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readLASheader}
\alias{readLASheader}
\title{Read header in a las or laz file}
\usage{
readLASheader(file)
}
\arguments{
\item{file}{character. the name of the file which the data are to be read from}
}
\value{
A list
}
\description{
Read data from a las  or laz file in format 1 to 4 according to LAS specification and return a list.
}
\details{
This function musn't be used as is. It is an internal function. Please use \link[lidR:readLAS]{readLAS} abstraction.
}

