% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLASheader.r
\name{readLASheader}
\alias{readLASheader}
\title{Read a .las file header}
\usage{
readLASheader(LASfile)
}
\arguments{
\item{LASfile}{character. filename of .las file}
}
\value{
A list with the las header fields
}
\description{
Methods to read .las file header
}

