% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_readMSLAS.R
\name{readMSLAS}
\alias{readMSLAS}
\title{Read multispectral .las or .laz files}
\usage{
readMSLAS(files1, files2, files3, select = "*", filter = "")
}
\arguments{
\item{files1, files2, files3}{characters. Path(s) to one or several a file(s). Each argument being
one channel.}

\item{select, filter}{character. See \link{readLAS}.}
}
\value{
A LAS object
}
\description{
Multispectral laser data are often stored in 3 differents files. If this is the case this function
reads the .las or .laz files of each channel and merges them into an object of class
\link[=LAS-class]{LAS} and takes care of attributing an ID to each channel. If the
multisprectral point cloud is already stored in a single file, use \link{readLAS}. This function
is somewhat experimental and its names could change.
}
