% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGeneric.r, R/ClassCatalog.r
\docType{methods}
\name{retrieveInventoryTiles}
\alias{retrieveInventoryTiles}
\alias{retrieveInventoryTiles,Catalog-method}
\title{Retrieve the tiles containing plots}
\usage{
retrieveInventoryTiles(obj, plotnames, x, y, radius, buffer = 2)

\S4method{retrieveInventoryTiles}{Catalog}(obj, plotnames, x, y, radius,
  buffer = 2)
}
\arguments{
\item{obj}{A Catalog object}

\item{plotnames}{vector. A set of plot names}

\item{x}{vector. A set of x plot coordinates}

\item{y}{vector. A set of y plot coordinates}

\item{radius}{numeric or vector. A radius or a set of radiuses of plots}

\item{buffer}{numeric. A buffer value to expend the extent of search}
}
\description{
When the user have a set of coordinates of ground inventory, he can extract the
lidar data associated with these plot automatically. This function retrieve the tiles
for each plot. This function is used by \link[lidR:extractGroundInventory]{extractGroundInventory}.
Users do not rally need it.
}
\details{
When the user have a set of coordinates of ground inventory, he can extract the
lidar data associated with these plot automatically. The algorithm is able to find tiles
for plots falling between several tiles (on the edges) if tiles are organized in a grid.
}

