% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_canopy_roughtness.r
\name{rumple_index}
\alias{rumple_index}
\title{Rumple index of roughness}
\usage{
rumple_index(x, y = NULL, z = NULL, ...)
}
\arguments{
\item{x}{A RasterLayer or a lasmetrics or a vector of x point coordinates.}

\item{y}{numeric if x is a vector of coordinates: the y coordinates.}

\item{z}{numeric if x is a vector of coordinates: the z coordinates.}

\item{...}{unused}
}
\value{
numeric.
}
\description{
Computes the roughness of a surface as the ratio between visible area over the ground area.
For each type of input the function make the Delaunay triangulation of the points and computes
the ratio between the area of the triangle and the area or the convexhull (area of the
triangles projected on plane X-Y).
}
\examples{
x = runif(20, 0, 100)
y = runif(20, 0, 100)

# Perfectly flat surface, rumple_index = 1
z = rep(10, 20)
rumple_index(x, y, z)

# Rought surface, rumple_index > 1
z = runif(20, 0, 10)
rumple_index(x, y, z)

# More rought surface, rumple_index increases
z = runif(20, 0, 50)
rumple_index(x, y, z)

# Measure of roughness is scale dependent
rumple_index(x, y, z)
rumple_index(x/10, y/10, z)

# Use with a canopy height model
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)
chm = las \%>\% grid_canopy
rumple_index(chm)
}
