% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_threads.r
\name{set_lidr_threads}
\alias{set_lidr_threads}
\alias{get_lidr_threads}
\title{Set or get number of threads that lidR should use}
\usage{
set_lidr_threads(n)

get_lidr_threads()
}
\description{
Set or get number of threads that lidR should use
}
