% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasmetrics.r
\docType{data}
\name{stdmetrics}
\alias{.stdmetrics}
\alias{.stdmetrics_ctrl}
\alias{.stdmetrics_i}
\alias{.stdmetrics_pulse}
\alias{.stdmetrics_rn}
\alias{.stdmetrics_z}
\alias{stdmetrics}
\alias{stdmetrics_ctrl}
\alias{stdmetrics_i}
\alias{stdmetrics_pulse}
\alias{stdmetrics_rn}
\alias{stdmetrics_z}
\title{Predefined standard metrics function}
\format{An object of class \code{expression} of length 1.}
\usage{
stdmetrics(x, y, z, i, a, rn, class, pulseID, dz = 1)

.stdmetrics

stdmetrics_z(z, dz = 1)

.stdmetrics_z

stdmetrics_i(i, z = NULL, class = NULL, rn = NULL)

.stdmetrics_i

stdmetrics_rn(rn, class = NULL)

.stdmetrics_rn

stdmetrics_pulse(pulseID, rn)

.stdmetrics_pulse

stdmetrics_ctrl(x, y, z, a)

.stdmetrics_ctrl
}
\arguments{
\item{x, y, z, i, a}{Coordinates of the points, Intensity and ScanAngle}

\item{rn, class}{ReturnNumber, Classification}

\item{pulseID}{The number referencing each pulse}

\item{dz}{Layer thickness for metric requiering such data such as \link[lidR:entropy]{entropy}}
}
\description{
Predefined function usable in \link[lidR:grid_metrics]{grid_metrics} and \link[lidR:cloud_metrics]{cloud_metrics}
and their convenient shortcuts. The philosophy of \code{lidR} package is to provide an easy way
to compute your own new metrics not to provide them. However a set of classical metric is
already predefined to save time, save pain and to ensure computation efficiency. To use
these function please read Details and examples sections.
}
\details{
The function names, their parameters and the output names  of the metrics rely on a nomenclature chose to get short
names and labels:
\itemize{
\item{\code{z}: refer to the elevation}
\item{\code{i}: refer to the intensity}
\item{\code{rn}: refer to the return number}
\item{\code{q}: refer to quantile}
\item{\code{a}: refer to the ScanAngle}
\item{\code{n}: refer to a number (a count)}
\item{\code{p}: refer to a percentatge}
}
For example the metric named \code{zq60} refer to the elavation, quantile, 60.
It's the 60th percentile of elevations. The metric \code{pground} refer to a percentage. It's the
percentage of returnclassified as ground. The function \code{stdmetric_i} refer to metrics of
intensity. If a doubt subsist you can retrive a description of each existing metric
on the lidR wiki page: https://github.com/Jean-Romain/lidR/wiki/stdmetrics.\cr\cr
Some functions have optional parameters. If these parameter are not provided the function
computes only a subset of existing metrics. For example \code{stdmetrics_i} requieres the intensity
values, but if the elevation values are provided it can computed more metrics such as cumulative
intensity at a given percentile of height.\cr\cr
Each function have a convenient variable associated. It is the name of the function with a
dot before the name. It enables to use the function without writing parametrers. The cost
of such feature is a non flexibility. It corresponds to a predefined behaviour (see examples)
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
lidar = readLAS(LASfile, all = TRUE)

# All the predefined function
lidar \%>\% grid_metrics(stdmetrics(X,Y,Z,Intensity, ScanAngle,
                                  ReturnNumber, Classification, pulseID,
                                  dz = 1))

# Super convenient shortcut
lidar \%>\% grid_metrics(.stdmetrics)

# Basic metrics from intensities
lidar \%>\% grid_metrics(stdmetrics_i(Intensity))

# All the metrics from intensities
lidar \%>\% grid_metrics(stdmetrics_i(Intensity, Z, Classification, ReturnNumber))

# Super convinient shortcut
lidar \%>\% grid_metrics(.stdmetrics_i)

# Compute the metrics only on first return
lidar \%>\% lasfilterfirst \%>\% grid_metrics(.stdmetrics_z)

# Compute the metrics with a threshold at 2 meters
lidar \%>\% lasfilter(Z > 2) \%>\% grid_metrics(.stdmetrics_z)

# Works also with cloud_metrics
lidar \%>\% cloud_metrics(.stdmetrics)

# Combine some predefined function with your own new metrics
# Here convenient shortcut are no more usable.
myMetrics = function(z, i)
{
  metrics = list(
    mymetric1 = mean(z*i),       # Mean of products of z by intensity
    mymetric2 = mean(z[i > 20])  # Mean elevation of point with an intensity > 20
  )

  return( c(metrics, stdmetrics_z(z)) )
}

lidar \%>\% grid_metrics(myMetrics(Z, Intensity))

# You can write your own convenient shorcut like that
.myMetrics = expression(myMetrics(Z,Intensity))

lidar \%>\% grid_metrics(.myMetrics)
}
\seealso{
\link[lidR:grid_metrics]{grid_metrics}
\link[lidR:cloud_metrics]{cloud_metrics}
}
\keyword{datasets}

