% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassummary.r
\name{summary.LAS}
\alias{summary.LAS}
\title{Summary of LAS data}
\usage{
\method{summary}{LAS}(object, ...)
}
\arguments{
\item{object}{An object of the class \code{LAS}}

\item{\dots}{Unused (inherited from R base)}
}
\description{
This functions implements a \link[base:summary]{summary} method for LAS objects
}
\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")

lidar = readLAS(LASfile)

summary(lidar)

}
\seealso{
\link[lidR:LAS]{Class LAS}
}
