% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/utils_base.R
\name{summary,LAS-method}
\alias{summary,LAS-method}
\alias{summary}
\alias{print,LAS-method}
\alias{summary,LAScatalog-method}
\alias{print.lidRAlgorithm}
\alias{tools}
\alias{dim.LAS}
\alias{dim.LAScatalog}
\alias{ncol.LAS}
\alias{nrow.LAScatalog}
\alias{names.LAS}
\alias{rbind.LAS}
\alias{npoints}
\alias{density}
\title{Tools inherited from base R for LAS* objects}
\usage{
\S4method{summary}{LAS}(object, ...)

\S4method{print}{LAS}(x)

\S4method{summary}{LAScatalog}(object, ...)

\method{print}{lidRAlgorithm}(x, ...)

\method{dim}{LAS}(x)

\method{dim}{LAScatalog}(x)

ncol.LAS(x)

nrow.LAScatalog(x)

\method{names}{LAS}(x)

\method{rbind}{LAS}(...)

npoints(x, ...)

density(x, ...)
}
\arguments{
\item{object}{A \code{LAS*} object or other lidR related objects.}

\item{...}{LAS* objects if it is the sole argurment (e.g. in rbind())}

\item{x}{a LAS* object}
}
\description{
Tools inherited from base R for LAS* objects
}
