% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_base.R
\name{tools}
\alias{tools}
\alias{dim.LAS}
\alias{dim.LAScatalog}
\alias{ncol.LAS}
\alias{nrow.LAScatalog}
\alias{names.LAS}
\alias{rbind.LAS}
\alias{npoints}
\alias{density}
\title{Tools inherited from base R for LAS* objects}
\usage{
\method{dim}{LAS}(x)

\method{dim}{LAScatalog}(x)

ncol.LAS(x)

nrow.LAScatalog(x)

\method{names}{LAS}(x)

\method{rbind}{LAS}(...)

npoints(x, ...)

density(x, ...)
}
\arguments{
\item{x}{a LAS* object}

\item{...}{LAS* objects if it is the sole argurment (e.g. in rbind())}
}
\description{
Tools inherited from base R for LAS* objects
}
