% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_detection.r
\name{tree_detection}
\alias{tree_detection}
\title{Tree top detection}
\usage{
tree_detection(x, algorithm, ...)
}
\arguments{
\item{x}{A object of class \code{LAS} or an object representing a canopy height model
such as a \code{RasterLayer} or a \code{lasmetrics} or a \code{matrix} depending on the algorithm
used (see respective documentation)}

\item{algorithm}{character. Can be either \code{"lmf"} or \code{"ptrees"}.}

\item{...}{Other parameters for each respective algorithm (see section "see also".}
}
\value{
The output of each algorithm as documedevtools::chented in each method.
}
\description{
This function is a wrapper for all the implemented tree detection methods. Tree dectection function
find the position of the trees before the segmentation process. Several methods may be used
(see documentation of each method)
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")

ttops = tree_detection(las, "lmf", ws = 5)

plot(las)
with(ttops, rgl::points3d(X, Y, Z, col = "red", size = 5, add = TRUE))
}
\seealso{
\link{tree_detection_lmf} \link{tree_detection_ptrees} \link{tree_detection_manual}
}
