% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_detection.r
\name{tree_detection}
\alias{tree_detection}
\title{Individual tree detection}
\usage{
tree_detection(las, algorithm)
}
\arguments{
\item{las}{An object of class \code{LAS} or \code{LAScatalog} or a \code{RasterLayer} representing a
canopy height model. depending on the algorithm used (see respective documentation)}

\item{algorithm}{An algorithm for individual tree detection. lidR have four of them: \link{lmf}, \link{ptrees},
\link{multichm} and \link{manual}.}
}
\value{
A SpatialPointsDataFrame with an attribute Z for the tree tops and treeID for an individual
ID for each tree.
}
\description{
Individual Tree detection function that find the position of the trees before the segmentation process. Several
methods may be used.
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile, select = "xyz", filter = "-drop_z_below 0")

ttops = tree_detection(las, lmf(ws = 5))

plot(las)
rgl::spheres3d(ttops@coords[,1], ttops@coords[,2], ttops@data$Z, col = "red", size = 5, add = TRUE)
}
