% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLAS.r
\docType{methods}
\name{writeLAS}
\alias{writeLAS}
\alias{writeLAS,LAS-method}
\title{Write a las file}
\usage{
writeLAS(obj, file)

\S4method{writeLAS}{LAS}(obj, file)
}
\arguments{
\item{obj}{an object of class LAS}

\item{file}{character. a character string naming an output file}
}
\value{
void
}
\description{
Write a LAS object into a binary file (.las if specified in filename)
}

