% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeLAS}
\alias{writeLAS}
\title{Read a .las file}
\usage{
writeLAS(file, LASheader)
}
\arguments{
\item{LASfile}{character. filename of .las file}

\item{fields}{character. \code{"minimal"}, \code{"standard"}, \code{"all"}.}
}
\value{
A data.table
}
\description{
Methods to read .las files
}

