% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLAS.r
\name{writeLAS}
\alias{writeLAS}
\title{Write a las or laz file}
\usage{
writeLAS(.las, file)
}
\arguments{
\item{.las}{an object of class LAS}

\item{file}{character. A character string naming an output file}
}
\value{
void
}
\description{
Write a LAS object into a binary file (.las or .laz specified in filename)
}
